### 2018-11-12 #1079 @whedon
[PRE REVIEW]: OApackage: A Python package for generation and analysis of orthogonal arrays and conference designs
**Submitting author:** @eendebakpt (<a href="http://orcid.org/0000-0001-7018-1124">Pieter Eendebak</a>)
**Repository:** <a href="https://github.com/eendebakpt/oapackage" target ="_blank">https://github.com/eendebakpt/oapackage</a>
**Version:** 2.5.1
