### 2018-11-12 #1079 @danielskatz
👋 @eendebakpt

1) there seem to be a bunch of reference problems in the compiled PDF.  If you fix the .md, you should be able recompile here by adding a new comment that starts with:
```@whedon generate pdf```

2) Can you suggest reviewers, perhaps from [this list of people](https://bit.ly/joss-reviewers) who have already agreed to review for JOSS, or from outside the list as well.  I think it would be good to get one reviewer who can check the Python API as well as the C++ source.
