### 2018-11-12 10a0d69
Remove symmetry method  (#73)
* cleanup oaclustergather

* remove symmetry method from code
