+++ src/lmc.h
@@ -264,41 +264,6 @@ typedef larray< colindex_t > colpermtypelight;
 typedef std::vector< int > colpermtype;
 typedef std::vector< colpermtype > colpermset;
 
-#define USE_ROWPERM_POOL
-
-/// class representing array symmetries
-struct arraysymmetry {
-        larray< rowindex_t > *rowperm;
-
-        larray< colindex_t > colperm;
-
-      public:
-        static object_pool< larray< rowindex_t > > rowpermpool;
-
-      public:
-        ~arraysymmetry ();
-
-        arraysymmetry (const dyndata_t *dyndata);
-
-        /// copy constructor
-        arraysymmetry (const arraysymmetry &rhs);
-
-        // Copy assignment operator
-        arraysymmetry &operator= (const arraysymmetry &rhs);
-
-        inline int operator== (const arraysymmetry &as2) {
-                return (((*rowperm) == *(as2.rowperm)) && (this->colperm == as2.colperm));
-        }
-
-        int N () const { return rowperm->size (); }
-        void show () const {
-                myprintf ("arraysymmetry: rowperm ");
-                print_perm< rowindex_t > (*rowperm, 24);
-                myprintf ("             : colperm ");
-                print_perm (colperm);
-        }
-};
-
 #ifdef _WIN32
 // on windows do not use smart pointers, it is a mess
 //#if _MSC_VER >= 1600
@@ -339,8 +304,6 @@ typedef std::tr1::shared_ptr< symmdata > symmdataPointer;
 #else
 typedef symmdata *symmdataPointer;
 #endif
-
-typedef std::vector< arraysymmetry > symmetryset;
 
 enum INIT_STATE { INIT_STATE_INVALID, COPY, INIT, SETROOT };
 
@@ -384,112 +347,6 @@ struct LMCreduction_t {
 
         LMC_static_struct_t *staticdata;
 
-        class symm_t {
-              public:
-		/// which kind of data to store
-                int store;
-                int ncols;
-                std::vector< colpermset > colperms;
-                std::vector< colpermset > colcombs;
-                std::vector< symmetryset > symmetries;
-
-                symm_t () { ncols = -1; }
-                void show (int verbose = 1) const {
-                        long ns = 0, ncp = 0, ncc = 0;
-                        for (size_t i = 0; i < symmetries.size (); i++) {
-                                if (verbose >= 2) {
-                                        myprintf ("  symm_t: k %ld: symms %ld\n", (long)i,
-                                                  (long)symmetries[i].size ());
-                                }
-                                ns += symmetries[i].size ();
-                        }
-                        for (size_t i = 0; i < colperms.size (); i++)
-                                ncp += colperms[i].size ();
-                        for (size_t i = 0; i < colcombs.size (); i++)
-                                ncc += colcombs[i].size ();
-
-                        myprintf ("symm_t: store %d, symms %ld, colperms %ld, colcombs %ld, ncols %d\n", store, ns,
-                                  ncp, ncc, ncols);
-                }
-                bool valid () const { return ncols > 0; }
-                void makeColpermsUnique (int dverbose = 0) {
-                        symm_t &xx = *this;
-                        // create unique elements
-                        for (size_t i = 0; i < xx.colperms.size (); i++) {
-                                std::sort (xx.colperms[i].begin (), xx.colperms[i].end ());
-
-                                std::vector< colpermtype >::iterator last =
-                                    std::unique (xx.colperms[i].begin (), xx.colperms[i].end ());
-                                if (dverbose >= 2) {
-                                        myprintf ("makeColpermsUnique: i %ld vals %ld, unique %d\n", (long)i,
-                                                  (long)xx.colperms[i].size (), (int)(last - xx.colperms[i].begin ()));
[…]
