### 2018-11-12 9c405e0
update paper references
