### 2018-11-12 89bddf5
update paper references (#74)
