+++ paper/paper.bib
@@ -7,7 +7,7 @@
 
 @article{Schoen2018dsd,
 title={A Classification Criterion for Definitive Screening Designs},
-author={Eric Schoen and Pieter Thijs Eendebak and Peter Goos},
+author={Eric D. Schoen and Pieter Thijs Eendebak and Peter Goos},
 journal = {The Annals of Statistics},
 year={2018}
 }
@@ -58,8 +58,8 @@ year={2018}
 }
 
 
-@misc{ OAdocumentation,
-    author = "P.T. Eendebak and Alan Vazquez",
+@misc{OAdocumentation,
+    author = "P.T. Eendebak and Alan Roberto Vazquez",
     title = "The Orthogonal Array package",
     year = "2018",
     url = "https://oapackage.readthedocs.io/en/latest/",
@@ -85,7 +85,7 @@ year={2018}
 }
 
 @article{Elster1995,
-author = {ELSTER, CLEMENS and NEUMAIER, ARNOLD},
+author = {Elster, Cliemens and Neumaier, Arnold},
 title = {Screening by conference designs},
 journal = {Biometrika},
 volume = {82},
@@ -95,5 +95,60 @@ year = {1995},
 doi = {10.1093/biomet/82.3.589},
 URL = {http://dx.doi.org/10.1093/biomet/82.3.589},
 eprint = {/oup/backfile/content_public/journal/biomet/82/3/10.1093/biomet/82.3.589/2/82-3-589.pdf}
+}
+
+@book{hedayat2012orthogonal,
+  title={Orthogonal arrays: theory and applications},
+  author={Hedayat, A Samad and Sloane, Neil James Alexander and Stufken, John},
+  year={2012},
+  publisher={Springer Science \& Business Media}
+}
+
+
+@online{EendebakOAwebsite,
+  author = {P.T. Eendebak},
+  title = {The Orthogonal Array Package},
+  year = 2018,
+  url = {http://www.pietereendebak.nl/oapackage/},
+  urldate = {2018-11-12}
+}
+
+@online{SloaneOAwebsite,
+  author = {N.J.A. Sloane},
+  title = {Orthogonal Arrays: Theory and Applications},
+  year = 1999,
+  url = {http://neilsloane.com/doc/OA.html},
+  urldate = {2018-11-12}
+}
+
+
+
+@article{Tang1999,
+author = "Tang, Boxin and Deng, Lih-Yuan",
+doi = "10.1214/aos/1017939244",
+fjournal = "The Annals of Statistics",
+journal = "Ann. Statist.",
+month = "12",
+number = "6",
+pages = "1914--1926",
+publisher = "The Institute of Mathematical Statistics",
+title = "Minimum $G_2$-aberration for nonregular fractional factorial designs",
+url = "https://doi.org/10.1214/aos/1017939244",
+volume = "27",
+year = "1999"
+}
+
+@article{Deng1999,
+ ISSN = {10170405, 19968507},
+ URL = {http://www.jstor.org/stable/24306636},
[…]
