### 2018-11-13 #1079 @eendebakpt
@danielskatz Thanks for picking this up. I updated the references in the paper. As for reviewers:

Based on Python/C++ experience: pragyansmita, katyhuff or djmitche
Based on Python and content experience: tupi
