### 2018-11-13 64f39fa
attempt at fix
better would be to leave this blank:   year = {},
but if that doesn't work, which I guess might be the case, n.y. as an abbreviation for no year is better than YEAR
