### 2018-11-13 79ae4ca
add example arrays
