### 2018-11-13 33cb651
remove unused functions
