### 2018-11-14 01a4fe4
cleanup code; add tests to travis
