+++ src/lmc.cpp
@@ -38,22 +38,22 @@ using namespace std;
 
 // worker pool of static objects
 
-static object_pool< LMC_static_struct_t > staticDataPool;
+static object_pool< LMCreduction_helper_t > LMCreduction_pool;
 
-LMC_static_struct_t *getGlobalStatic () {
+LMCreduction_helper_t *acquire_LMCreduction_object () {
 #ifdef NOREUSE
-        LMC_static_struct_t *pp = new LMC_static_struct_t ();
+        LMCreduction_helper_t *pp = new LMCreduction_helper_t ();
         pp->id = 100 * omp_get_thread_num () + int(random () % 40);
         return pp;
 #endif
 
 
-        LMC_static_struct_t *p = 0;
+        LMCreduction_helper_t *p = 0;
 #pragma omp critical
-        { p = staticDataPool.New (); }
+        { p = LMCreduction_pool.New (); }
         return p;
 }
-void releaseGlobalStatic (LMC_static_struct_t *p) {
+void release_LMCreduction_object (LMCreduction_helper_t *p) {
 #ifdef NOREUSE
         delete p;
         return;
@@ -61,47 +61,47 @@ void releaseGlobalStatic (LMC_static_struct_t *p) {
 
 #pragma omp critical
         {
-                staticDataPool.Delete (p);
+                LMCreduction_pool.Delete (p);
         }
 }
 
-void cleanGlobalStatic () { staticDataPool.reset (); }
+void clear_LMCreduction_pool () { LMCreduction_pool.reset (); }
 
 // indexed pool of static objects
-std::vector< LMC_static_struct_t * > globalStaticPool;
-LMC_static_struct_t *getGlobalStaticIndexed (int n) {
-        const int verbose = 0;
+// std::vector< LMC_static_struct_t * > globalStaticPool;
+// LMC_static_struct_t *getGlobalStaticIndexed (int n) {
+//         const int verbose = 0;
+// 
+//         if (verbose)
+//                 myprintf ("getGlobalStatic: n %d, pool size %ld\n", n, (long)globalStaticPool.size ());
+//         if (n >= (int(globalStaticPool.size ()))) {
+//                 myprintf ("  allocating new element in globalStaticPool: n %d, pool size %ld\n", n,
+//                           (long)globalStaticPool.size ());
+//                 size_t psize = globalStaticPool.size ();
+//                 for (int jj = psize; jj <= n; jj++) {
+//                         LMC_static_struct_t *p = new LMC_static_struct_t ();
+//                         if (verbose)
+//                                 myprintf ("new element jj %d\n", jj);
+//                         globalStaticPool.push_back (p);
+//                 }
+//         }
+// 
+//         if (verbose) {
+//                 myprintf ("   ");
+//                 globalStaticPool[n]->show ();
+//         }
+//         return globalStaticPool.at (n);
+// }
+// 
+// void cleanGlobalStaticIndexed () {
+//         printfd ("cleanGlobalStaticIndexed: delete %d items\n", globalStaticPool.size ());
+//         for (size_t jj = 0; jj < globalStaticPool.size (); jj++) {
+//                 delete globalStaticPool[jj];
+//         }
+//         globalStaticPool.resize (0);
+// }
 
-        if (verbose)
-                myprintf ("getGlobalStatic: n %d, pool size %ld\n", n, (long)globalStaticPool.size ());
-        if (n >= (int(globalStaticPool.size ()))) {
-                myprintf ("  allocating new element in globalStaticPool: n %d, pool size %ld\n", n,
-                          (long)globalStaticPool.size ());
[…]
