### 2018-11-14 8b6a7f4
fix python compilation; rename variables
