### 2018-11-14 a39c7f2
refactor code; fix nauty warning
