### 2018-11-14 353acbc
remove dyndata from conference.cpp
