### 2018-11-14 bc0cd6c
update documentation
