### 2018-11-14 7aee311
refactor code; add test
