+++ src/arraytools.cpp
@@ -5500,78 +5500,78 @@ array_link selectArrays (const std::string filename, int ii) {
 
 #endif // FULLPACKAGE, related to arrayfile_t
 
-void selectArrays (const arraylist_t &al, std::vector< int > &idx, arraylist_t &rl) {
+void selectArrays (const arraylist_t &input_list, std::vector< int > &idx, arraylist_t &output_list) {
         for (std::vector< int >::iterator it = idx.begin (); it < idx.end (); it++) {
-                rl.push_back (al.at (*it));
+                output_list.push_back (input_list.at (*it));
         }
 }
-void selectArrays (const arraylist_t &al, std::vector< long > &idx, arraylist_t &rl) {
+void selectArrays (const arraylist_t &input_list, std::vector< long > &idx, arraylist_t &output_list) {
         for (std::vector< long >::iterator it = idx.begin (); it < idx.end (); it++) {
-                if ((*it) >= 0 && ((*it) < (int)al.size ())) {
-                        rl.push_back (al.at (*it));
+                if ((*it) >= 0 && ((*it) < (int)input_list.size ())) {
+                        output_list.push_back (input_list.at (*it));
                 } else {
                         myprintf ("selectArrays: index %ld out of bounds!\n", *it);
                 }
         }
 }
 
-arraylist_t selectArrays (const arraylist_t &al, std::vector< int > &idx) {
+arraylist_t selectArrays (const arraylist_t &input_list, std::vector< int > &idx) {
         arraylist_t rl;
         for (std::vector< int >::iterator it = idx.begin (); it < idx.end (); ++it) {
                 int val = *it;
-                if (val >= 0 && val <= (int)al.size ()) {
-                        rl.push_back (al.at (val));
+                if (val >= 0 && val <= (int)input_list.size ()) {
+                        rl.push_back (input_list.at (val));
                 } else {
                         myprintf ("selectArrays: error: index out of bounds: index %d, size %d\n", val,
-                                  (int)al.size ());
+                                  (int)input_list.size ());
                 }
         }
         return rl;
 }
 
-arraylist_t selectArrays (const arraylist_t &al, std::vector< long > &idx) {
+arraylist_t selectArrays (const arraylist_t &input_list, std::vector< long > &idx) {
         arraylist_t rl;
         for (std::vector< long >::iterator it = idx.begin (); it < idx.end (); ++it) {
                 int val = *it;
-                if (val >= 0 && val <= (int)al.size ()) {
-                        rl.push_back (al.at (val));
+                if (val >= 0 && val <= (int)input_list.size ()) {
+                        rl.push_back (input_list.at (val));
                 } else {
                         myprintf ("selectArrays: error: index out of bounds: index %d, size %ld\n", val,
-                                  (long)al.size ());
+                                  (long)input_list.size ());
                 }
         }
         return rl;
 }
 
-array_link array_link::operator+ (array_t v) const {
+array_link array_link::operator+ (array_t value) const {
         array_link tmp = (*this);
         for (int i = 0; i < tmp.n_columns * tmp.n_rows; i++) {
-                tmp.array[i] += v;
+                tmp.array[i] += value;
         }
         return tmp;
 }
-array_link array_link::operator- (array_t v) const {
+array_link array_link::operator- (array_t value) const {
         array_link tmp = (*this);
         for (int i = 0; i < tmp.n_columns * tmp.n_rows; i++) {
-                tmp.array[i] -= v;
-        }
-        return tmp;
-}
-
-array_link array_link::operator+ (const array_link &b) const {
-        assert (this->equalsize (b));
[…]
