### 2018-11-15 6d17282
refactor to have unit tests in a single file
