### 2018-11-15 c2f7794
fix bug in rowsort init; add tests
