+++ .travis.yml
@@ -23,6 +23,8 @@ script:
   - ./oaranktest -I ../docs/examples/OAN56K28.oa
   - ./oainfo ../docs/examples/OAN56K28.oa
   - ./oaextendsingle  -c ../oaconfig.txt -l 1
+  - ./oasplit -i result-20.2-2-2.oa
+  - ./oajoin split-split-0.oa split-split-1.oa -o temp.oa
   - ./oaextendsingle -m 3 -l 1 -c ../oaconfig3.txt --maxk 5 # J4 algorithm
   - ./oaconference -N 12 -k 6 -v 0
 
