### 2018-11-15 4711e7e
refactor code; add tests
