+++ src/unittests.h
@@ -1,17 +1,15 @@
-/** \file arraytools.h
+/** \file unittests.h
 
- \brief Contains the array_link class and related classes.
-
- C++ Interface: arraytools
-
- This file contains definitions are functions to work with (orthogonal) arrays.
- The code is generic (with templates) and partly inlined for speed.
-
- Author: Pieter Eendebak <pieter.eendebak@gmail.com>
-
- Copyright: See LICENSE.txt file that comes with this distribution
-*/
+ \brief Contains unit tests
+ */
 
 #pragma once
+
+int checkConferenceComposition(const array_link &al, int verbose = 0);
+
+void test_conference_candidate_generators(int verbose = 1);
+int checkTransformationInverse(const array_link &al);
+
+int checkConferenceInverse(const array_link &array);
 
 bool testLMC0checkDC(const array_link &al, int verbose = 1);
