### 2018-11-15 8ba110b
refactor code
