+++ src/unittests.h
@@ -5,9 +5,18 @@
 
 #pragma once
 
+int unittest_reduceConferenceTransformation();
+
+int unittest_nautynormalform(const array_link &al, int verbose);
+
+int checkTransformationComposition(const array_link &al, int verbose = 0);
+
+void test_array_manipulation(int verbose = 1);
+
 int checkConferenceComposition(const array_link &al, int verbose = 0);
 
 void test_conference_candidate_generators(int verbose = 1);
+
 int checkTransformationInverse(const array_link &al);
 
 int checkConferenceInverse(const array_link &array);
