### 2018-11-15 79c6ddd
cleanup code
