+++ .travis.yml
@@ -35,18 +35,3 @@ script:
   - ./oapareto.exe result-20.2-2-2.oa -f B -o dummy.oa
 
 
-#addons:
-#  artifacts: true
-  
-# Python build
-#language: python
-#python:
-#  - "2.7"
-#  - "3.4"
-#
-#before_install:
-#  - sudo apt-get update -qq
-#  - sudo apt-get install swig
-#install:
-#  - python setup.py build
-
