### 2018-11-15 ac55e12
bump version; small fixes
