+++ .travis.yml
@@ -26,9 +26,9 @@ script:
 # tools
   - ./oasplit -i result-20.2-2-2.oa
   - ./oajoin split-split-0.oa split-split-1.oa -o temp.oa
-  - ./oacat dummy.oa
-  - ./oaanalyse dummy.oa
-  - ./oaconvert dummy.oa -f B dummy-binary.oa
+  - ./oacat result-20.2-2-2.oa
+  - ./oaanalyse result-20.2-2-2.oa
+  - ./oaconvert result-20.2-2-2.oa -f B dummy-binary.oa
 #
   - ./oaextendsingle -m 3 -l 1 -c ../oaconfig3.txt --maxk 5 # J4 algorithm
   - ./oaconference -N 12 -k 6 -v 0
