### 2018-11-15 d0ffe28
refactor code; add tests (#80)
* refactor code; add tests

* refactor code

* add tests

* cleanup code

* fix unsaved file

* bump version; small fixes

* fix issue with travis

* fix travis
