+++ utils/oaunittest.cpp
@@ -33,215 +33,13 @@ Copyright: See LICENSE.txt file that comes with this distribution
 #include <string>
 #endif
 
-template < class Type > void permute (Type *source, Type *target, const std::vector< int > p) {
-        for (size_t i = 0; i < p.size (); i++)
-                target[i] = source[p[i]];
-}
 
-/// check transformation inverse. return 0 if test is good
-int checkConferenceInverse (const array_link &al) {
-        conference_transformation_t T1 (al);
-        T1.randomize ();
 
-        conference_transformation_t T1i = T1.inverse ();
-        conference_transformation_t II = T1i * T1;
-
-        myassert (II.isIdentity (), "unittest error: inverse of conference matrix transformation\n");
-        myassert (al == II.apply (al), "unittest error: inverse of conference matrix transformation\n");
-
-        return 0;
-}
-
-int unittest_reduceConferenceTransformation()
-{
-	array_link conference_design = exampleArray(45);
-	conference_transformation_t tranformation = reduceConferenceTransformation(conference_design, 0);
-
-	return 0;
-}
-/// unittest: return 1 if all tests are good
-int unittest_nautynormalform(const array_link &al, int verbose) {
-	arraydata_t arrayclass = arraylink2arraydata(al);
-
-	if (verbose >= 2) {
-		myprintf("unittest_nautynormalform: testing on array\n");
-		al.showarray();
-	}
-
-	array_link alr1 = al.randomperm();
-	array_link alr2 = al.randomperm();
-
-	array_transformation_t ttx1 = reduceOAnauty(alr1, 0);
-	array_link alx1 = ttx1.apply(alr1);
-
-	array_transformation_t ttx2 = reduceOAnauty(alr2, 0);
-	array_link alx2 = ttx2.apply(alr2);
-
-	if (alx1 != alx2) {
-		printfd("unittest_nautynormalform: error: transformed graphs unequal!\n");
-
-		myprintf("alx1: \n");
-		alx1.showarray();
-		myprintf("alx2: \n");
-		alx2.showarray();
-
-		return 0;
-	}
-
-	return 1;
-}
-
-/// check composition operator. returns 0 if test id good
-int checkConferenceComposition (const array_link &al, int verbose = 0) {
-        conference_transformation_t T1 (al);
-        T1.randomize ();
-
-        conference_transformation_t T2 (al);
-        T2.randomize ();
-
-        conference_transformation_t T3 = T2 * T1;
-
-        array_link al1 = T1.apply (al);
-        array_link al1t2 = T2.apply (al1);
-        array_link al3 = T3.apply (al);
-
-        if (verbose) {
-                printfd ("checkTransformationComposition: transforms\n");
-                T1.show ();
-                T2.show ();
-                T3.show ();
-                printfd ("checkTransformationComposition: arrays\n");
-                al.showarray ();
-                al1.showarray ();
-                al1t2.showarray ();
-                al3.showarray ();
-        }
-
-        myassert (al3 == al1t2, "unittest error: composition of conference transformations\n");
-
-        return 0;
-}
-
-void checkGenerators (int verbose = 1) {
-        int N = 20;
-        conference_t ct (N, 4, 1);
-        ct.ctype = conference_t::DCONFERENCE;
-        ct.j3zero = 1;
-
-        array_link al = exampleArray (35, 1);
-
-        CandidateGeneratorDouble cgenerator (array_link (), ct);
-        cgenerator.verbose = 0;
-        for (int i = 0; i < 2; i++) {
-                {
-                        const std::vector< conference_column > &cl = cgenerator.generateCandidates (al);
-                        myassert (cl.size () == 3, "unittest error: inverse of array transformation\n");
-
-                        if (verbose >= 2) {
[…]
