### 2018-11-15 46c2973
fix reference to R
