+++ paper/paper.bib
@@ -80,7 +80,7 @@ year={2018}
   author       = {{R Core Team}},
   organization = {R Foundation for Statistical Computing},
   address      = {Vienna, Austria},
-  year         = YEAR,
+  year         = {2018},
   url          = {https://www.R-project.org}
 }
 
