### 2018-11-15 14906ce
Merge dev into master (#79)
* cleanup code

* add example arrays

* remove unused functions

* cleanup code; add tests to travis

* rename LMC helper object

* fix python compilation; rename variables

* refactor code; fix nauty warning

* remove dyndata from conference.cpp

* update documentation

* update docs

* update docs

* refactor code; add test

* refactor to have unit tests in a single file

* fix bug in rowsort init; add tests

* refactor code; add tests (#80)

* refactor code; add tests

* refactor code

* add tests

* cleanup code

* fix unsaved file

* bump version; small fixes


[…]
