### 2018-11-18 #1079 @djmitche
For my reference, since I haven't reviewed anything here since volunteering, [these](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) are the reviewer guidelines.
