### 2018-11-18 e920965
cleanup code
