### 2018-11-19 eendebakpt/oapackage#82 @djmitche
Add community guidelines
The JOSS guidelines for the review in https://github.com/openjournals/joss-reviews/issues/1079  suggest documenting how to:

*   Contribute to the software
*   Report issues or problems with the software
*   Seek support

I suspect all three of these could be handled with some short `SUPPORT.md` and `CONTRIBUTING.md` files in the repo.
