### 2018-11-19 #1079 @djmitche
Aside from the community guidelines noted above, everything looks on the up-and-up.  I would like a general pointer in the docs or README to what the heck a Conference Design is -- it's a particularly difficult thing to search for, since there are a lot of Design Conferences out there.   A quick sentence about how this might be used would be helpful, too.  The paragraph beginning "As an example" in [EN95] was tantalizing, but of course I can't see the second page!

I took some extra time to stare at the `LICENSE` file, as it identifies some separately-licensed bits (the MPL2 bits, in particular, concern me).  However, I think this is fine: anyone using this software in an open-source context will automatically be compliant with the MPL2 license, and anyone using it in a proprietary context has the information they need available to them in `LICENSE`.  Whether that means that the package is essentially MPL2-licensed is neither something I can figure out, nor particularly important to the review.  Which is a long way of saying that it's nice to see proper attention paid to licensing, and I see nothing to be concerned about here.

I perused the C++ and Python source.  I didn't look at everything, and certainly didn't understand the meaning of much of what I saw.  But the code looks well-structured and well-documented.  Nothing seems alarming or suggestive of any systemic issues.  This looks like solid, efficient numerical software.
