### 2018-11-19 #1079 @danielskatz
Thanks @djmitche  - let me try to find another reviewer, then I'll open a review issue, where you can check off stuff like you've started to do above...
Sorry for the delay.
