### 2018-11-21 3e0a5bc
add CONTRIBUTING.md
