+++ CONTRIBUTING.md
@@ -1,0 +1,23 @@
+# Contributing 
+
+### Code style
+
+* Try to follow the [PEP 8](https://www.python.org/dev/peps/pep-0008/) style guide. A usefull tool for automated formatting is [autopep8](https://pypi.python.org/pypi/autopep8). We do allow lines upto 120 characters.
+* Document functions using the [Google docstring](https://sphinxcontrib-napoleon.readthedocs.io/en/latest/example_google.html) convention
+* Add unit tests using the [unittest](https://docs.python.org/3/library/unittest.html) framework
+
+### Submitting code
+
+If you would like to contribute, please submit a pull request. (See the [Github Hello World](https://guides.github.com/activities/hello-world/) example, if you are new to Github).
+
+By contributing to the repository you state you own the copyright to those contributions and agree to include your contributions as part of this project under the BSD license.
+
+### Bugs reports and feature requests
+
+To submit a bug report or feature request use the [Github issue tracker](https://github.com/eendebakpt/oapackage/issues) are for. Search for existing and closed issues. If your problem or idea is not yet addressed, please open a new issue.
+
+
+### Contact
+
+For further information please contact Pieter Eendebak, pieter.eendebak at tno.nl
+
