+++ README.md
@@ -80,14 +80,8 @@ $ make
 $ make install
 ````
 
-Testing
--------
+Contributions, unit testing and support
+---------------------------------------
 
-To perform tests run `pytest`. To obtain a coverage report, run
-```
-$ coverage run --source='./oapackage' -m pytest
-$ coverage report --omit oapackage/markup.py,oapackage/tests/*.py,oapackage/deprecated.py
-```
-
-Continuous integration and testing for the C++ library is performed on [Travis](https://travis-ci.org/eendebakpt/oapackage) and for the Python package on
-[AppVeyor](https://ci.appveyor.com/project/eendebakpt/oapackage-4lws8).
+See the file [CONTRIBUTING.md](https://github.com/eendebakpt/oapackage/blob/master/CONTRIBUTING.md) on GitHub.
+             
