### 2018-11-21 e3c950c
improve coding style
