### 2018-11-21 8898bdf
Cleanup code (#81)
* cleanup code

* cleanup code

* improve coding style

* improve coding style
