+++ oapackage/oahelper.py
@@ -62,8 +62,7 @@ def deprecated(func):
         return func(*args, **kwargs)
     return new_func
 
-#%% Load Qt support
-
+# %%
 
 try:
     try:
@@ -174,8 +173,7 @@ def plot2Dline(line, *args, **kwargs):
         plt.plot(xx, yy, *args, **kwargs)
 
 
-#%% Make nice plots
-# http://blog.olgabotvinnik.com/prettyplotlib/
+# %% Make nice plots
 
 def niceplot(ax, fig=None, despine=True, verbose=0, figurebg=True,
              tightlayout=True, legend=None, almost_black='#222222'):
@@ -210,7 +208,7 @@ def niceplot(ax, fig=None, despine=True, verbose=0, figurebg=True,
 
     ax.tick_params(axis='both', direction='out')
 
-    if not legend is None:
+    if legend is not None:
         if verbose:
             print('niceplot: adjust legend')
 
@@ -229,7 +227,7 @@ def niceplot(ax, fig=None, despine=True, verbose=0, figurebg=True,
         for t in texts:
             t.set_color(almost_black)
 
-    if not fig is None and tightlayout:
+    if (fig is not None) and tightlayout:
         fig.tight_layout(pad=1.0)
 
     if figurebg and fig is not None:
@@ -253,11 +251,11 @@ def enlargelims(factor=1.05):
     yl = (yl[0] - d, yl[1] + d)
     plt.ylim(yl)
 
-#%%
+# %%
 
 
 def selectParetoArrays(array_list, pareto_object):
-    """ Select arrays using a Pareto object 
+    """ Select arrays using a Pareto object
 
     Args:
         array_list (list): list of arrays
@@ -294,7 +292,7 @@ def createPareto(dds, verbose=1):
     return pp
 
 
-#%% Utils
+# %% Utils
 
 
 def static_var(varname, value):
@@ -354,7 +352,7 @@ def findfilesR(directory, pattern):
 
 
 def findfiles(directory, pattern=None):
-    """ Get a list of files 
+    """ Get a list of files
     Args:
         directory (str): directory to search
         patt (str): pattern
@@ -362,14 +360,14 @@ def findfiles(directory, pattern=None):
         list: list of matched files
     """
     lst = os.listdir(directory)
-    if not pattern is None:
+    if pattern is not None:
         rr = re.compile(pattern)
         lst = [l for l in lst if re.match(rr, l)]
     return lst
 
 
 def finddirectories(directory, pattern=None):
-    """ Get a list of directories 
+    """ Get a list of directories
     Args:
         directory (str): directory to search
         patt (str): pattern
@@ -377,7 +375,7 @@ def finddirectories(directory, pattern=None):
         list: list of matched directories
     """
     lst = os.listdir(directory)
-    if not pattern is None:
+    if pattern is not None:
         rr = re.compile(pattern)
         lst = [l for l in lst if re.match(rr, l)]
     lst = [l for l in lst if os.path.isdir(os.path.join(directory, l))]
@@ -400,7 +398,7 @@ def oaIsBinary(filename):
 
 
 def fac(n):
-    """ Return n! (factorial) 
+    """ Return n! (factorial)
 
     Args:
         n (int): number to calculate factorial
@@ -528,7 +526,7 @@ def runcommand(cmd, dryrun=0, idstr=None, verbose=1, logfile=None, shell=True):
     Returns:
         r (int): return value of the shell command
     """
-    if not idstr is None:
+    if idstr is not None:
         cmd = 'echo "idstr: %s";\n' % idstr + cmd
[…]
