### 2018-11-23 17ad2e2
add double_conference_foldover_permutation
