### 2018-11-25 cb2ec0e
cleanup code
