### 2018-11-26 #1097 @danielskatz
👋 @djmitche, @tanaken-basis - thanks for agreeing to review this submission.  Please see the instructions above and feel free to ask me if you have any questions.  Otherwise, I will assume that you will work through the checklist items above, and will bring up any issues with the submission with the author (via very short discussion here or via new issues in [the source repo](https://github.com/eendebakpt/oapackage)), who will resolve them.
