### 2018-11-26 #1097 @djmitche
It's worth noting that the most recent version is 2.5.2 now.  That is the same in the repo and the package repository.
