### 2018-11-27 #1097 @peendebak
Any updates to the package to improve based on review comments will go to the `dev` branch. On a regular basis I will merge `dev` to `master` and update the PyPi packages.
