### 2018-11-28 c2e1ad9
cleanup code; update docs
