### 2018-12-02 e651e25
cleanup code
