### 2018-12-03 a3376e1
Add functionality to reduce double conference designs to single conference designs (#86)
* add double_conference_foldover_permutation

* autopep

* small fixes

* remove print statement

* add example arrays

* cleanup code

* cleanup code; update docs

* cleanup code
