### 2018-12-04 4516182
update paper with comments from alan
