+++ paper/paper.bib
@@ -9,9 +9,17 @@
 title={A Classification Criterion for Definitive Screening Designs},
 author={Eric D. Schoen and Pieter Thijs Eendebak and Peter Goos},
 journal = {The Annals of Statistics},
-year={2018}
+year={2018},
+note={Published online}
 }
 
+@Article{Goos2011,
+ title                = {Optimal Design of Experiments: A Case Study Approach},
+ author               = {Goos, Peter and Jones, Bradley},
+ year                 = 2011,
+ month                = jun,
+ doi                  = {10.1002/9781119974017}
+}
 
 @misc{Eendebak2016rank,
   author = {P.T. Eendebak},
@@ -23,22 +31,19 @@ year={2018}
   title = {Two-Level Designs to Estimate All Main Effects and Two-Factor Interactions},
   author = {Pieter T. Eendebak and Eric D. Schoen},
   journal = {Technometrics},
-  year = {2016},
-  number = {ja},
-  pages = {1-30},
-  volume = {0},
+  year = {2017},
+  number = {1},
+  pages = {69-79},
+  volume = {59},
   doi = {10.1080/00401706.2016.1142903},
   eprintx = { http://dx.doi.org/10.1080/00401706.2016.1142903},
-  url = { http://dx.doi.org/10.1080/00401706.2016.1142903},
-  urlx = {files/Eendebak2015technometrics.pdf}
+  url = {http://dx.doi.org/10.1080/00401706.2016.1142903},
 }
 @misc{EendebakDOF,
-  author = {P.T. Eendebak},
+  author = {Eendebak, P.T.},
   year = {2014},
   title = {A canonical form for non-regular arrays based on generalized worldlength pattern values of delete-one-factor projections},
   note = {Working paper University of Antwerp},
-  url2 = {http://www.pietereendebak.nl/oapage/index.html},
-  url = {files/Eendebak2014-delete-one-factor.pdf}
 }
 @article{Eendebak2009,
   author = {Schoen, Eric D. and Eendebak, Pieter T. and Nguyen, Man V. M.},
@@ -127,7 +132,7 @@ eprint = {/oup/backfile/content_public/journal/biomet/82/3/10.1093/biomet/82.3.5
 author = "Tang, Boxin and Deng, Lih-Yuan",
 doi = "10.1214/aos/1017939244",
 fjournal = "The Annals of Statistics",
-journal = "Ann. Statist.",
+journal = "The Annals of Statistics",
 month = "12",
 number = "6",
 pages = "1914--1926",
@@ -147,7 +152,7 @@ year = "1999"
  number = {4},
  pages = {1071--1082},
  publisher = {Institute of Statistical Science, Academia Sinica},
- title = {GENERALIZED RESOLUTION AND MINIMUM ABERRATION CRITERIA FOR PLACKETT-BURMAN AND OTHER NONREGULAR FACTORIAL DESIGNS},
+ title = {Generalized resolution and minimum Aberration Criteria for {P}lackett-{B}urman and other nonregular factorial designs},
  volume = {9},
  year = {1999}
 }
