+++ paper/paper.bib
@@ -65,7 +65,7 @@ note={Published online}
 
 @misc{OAdocumentation,
     author = "P.T. Eendebak and Alan Roberto Vazquez",
-    title = "The Orthogonal Array package",
+    title = "The {O}rthogonal {A}rray package",
     year = "2018",
     url = "https://oapackage.readthedocs.io/en/latest/",
   }
