### 2018-12-04 #1097 @arfon
> @whedon generate pdf from branch feat/updates_paper_alan

Just a heads up @eendebakpt - compiling from a branch doesn't seem to work very well. You might have to merge to `master` to see new changes with @whedon.
