+++ paper/paper.bib
@@ -152,7 +152,7 @@ year = "1999"
  number = {4},
  pages = {1071--1082},
  publisher = {Institute of Statistical Science, Academia Sinica},
- title = {Generalized resolution and minimum Aberration Criteria for {P}lackett-{B}urman and other nonregular factorial designs},
+ title = {Generalized resolution and minimum aberration criteria for {P}lackett-{B}urman and other nonregular factorial designs},
  volume = {9},
  year = {1999}
 }
