### 2018-12-06 #1097 @djmitche
This is my first JOSS review so no, I don't have ideas :/
