### 2018-12-06 1f5ff0f
Small update to paper (#89)
* small update to paper

* bump version
