+++ docs/references.bib
@@ -97,7 +97,7 @@ note={Published online}
 @misc{EendebakOA,
 author={P. T. Eendebak},
 year={2012},
-title={{O|rthogonal {A}rray page},
+title={{O}rthogonal {A}rray page},
 url={http://www.pietereendebak.nl/oapage/index.html}
 }
 
