### 2018-12-08 efcd20e
throw error for incorrent use of jstructconference_t
