### 2018-12-08 a558a10
add tests
