### 2018-12-10 2a8c49f
cleanup outout of unit tests
