+++ oapackage/tests/test_cpplibrary.py
@@ -20,6 +20,15 @@ def is_sorted(l):
     return all(a <= b for a, b in zip(l, l[1:]))
 
 
+def only_python3(function):
+    if python3:
+        def only_python3_function(*args, **kwargs):
+            return function(*args, **kwargs)
+    else:
+        def only_python3_function(*args, **kwargs):
+            return None
+    return only_python3_function
+    
 import oapackage
 
 
@@ -43,16 +52,21 @@ class TestReductions(unittest.TestCase):
         pvalues = [pvalue.raw_values() for pvalue in dof_values]
         self.assertTrue(is_sorted(pvalues))
 
-        al = oapackage.exampleArray(7, 1)
+        al = oapackage.exampleArray(7, 0)
         transformation = oapackage.reductionDOP(al)
         self.assertTrue(transformation.isIdentity())
 
-        al = oapackage.exampleArray(8, 1)
+        al = oapackage.exampleArray(8, 0)
         transformation = oapackage.reductionDOP(al)
-        transformation.show()
+        
+        with mock.patch('sys.stdout', new_callable=io.StringIO) as mock_stdout:
+            transformation.show()
+            stdout = mock_stdout.getvalue()
+            self.assertTrue(stdout.startswith('array transformation: N 40'))
+            self.assertIn('column permutation: {2,0,3,1,4,5,6}', stdout)
         alr = oapackage.reduceDOPform(al)
         self.assertTrue(transformation.apply(al) == alr)
-
+            
 
 class TestModelmatrix(unittest.TestCase):
 
@@ -66,7 +80,7 @@ class TestModelmatrix(unittest.TestCase):
         model_matrix = oapackage.array2modelmatrix(al, "main", 1)
         np.testing.assert_array_equal(2 * np.array(al) - 1, model_matrix[:, 1:])
 
-        conf_design = oapackage.exampleArray(41, 1)
+        conf_design = oapackage.exampleArray(41, 0)
         sizes = oapackage.array2modelmatrix_sizes(conf_design)
         k = conf_design.n_columns
         self.assertEqual(sizes, (1, 1 + k, int(1 + k + k * (k - 1) / 2), int(1 + k + k * (k + 1) / 2)))
@@ -84,11 +98,11 @@ class TestModelmatrix(unittest.TestCase):
 class TestArrayLink(unittest.TestCase):
 
     def test_selectFirstColumns(self):
[…]
