### 2018-12-10 69afa29
Cleanup jstructconference_t (#90)
* throw error for incorrent use of jstructconference_t

* add tests

* cleanup outout of unit tests

* update tests
