### 2018-12-11 3895c52
cleanup code; improve docs
