### 2018-12-11 1011adc
cleanup code
