### 2018-12-12 b4e1ce4
cleanup code
