### 2018-12-16 c75e14e
refactor code (#91)
* refactor code

* cleanup code; improve docs

* cleanup code

* cleanup code
