+++ src/arraytools.cpp
@@ -9,6 +9,7 @@
 #include "mathtools.h"
 #include "strength.h"
 #include "tools.h"
+#include <errno.h>
 
 #ifdef RPACKAGE
 #define printf notallowed
@@ -935,6 +936,22 @@ void showArrayList (const arraylist_t &lst) {
                 myprintf ("array %d:\n", (int)i);
                 lst[i].showarray ();
         }
+}
+
+template < class atype >
+/// write array to output stream
+void write_array_format(std::ostream &ss, const atype *array, const int nrows, const int ncols, int width = 3) {
+	assert(array != 0 || ncols == 0);
+
+	int count;
+	for (int j = 0; j < nrows; j++) {
+		count = j;
+		for (int k = 0; k < ncols; k++) {
+			const char *s = (k < ncols - 1) ? " " : "\n";
+			ss << std::setw(width) << array[count] << s;
+			count += nrows;
+		}
+	}
 }
 
 #ifdef FULLPACKAGE
@@ -4482,6 +4499,13 @@ int arrayfile_t::read_array (array_t *array, const int nrows, const int ncols) {
         return index;
 }
 
+void arrayfile_t::finisharrayfile() {
+	if (this->mode == ATEXT) {
+		fprintf(this->nfid, "-1\n");
+	}
+	this->closefile();
+}
+
 void arrayfile_t::writeheader () {
         assert (this->nfid);
 
@@ -4600,6 +4624,14 @@ long nArrays (const char *fname) {
         return n;
 }
 
+void arrayfileinfo(const char *filename, int &number_of_arrays, int &number_of_rows, int &number_of_columns) {
+	arrayfile_t af(filename, 0);
+	number_of_arrays = af.narrays;
+	number_of_rows = af.nrows;
+	number_of_columns = af.ncols;
+	af.closefile();
+}
+
 void appendArrays(const arraylist_t &arrays_to_append, arraylist_t &dst) {
 	for (arraylist_t::const_iterator it = arrays_to_append.begin(); it < arrays_to_append.end(); ++it) {
 		dst.push_back(*it);
@@ -4661,13 +4693,8 @@ arraylist_t readarrayfile (const char *fname, int verbose, int *setcols) {
         return v;
 }
 
-/**
- * @brief Read all arrays in a file and store then in an array list
- * @param fname
- * @param arraylist
- * @return
- */
-int readarrayfile (const char *fname, arraylist_t *arraylist, int verbose, colindex_t *setcols, rowindex_t *setrows,
+
+int readarrayfile (const char *fname, arraylist_t *arraylist, int verbose, int *setcols, int *setrows,
                    int *setbits) {
         if (arraylist == 0) {
                 arraylist = new arraylist_t;
@@ -5032,34 +5059,6 @@ arrayfile_t::arrayfile_t (const std::string fnamein, int verbose) {
         }
 }
 
-/*!
-  save_arrays writes all the arrays from solutions to a file. The filename is obtained from the number of processors,
-  the number of factors and the number of columns so far. Then the file header contains the number of columns in the
-  design,
-  the number of runs and the number of arrays in the file.
-  \brief Saves the arrays from solutions
-  \param solutions List of arrays
-  \param p Characteristic numbers of OA
-  \param n_arrays Number of arrays found
-  \param n_procs Number of processors in system, for filename
-  */
-int save_arrays (arraylist_t &solutions, const arraydata_t *ad, const int n_arrays, const int n_procs,
-                 const char *resultprefix, arrayfile::arrayfilemode_t mode) {
[…]
