+++ src/arraytools.h
@@ -1118,7 +1118,8 @@ namespace arrayfile {
 enum arrayfilemode_t {
 	/// text based format
         ATEXT,
-        ALATEX,
+	/// write arrays to a text file in a format that can be parsed by LaTeX
+		ALATEX,
 	/// binary format
         ABINARY,
 	/// binary format storing differences of arrays
@@ -1258,12 +1259,7 @@ struct arrayfile_t {
         /// read array and return index
         int read_array (array_t *array, const int nrows, const int ncols);
 
-        void finisharrayfile () {
-                if (this->mode == ATEXT) {
-                        fprintf (this->nfid, "-1\n");
-                }
-                this->closefile ();
-        }
+		void finisharrayfile();
 
         void setVerbose (int v) { this->verbose = v; }
 
@@ -1324,25 +1320,30 @@ long nArrays (const char *fname);
 
 /** return information about file with arrays
  *
- * \param fname Filename of array file
+ * \param filename Filename of array file
  * \param number_of_arrays Variable is set with number of arrays
  * \param number_of_rows Variable is set with number of rows
  * \param number_of_columns Variable is set with number of columns
  */
-inline void arrayfileinfo (const char *fname, int &number_of_arrays, int &number_of_rows, int &number_of_columns) {
-        arrayfile_t af (fname, 0);
-        number_of_arrays = af.narrays;
-        number_of_rows = af.nrows;
-        number_of_columns = af.ncols;
-        af.closefile ();
-}
+void arrayfileinfo(const char *filename, int &number_of_arrays, int &number_of_rows, int &number_of_columns);
 
-/// read list of arrays from file and append to list
-int readarrayfile (const char *fname, arraylist_t *arraylist, int verbose = 1, int *setcols = 0,
-                   rowindex_t *setrows = 0, int *setbits = 0);
-
-/// read list of arrays from file
+/** Read all arrays in a file
+*
+* @param fname Filename to read from
+* @param verbose Verbosity level
+* @param setcols Pointer to return number of columns from array file
+* @return List of arrays
+*/
 arraylist_t readarrayfile (const char *fname, int verbose = 1, int *setcols = 0);
+
+/** Read all arrays in a file and append then to an array list
+*
+* @param fname Filename to read from
+* @param arraylist Pointer to list of arrays
+* @return
+*/
+int readarrayfile(const char *fname, arraylist_t *arraylist, int verbose = 1, int *setcols = 0,
+	int *setrows = 0, int *setbits = 0);
 
 const int NRAUTO = 0;
 /// write a list of arrays to file on disk
@@ -1350,14 +1351,14 @@ int writearrayfile (const char *fname, const arraylist_t *arraylist,
                     arrayfile::arrayfilemode_t mode = arrayfile::ATEXT, int nrows = NRAUTO, int ncols = NRAUTO);
 
 /// write a list of arrays to file on disk
-int writearrayfile (const char *fname, const arraylist_t arraylist, arrayfile::arrayfilemode_t mode = arrayfile::ATEXT,
+int writearrayfile (const char *filename, const arraylist_t arraylist, arrayfile::arrayfilemode_t mode = arrayfile::ATEXT,
                     int nrows = NRAUTO, int ncols = NRAUTO);
 
 /// write a single array to file
-int writearrayfile (const char *fname, const array_link &al, arrayfile::arrayfilemode_t mode = arrayfile::ATEXT);
+int writearrayfile (const char *filename, const array_link &array, arrayfile::arrayfilemode_t mode = arrayfile::ATEXT);
 
 /// append a single array to an array file. creates a new file if no file exists
-int appendarrayfile (const char *fname, const array_link al);
+int append_arrayfile (const char *filename, const array_link array);
 
[…]
