### 2018-12-16 #1097 @danielskatz
👋 @gsagnol - how are things going?  When will you be able to start on the review (at least the first 2 checkboxes regarding COIs and the JOSS CoC?)
