### 2018-12-17 #1097 @gsagnol
Hi @danielskatz , I wanted to start with the review this week.

I tried to check some of the boxes, but this will not work, although I'm logged in to github. Is google-chrome under linux not supported ? I tried under firefox, but this did not work neither.
