### 2018-12-17 #1097 @arfon
@gsagnol - I've re-made the invitation here: https://github.com/openjournals/joss-reviews/invitations
