### 2018-12-18 #1097 @gsagnol
I installed the latest release from github, and `oapackage.__version__` returns 2.5.2, while the check list above asks for 2.5.1. Should the checklist be updated, or should I review a former version ?
