### 2018-12-18 #1097 @peendebak
@gsagnol I don't know about JOSS policy on this matter. I submitted version 2.5.1 to JOSS, but since that moment (also based on reviewers comments) I updated the code and the paper and made a new release. I will continue to push updates based on the review to `dev` and once in a while merge `dev` to `master` (with an increase in version number)
