### 2018-12-18 eendebakpt/oapackage#92 @gsagnol
Location of Documentation
Is the documentation of OApackage available online ? I see there are several notebooks with examples in this repository, but they do not seem to be listed on the webpage http://pietereendebak.nl/oapackage/index.html

Also, there does not seem to be an API documentation which lists the features of OApackage. The closest thing I could find is the following pdf file `http://www.pietereendebak.nl/oapackage/files/OApackage.pdf`, but this pdf was generated more than 3 years ago, and I had to use google to find it. Or is there location other than http://pietereendebak.nl/oapackage/index.html for the actual doc of OApackage ?
