### 2018-12-18 eendebakpt/oapackage#92 @gsagnol
OK, finally I found the webpage for the documentation:
https://oapackage.readthedocs.io/en/latest/index.html

But I think you should update the link on the homepage of this repository, and also add a link from your "oapackage" website to the doc webpage.
