### 2018-12-18 #1097 @gsagnol
@peendebak : I posted an issue on the OApackage repository about confusing location for the documentation (readthedocs.io vs. http://pietereendebak.nl)
