### 2018-12-18 #1097 @peendebak
> @peendebak : I would like to have a look at your unit tests. I tried to run
> `coverage run --source='./oapackage' -m pytest`,
> but numpy raises some error:
> ImportError: /homes/combi/sagnol/.local/lib/python2.7/site-packages/numpy/core/multiarray.so: undefined symbol: _Py_ZeroStruct
>
> Do I need a particular version of numpy to execute the testing suite ? Or is there another way to run the unit tests ?

The minimal version of numpy is specified in the `setup.py` (1.13, although perhaps earlier versions of numpy work as well). Can you provide me with some more details of your system (e.g. platform, version of numpy installed, how did you install (pip or compile from source), send output of installation process, version of swig installed)

Can you also send me the output of the following commands?
```
python -c "import numpy; print(numpy); print(numpy.__version__')
python -c "import oapackage; print(oapackage)"
pytest - v
```
@gsagnol
