### 2018-12-18 29e02ee
add tests for conference statistics=a
