### 2018-12-18 5a6cabc
add tests for conference statistics
