+++ docs/references.bib
@@ -58,12 +58,13 @@ url={http://www.pietereendebak.nl/oapage/},
 note={\url{http://www.pietereendebak.nl/oapage/}}
 }
 
-@TECHREPORT{EendebakOApackage,
+@TECHREPORT{EendebakOApackageResults,
 author={P. T. Eendebak},
-year={2013},
-title={The {O}rthogonal {A}rray package},
-url={http://www.pietereendebak.nl/oapage/software.html}
+year={2018},
+title={The {O}rthogonal {A}rray package: results},
+url={http://www.pietereendebak.nl/oapackage/}
 }
+
 
 @misc{EendebakDOF,
 author={P. T. Eendebak},
