### 2018-12-19 eendebakpt/oapackage#95 @gsagnol
Problem with testing
Here is the output of the following commands on my PC:

    $ python -c "import numpy; print(numpy); print(numpy.__version__)"
    <module 'numpy' from '/homes/-hidden-/.local/lib/python2.7/site-packages/numpy/__init__.pyc'>
    1.15.4

    $ python -c "import oapackage; print(oapackage)"
    <module 'oapackage' from '/homes/-hidden-/.local/lib/python2.7/site-packages/oapackage/__init__.pyc'>

    $ pytest -v
platform linux -- Python 3.6.5, pytest-3.3.2, py-1.5.2, pluggy-0.6.0 -- /usr/bin/python3
cachedir: .cache
rootdir: /homes/-hidden-/.local/lib/python2.7/site-packages/oapackage, inifile:
collected 0 items / 2 errors

==================================== ERRORS ====================================
__________________ ERROR collecting tests/test_cpplibrary.py ___________________
ImportError while importing test module '/homes/-hidden-/.local/lib/python2.7/site-packages/oapackage/tests/test_cpplibrary.py'.
Traceback:
../oalib.py:17: in swig_import_helper
    return importlib.import_module(mname)
/usr/lib64/python3.6/importlib/__init__.py:126: in import_module
<frozen importlib._bootstrap>:994: in _gcd_import
    ???
<frozen importlib._bootstrap>:971: in _find_and_load
    ???
<frozen importlib._bootstrap>:955: in _find_and_load_unlocked
    ???
<frozen importlib._bootstrap>:658: in _load_unlocked
    ???
<frozen importlib._bootstrap>:571: in module_from_spec
    ???
<frozen importlib._bootstrap_external>:922: in create_module
    ???
<frozen importlib._bootstrap>:219: in _call_with_frames_removed
    ???
E   ImportError: dynamic module does not define module export function (PyInit__oalib)

__init__.py:10: in <module>
    import oalib
../oalib.py:20: in <module>
    _oalib = swig_import_helper()
../oalib.py:19: in swig_import_helper
    return importlib.import_module('_oalib')
/usr/lib64/python3.6/importlib/__init__.py:126: in import_module
E   ImportError: dynamic module does not define module export function (PyInit__oalib)
___________________ ERROR collecting tests/test_oapackage.py ___________________
ImportError while importing test module '/homes/-hidden-/.local/lib/python2.7/site-packages/oapackage/tests/test_oapackage.py'.
Traceback:
../oalib.py:17: in swig_import_helper
    return importlib.import_module(mname)
/usr/lib64/python3.6/importlib/__init__.py:126: in import_module
<frozen importlib._bootstrap>:994: in _gcd_import
    ???
<frozen importlib._bootstrap>:971: in _find_and_load
    ???
<frozen importlib._bootstrap>:955: in _find_and_load_unlocked
    ???
<frozen importlib._bootstrap>:658: in _load_unlocked
    ???
<frozen importlib._bootstrap>:571: in module_from_spec
    ???
<frozen importlib._bootstrap_external>:922: in create_module
    ???
<frozen importlib._bootstrap>:219: in _call_with_frames_removed
    ???
E   ImportError: dynamic module does not define module export function (PyInit__oalib)

__init__.py:10: in <module>
    import oalib
../oalib.py:20: in <module>
    _oalib = swig_import_helper()
../oalib.py:19: in swig_import_helper
    return importlib.import_module('_oalib')
/usr/lib64/python3.6/importlib/__init__.py:126: in import_module
E   ImportError: dynamic module does not define module export function (PyInit__oalib)
!!!!!!!!!!!!!!!!!!! Interrupted: 2 errors during collection !!!!!!!!!!!!!!!!!!!!
=========================== 2 error in 0.57 seconds ============================
