### 2018-12-19 eendebakpt/oapackage#95 @gsagnol
I also installed `OApackage` on my laptop (with a global installation), and this time the tests work, but with several failures:

    $ pytest -v

==================================================== test session starts =====================================================
platform linux2 -- Python 2.7.15rc1, pytest-3.3.2, py-1.5.2, pluggy-0.6.0 -- /usr/bin/python2
cachedir: .cache
rootdir: /usr/local/lib/python2.7/dist-packages/oapackage, inifile:
collected 70 items

tests/test_cpplibrary.py::TestReductions::test_DOP PASSED                                                              [  1%]
tests/test_cpplibrary.py::TestReductions::test_LMC PASSED                                                              [  2%]
tests/test_cpplibrary.py::TestModelmatrix::test_modelmatrix PASSED                                                     [  4%]
tests/test_cpplibrary.py::TestArrayLink::test_array_class_functions PASSED                                             [  5%]
tests/test_cpplibrary.py::TestArrayLink::test_array_link_dimensions PASSED                                             [  7%]
tests/test_cpplibrary.py::TestArrayLink::test_basic_array_link_functionality PASSED                                    [  8%]
tests/test_cpplibrary.py::TestArrayLink::test_is_ortogonal_array PASSED                                                [ 10%]
tests/test_cpplibrary.py::TestArrayLink::test_selectFirstColumns PASSED                                                [ 11%]
tests/test_cpplibrary.py::TestArraydata_t::test_factor_levels PASSED                                                   [ 12%]
tests/test_cpplibrary.py::TestConferenceDesigns::test_conf2dsd PASSED                                                  [ 14%]
tests/test_cpplibrary.py::TestArrayFiles::test_write_latex_format PASSED                                               [ 15%]
tests/test_cpplibrary.py::TestCppLibrary::test_array2eigenModelMatrixMixed PASSED                                      [ 17%]
tests/test_cpplibrary.py::TestCppLibrary::test_array_transformation_t PASSED                                           [ 18%]
tests/test_cpplibrary.py::TestCppLibrary::test_arraylink PASSED                                                        [ 20%]
tests/test_cpplibrary.py::TestCppLibrary::test_arraylink2arraydata PASSED                                              [ 21%]
tests/test_cpplibrary.py::TestCppLibrary::test_arraylink_slicing PASSED                                                [ 22%]
tests/test_cpplibrary.py::TestCppLibrary::test_arrayrankInfo PASSED                                                    [ 24%]
tests/test_cpplibrary.py::TestCppLibrary::test_conference_generation PASSED                                            [ 25%]
tests/test_cpplibrary.py::TestCppLibrary::test_exception_handling hiFAILED                                               [ 27%]
tests/test_cpplibrary.py::TestCppLibrary::test_mvalue_t PASSED                                                         [ 28%]
tests/test_cpplibrary.py::TestCppLibrary::test_mycheck_handler blaPASSED                                                  [ 30%]
tests/test_cpplibrary.py::TestCppLibrary::test_projectionDOFvalues PASSED                                              [ 31%]
tests/test_cpplibrary.py::TestCppLibrary::test_projection_efficiencies PASSED                                          [ 32%]
tests/test_cpplibrary.py::TestCppLibrary::test_rankStructure PASSED                                                    [ 34%]
tests/test_cpplibrary.py::TestCppLibrary::test_splits PASSED                                                           [ 35%]
tests/test_oapackage.py::TestMisc::test_exampleArray PASSED                                                            [ 37%]
tests/test_oapackage.py::TestMisc::test_oa2graph PASSED                                                                [ 38%]
tests/test_oapackage.py::TestMisc::test_reduceGraphNauty PASSED                                                        [ 40%]
tests/test_oapackage.py::TestMisc::test_scanf FAILED                                                                   [ 41%]
tests/test_oapackage.py::test_numpy_interface PASSED                                                                   [ 42%]
tests/test_oapackage.py::test_nauty PASSED                                                                             [ 44%]
tests/test_oapackage.py::TestOAfiles::test_checkArrayFile PASSED                                                       [ 45%]
tests/test_oapackage.py::TestOAfiles::test_finddirectories PASSED                                                      [ 47%]
tests/test_oapackage.py::TestOAfiles::test_findfiles PASSED                                                            [ 48%]
tests/test_oapackage.py::TestOAfiles::test_findfilesR PASSED                                                           [ 50%]
tests/test_oapackage.py::TestOAfiles::test_misc_file_operations PASSED                                                 [ 51%]
tests/test_oapackage.py::TestOAfiles::test_nArrayFile PASSED                                                           [ 52%]
tests/test_oapackage.py::TestOAfiles::test_selectArrays PASSED                                                         [ 54%]
tests/test_oapackage.py::TestParetoFunctionality::test_selectParetoArrays FAILED                                       [ 55%]
tests/test_oapackage.py::TestOAhelper::test_argsort PASSED                                                             [ 57%]
tests/test_oapackage.py::TestOAhelper::test_array2html PASSED                                                          [ 58%]
tests/test_oapackage.py::TestOAhelper::test_array2latex PASSED                                                         [ 60%]
tests/test_oapackage.py::TestOAhelper::test_bounds PASSED                                                              [ 61%]
tests/test_oapackage.py::TestOAhelper::test_checkFiles PASSED                                                          [ 62%]
tests/test_oapackage.py::TestOAhelper::test_checkFilesOA PASSED                                                        [ 64%]
tests/test_oapackage.py::TestOAhelper::test_choose PASSED                                                              [ 65%]
tests/test_oapackage.py::TestOAhelper::test_create_pareto_element PASSED                                               [ 67%]
tests/test_oapackage.py::TestOAhelper::test_deprecated FAILED                                                          [ 68%]
tests/test_oapackage.py::TestOAhelper::test_designStandardError FAILED                                                 [ 70%]
tests/test_oapackage.py::TestOAhelper::test_fac PASSED                                                                 [ 71%]
tests/test_oapackage.py::TestOAhelper::test_floatformat PASSED                                                         [ 72%]
tests/test_oapackage.py::TestOAhelper::test_formatC PASSED                                                             [ 74%]
tests/test_oapackage.py::TestOAhelper::test_gwlp2str PASSED                                                            [ 75%]
tests/test_oapackage.py::TestOAhelper::test_joinArrayLists PASSED                                                      [ 77%]
tests/test_oapackage.py::TestOAhelper::test_parseProcessingTime PASSED                                                 [ 78%]
tests/test_oapackage.py::TestOAhelper::test_plot2Dline PASSED                                                          [ 80%]
tests/test_oapackage.py::TestOAhelper::test_runExtend PASSED                                                           [ 81%]
tests/test_oapackage.py::TestOAhelper::test_safemin_safemax PASSED                                                     [ 82%]
tests/test_oapackage.py::TestOAhelper::test_sortcols PASSED                                                            [ 84%]
tests/test_oapackage.py::TestOAhelper::test_sortrows PASSED                                                            [ 85%]
tests/test_oapackage.py::TestOAhelper::test_testHtml PASSED                                                            [ 87%]
tests/test_oapackage.py::TestDoptimize::test_array2Dtable PASSED                                                       [ 88%]
tests/test_oapackage.py::TestDoptimize::test_calcScore PASSED                                                          [ 90%]
tests/test_oapackage.py::TestDoptimize::test_custom_optim PASSED                                                       [ 91%]
tests/test_oapackage.py::TestDoptimize::test_filterPareto PASSED                                                       [ 92%]
tests/test_oapackage.py::TestDoptimize::test_generateDpage PASSED                                                      [ 94%]
tests/test_oapackage.py::TestDoptimize::test_generateDscatter PASSED                                                   [ 95%]
tests/test_oapackage.py::TestDoptimize::test_optimDeffPython PASSED                                                    [ 97%]
tests/test_oapackage.py::TestDoptimize::test_runcommand PASSED                                                         [ 98%]
tests/test_oapackage.py::TestDoptimize::test_unittest PASSED                                                           [100%]

========================================================== FAILURES ==========================================================
___________________________________________ TestCppLibrary.test_exception_handling ___________________________________________

self = <oapackage.tests.test_cpplibrary.TestCppLibrary testMethod=test_exception_handling>

    def test_exception_handling(self):
        with self.assertRaises(RuntimeError):
            oapackage.mycheck_handler("file", "function", 10, 0, "hi")
        oapackage.mycheck_handler("file", "function", 10, 1, "hi")

        with self.assertRaises(RuntimeError):
            oapackage.throw_runtime_exception("dsfs")

        al = oapackage.oalib.exampleArray(18, 1)
>       with self.assertRaisesRegex(RuntimeError, "array cannot have negative elements"):
E       AttributeError: 'TestCppLibrary' object has no attribute 'assertRaisesRegex'

tests/test_cpplibrary.py:240: AttributeError
---------------------------------------------------- Captured stdout call ----------------------------------------------------
mycheck: file: function (line 10): exampleArray: conference matrix of size 16, 7
____________________________________________________ TestMisc.test_scanf _____________________________________________________

self = <oapackage.tests.test_oapackage.TestMisc testMethod=test_scanf>

    def test_scanf(self):
>       r = oapackage.scanf.sscanf('1', '%d')

tests/test_oapackage.py:53:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

inputString = '1', formatString = '%d'

    def sscanf(inputString, formatString):
        """ Simple scanf function """
>       return _scanfpip.scanf(formatString, inputString)
E       AttributeError: 'module' object has no attribute 'scanf'

scanf.py:6: AttributeError
______________________________________ TestParetoFunctionality.test_selectParetoArrays _______________________________________

self = <oapackage.tests.test_oapackage.TestParetoFunctionality instance at 0x7ffb0214b200>

    def test_selectParetoArrays(self):

        arrays = [oapackage.array_link(np.array([[ii]])) for ii in range(5)]
        pareto_object = oapackage.ParetoLongLong()

        for ii in range(len(arrays)):
            value = [ii, ii % 2]
            pareto_object.addvalue(value, ii)
        pareto_object.show(2)

        selected = oapackage.oahelper.selectParetoArrays(arrays, pareto_object)
>       self.assertEqual(selected, arrays[4, 5])
E       AttributeError: TestParetoFunctionality instance has no attribute 'assertEqual'

tests/test_oapackage.py:241: AttributeError
---------------------------------------------------- Captured stdout call ----------------------------------------------------
Pareto: 2 optimal values, 2 objects
value 0: 3; 1
value 1: 4; 0
________________________________________________ TestOAhelper.test_deprecated ________________________________________________

self = <oapackage.tests.test_oapackage.TestOAhelper testMethod=test_deprecated>

    def test_deprecated(self):
        def func():
            return 'hi'
        deprecated_function = oapackage.oahelper.deprecated(func)
>       with self.assertWarns(Warning):
E       AttributeError: 'TestOAhelper' object has no attribute 'assertWarns'

tests/test_oapackage.py:290: AttributeError
---------------------------------------------------- Captured stdout call ----------------------------------------------------
exampleArray: array 3 in OA(16, 2, 2^5)
___________________________________________ TestOAhelper.test_designStandardError ____________________________________________

self = <oapackage.tests.test_oapackage.TestOAhelper testMethod=test_designStandardError>

    def test_designStandardError(self):
        al = oapackage.exampleArray(14, 0)
        v = oapackage.oahelper.designStandardError(al)
>       self.assertAlmostEqual(v[0], 0.3747931073686535)
E       AssertionError: 0.2288887948318749 != 0.3747931073686535 within 7 places

tests/test_oapackage.py:363: AssertionError
---------------------------------------------------- Captured stdout call ----------------------------------------------------
exampleArray: array 3 in OA(16, 2, 2^5)
====================================================== warnings summary ======================================================
tests/test_oapackage.py::TestOAhelper::test_designStandardError
  /usr/local/lib/python2.7/dist-packages/oapackage/oahelper.py:1167: RuntimeWarning: invalid value encountered in sqrt
    return np.sqrt(m0), np.sqrt(m1), np.sqrt(m2)

None
  could not create cache path /usr/local/lib/python2.7/dist-packages/oapackage/.cache/v/cache/lastfailed

-- Docs: http://doc.pytest.org/en/latest/warnings.html
====================================== 5 failed, 65 passed, 2 warnings in 0.81 seconds =======================================
