### 2018-12-19 #1097 @gsagnol
@peendebak I created an issue on the oapackage repo for the tests.
* On the PC of my institute, I am not able to run the tests. Here, I used a local information (pip install --user)
* On my laptop, the testing suite seems to work, but some tests fail. On my laptop I used a standard global installation without the --user flag.
