### 2018-12-19 eendebakpt/oapackage#95 @eendebakpt
@gsagnol About the first issue (on your PC). It seems the python 2.7 and 3.6.5 distrubutions are mixed up. Since I have no access to your system, it is hard for me to explain how this could happen.

I will try to make a few suggestions. In the output of `pytest -v` it reads
```
...
ImportError while importing test module '/homes/-hidden-/.local/lib/python2.7/site-packages/oapackage/tests/test_cpplibrary.py'
...
./oalib.py:19: in swig_import_helper
return importlib.import_module('_oalib')
/usr/lib64/python3.6/importlib/init.py:126: in import_module
...
```
The first which means that the `oapackage` was installed (and compiled) as version 2.7.  The last lines show that `oalib.py` tries to use the python 3.6 `importlib`.

It appears you are running linux (or at least some unix variation). Are there besides `pytest` also `pytest-2`  and `pytest-2` available. If so, can you run `pytest-2`?

Can you also try to create a [virtualenv ](https://virtualenv.pypa.io/en/latest/) environment and the the oapackage there?
