### 2018-12-19 8bb5067
include scanf code to prevent import conflicts with other scanf packages
