### 2018-12-19 150cd19
Add unit tests for conference design statistics (#94)
* add tests for conference statistics=a

* add tests for conference statistics
