### 2018-12-19 b98565a
fix unittests
