### 2018-12-19 4fdc8f7
for python 2.7 add explicit warning if mock is not installed
