### 2018-12-19 da7cf0a
fix exception definition for python 2.7
