### 2018-12-19 ccbe4ee
fix exception definition for python 2.7
