### 2018-12-19 7cfb40f
add mock to appveyor
