### 2018-12-19 64d89d2
fix appveyor tests
