### 2018-12-19 54ce614
fix appveyor file
