### 2018-12-20 eendebakpt/oapackage#95 @eendebakpt
@gsagnol In the second report there seem to be multiple things going on. Some of them I fixed in #96, but for the others I need some more information. Can you run for me the following python script and send me the output:
```
import numpy as np; np.show_config()

import oapackage
al = oapackage.exampleArray(14, 1)
al.showarray()
print('Defficiencies: %s' % (al.Defficiencies(), ) )

scalefac=1
X=al.getModelMatrix(2)
print('model matrix:')
print(X)
M = np.linalg.inv(X.transpose().dot(X) / scalefac)
print('information matrix:')
print(M)
v = oapackage.oahelper.designStandardError(al)
print('output of designStandardError')
print(v)
print(v[0])
```
And also the following script:
```
import unittest
print(unittest)
print(dir(unittest.TestCase))
print(unittest.TestCase.assertWarns)
```
