### 2018-12-20 d3a6e47
remove scanf requirement
