### 2018-12-20 d3af301
fix python2 int division bug
