### 2018-12-20 88f7956
fix py27 tests
