### 2018-12-20 8f391f6
fix matplotlib dep in tests
