### 2018-12-20 448916d
fix appveyor tests
