+++ setup.py
@@ -362,7 +362,7 @@ setup(name='OApackage',
       scripts=scripts,
       tests_require=['numpy', 'nose>=1.3', 'coverage>=4.0', 'mock' , 'python-dateutil'],
       zip_safe=False,
-      install_requires=['numpy>=1.13', 'mock; python_version <"3.0"'],
+      install_requires=['numpy>=1.13', 'mock; python_version <"3.0"', 'python-dateutil'],
       extras_require={
           'GUI':  ["qtpy", 'matplotlib'],
       },
