### 2018-12-20 1f74c2f
updates to tests
