### 2018-12-20 eendebakpt/oapackage#95 @gsagnol
Hi ,

You're correct the python installation of my PC's institute is mixed up. Since I am not administrating it, the simplest solution would indeed be to use virtualenvs, for now I will focus on the results I obtain on my laptop.
