### 2018-12-20 eendebakpt/oapackage#95 @gsagnol
Here is the ouput of your first script

Python 2.7.15rc1 (default, Nov 12 2018, 14:31:15)
Type "copyright", "credits" or "license" for more information.

IPython 5.5.0 -- An enhanced Interactive Python.
?         -> Introduction and overview of IPython's features.
%quickref -> Quick reference.
help      -> Python's own help system.
object?   -> Details about 'object', use 'object??' for extra details.
Using matplotlib backend: TkAgg

In [1]: import numpy as np; np.show_config()
   ...:
   ...: import oapackage
   ...: al = oapackage.exampleArray(14, 1)
   ...: al.showarray()
   ...: print('Defficiencies: %s' % (al.Defficiencies(), ) )
   ...:
   ...: scalefac=1
   ...: X=al.getModelMatrix(2)
   ...: print('model matrix:')
   ...: print(X)
   ...: M = np.linalg.inv(X.transpose().dot(X) / scalefac)
   ...: print('information matrix:')
   ...: print(M)
   ...: v = oapackage.oahelper.designStandardError(al)
   ...: print('output of designStandardError')
   ...: print(v)
   ...: print(v[0])
   ...:
lapack_opt_info:
    libraries = ['openblas', 'openblas']
    library_dirs = ['/usr/local/lib']
    define_macros = [('HAVE_CBLAS', None)]
    language = c
blas_opt_info:
    libraries = ['openblas', 'openblas']
    library_dirs = ['/usr/local/lib']
    define_macros = [('HAVE_CBLAS', None)]
    language = c
openblas_info:
    libraries = ['openblas', 'openblas']
    library_dirs = ['/usr/local/lib']
    define_macros = [('HAVE_CBLAS', None)]
    language = c
blis_info:
  NOT AVAILABLE
openblas_lapack_info:
    libraries = ['openblas', 'openblas']
    library_dirs = ['/usr/local/lib']
    define_macros = [('HAVE_CBLAS', None)]
    language = c
lapack_mkl_info:
  NOT AVAILABLE
blas_mkl_info:
  NOT AVAILABLE
exampleArray: design in D(28, 2^5), D-efficiency is low
array:
  1   0   1   1   1
  1   0   1   0   0
  0   1   0   1   1
  0   0   1   0   1
  1   1   1   0   0
  0   0   0   0   0
  0   1   0   1   1
  1   0   1   0   0
  1   0   0   0   0
  1   0   0   1   1
  0   1   1   0   0
  0   1   0   1   1
  1   0   1   1   0
  1   0   1   0   1
  0   0   1   0   0
  0   1   0   0   0
  0   0   0   1   1
  0   1   0   0   1
  0   1   0   1   1
  1   1   0   0   0
  1   0   1   0   1
  1   1   0   0   1
  1   1   0   1   0
  0   0   1   0   1
  0   1   0   1   0
  1   1   0   0   0
  0   1   1   0   0
  1   0   1   0   1
Defficiencies: (0.0, 0.0, 0.8727271801272938)
model matrix:
[[ 1.  1. -1.  1.  1.  1. -1.  1.  1.  1. -1. -1. -1.  1.  1.  1.]
 [ 1.  1. -1.  1. -1. -1. -1.  1. -1. -1. -1.  1.  1. -1. -1.  1.]
 [ 1. -1.  1. -1.  1.  1. -1.  1. -1. -1. -1.  1.  1. -1. -1.  1.]
 [ 1. -1. -1.  1. -1.  1.  1. -1.  1. -1. -1.  1. -1. -1.  1. -1.]
 [ 1.  1.  1.  1. -1. -1.  1.  1. -1. -1.  1. -1. -1. -1. -1.  1.]
 [ 1. -1. -1. -1. -1. -1.  1.  1.  1.  1.  1.  1.  1.  1.  1.  1.]
 [ 1. -1.  1. -1.  1.  1. -1.  1. -1. -1. -1.  1.  1. -1. -1.  1.]
 [ 1.  1. -1.  1. -1. -1. -1.  1. -1. -1. -1.  1.  1. -1. -1.  1.]
 [ 1.  1. -1. -1. -1. -1. -1. -1. -1. -1.  1.  1.  1.  1.  1.  1.]
 [ 1.  1. -1. -1.  1.  1. -1. -1.  1.  1.  1. -1. -1. -1. -1.  1.]
 [ 1. -1.  1.  1. -1. -1. -1. -1.  1.  1.  1. -1. -1. -1. -1.  1.]
 [ 1. -1.  1. -1.  1.  1. -1.  1. -1. -1. -1.  1.  1. -1. -1.  1.]
 [ 1.  1. -1.  1.  1. -1. -1.  1.  1. -1. -1. -1.  1.  1. -1. -1.]
 [ 1.  1. -1.  1. -1.  1. -1.  1. -1.  1. -1.  1. -1. -1.  1. -1.]
 [ 1. -1. -1.  1. -1. -1.  1. -1.  1.  1. -1.  1.  1. -1. -1.  1.]
 [ 1. -1.  1. -1. -1. -1. -1.  1.  1.  1. -1. -1. -1.  1.  1.  1.]
 [ 1. -1. -1. -1.  1.  1.  1.  1. -1. -1.  1. -1. -1. -1. -1.  1.]
 [ 1. -1.  1. -1. -1.  1. -1.  1.  1. -1. -1. -1.  1.  1. -1. -1.]
 [ 1. -1.  1. -1.  1.  1. -1.  1. -1. -1. -1.  1.  1. -1. -1.  1.]
 [ 1.  1.  1. -1. -1. -1.  1. -1. -1. -1. -1. -1. -1.  1.  1.  1.]
 [ 1.  1. -1.  1. -1.  1. -1.  1. -1.  1. -1.  1. -1. -1.  1. -1.]
 [ 1.  1.  1. -1. -1.  1.  1. -1. -1.  1. -1. -1.  1.  1. -1. -1.]
 [ 1.  1.  1. -1.  1. -1.  1. -1.  1. -1. -1.  1. -1. -1.  1. -1.]
 [ 1. -1. -1.  1. -1.  1.  1. -1.  1. -1. -1.  1. -1. -1.  1. -1.]
 [ 1. -1.  1. -1.  1. -1. -1.  1. -1.  1. -1.  1. -1. -1.  1. -1.]
 [ 1.  1.  1. -1. -1. -1.  1. -1. -1. -1. -1. -1. -1.  1.  1.  1.]
 [ 1. -1.  1.  1. -1. -1. -1. -1.  1.  1.  1. -1. -1. -1. -1.  1.]
 [ 1.  1. -1.  1. -1.  1. -1.  1. -1.  1. -1.  1. -1. -1.  1. -1.]]
information matrix:
[[ 5.23900804e-02  4.40546476e-02  1.67294674e-02 -1.09175849e-01
  -1.48444038e-01  2.23386805e-02 -5.74357211e-02 -1.57955342e-01
   9.89626920e-02  1.50432120e-01 -7.70154665e-02 -6.07261281e+12
  -1.02114117e+15  2.37004064e+15  6.07261281e+12  1.03511731e-02]
 [-3.00539040e-01  7.91325583e-02  2.17834688e-02  5.55470361e-02
  -1.12884746e-01  2.98224844e-02  8.38313238e-03 -9.94862751e-03
   1.67319989e-01  5.67516911e-02 -9.06625238e-02  1.10820717e+15
  -2.38445672e+15  6.29712244e+14 -1.10820717e+15 -1.12712774e-02]
 [-2.51790796e-01  8.32877541e-02  1.90154884e-03  1.73068207e-01
  -1.41259009e-01  4.20598409e-02 -3.56481703e-02  2.53987119e-02
   2.49728228e-01  2.15668183e-01 -6.34812316e-02 -1.25687025e+14
  -2.74203216e+15  1.75667167e+15  1.25687025e+14  6.92378470e-02]
 [-1.66538874e-01  6.67650667e-02 -6.57874559e-02  7.15381077e-02
  -2.78216531e-01  5.02415274e-02 -6.81625949e-02  3.86252109e-02
   1.85477687e-01  5.60496856e-02 -2.78216531e-01  3.65343383e+14
  -1.23553745e+15  2.10573152e+15 -3.65343383e+14  1.44916424e-01]
 [ 8.94189110e-02  1.14838595e-02 -4.20368210e-02 -1.09903939e-01
  -2.40559882e-01  6.16232411e-03 -4.42429975e-02  1.08906387e-01
   3.39096208e-02 -6.78269211e-02 -1.56376473e-01  1.23938186e+14
  -4.95973065e+14  2.32908627e+15 -1.23938186e+14  8.48049897e-02]
 [ 1.36405808e-01  1.69138385e-02  5.27963944e-02  1.50032610e-01
  -7.01375452e-03  6.31803936e-02 -1.47972431e-03 -2.86275014e-02
   1.63007471e-02  1.21908990e-01  5.36394405e-02 -1.82781649e+14
  -1.00885820e+15  3.07491837e+14  1.82781649e+14  7.05747331e-05]
 [-1.69750291e-01  1.11327819e-02  1.32860823e-01  2.15464549e-01
  -1.26764511e-01 -4.08925699e-03  7.58802084e-02 -1.00631707e-01
  -1.36079609e-02  7.11941059e-02 -1.92882410e-03 -2.27405083e+14
   2.04778756e+13  7.60300676e+14  2.27405083e+14 -1.31388903e-01]
 [ 5.61039303e-01 -2.57272095e-02  8.00618769e-02  4.41731045e-01
   1.68131520e-01 -1.02032252e-02  6.76634297e-02 -1.16318801e-01
  -1.22789383e-01  1.55109319e-01  3.94263661e-01 -4.07319378e+14
  -8.75712038e+13 -1.36798933e+15  4.07319378e+14 -1.03423797e-01]
 [ 4.67550345e-02  1.10140760e-01 -2.11227201e-02  1.85914078e-02
  -1.66347320e-01  3.01582477e-02 -3.52045335e-03  4.70315497e-02
   1.89724837e-01  1.08805975e-01  6.56229574e-02  1.31993138e+15
  -3.23938675e+15  5.37350460e+14 -1.31993138e+15  2.81636268e-02]
 [-1.58590521e-01  1.98238151e-02  6.78479236e-02 -6.81266226e-02
  -2.14991107e-01 -4.25791391e-03  1.13079873e-02 -1.46864485e-01
   1.35695847e-01  1.67529567e-02 -2.14991107e-01 -8.65125428e+14
   3.57114105e+14  1.62965271e+15  8.65125428e+14 -9.04638981e-02]
 [ 4.62970717e-01 -5.78713396e-02  0.00000000e+00  4.62970717e-01
   2.31485359e-01  2.89356698e-02  0.00000000e+00 -2.31485359e-01
   0.00000000e+00  3.47228038e-01  2.31485359e-01 -1.04251737e+15
  -1.04251737e+15  0.00000000e+00  1.04251737e+15  0.00000000e+00]
 [ 5.97232515e+15 -1.04284550e+15  1.28274892e+15  4.80154535e+15
   2.74402402e+15 -3.48963336e+14  6.50097657e+14 -2.26528152e+15
  -2.39321582e+15  1.55167349e+15  3.19282295e+15 -1.49349093e+31
   1.49349093e+31 -1.49349093e+31  1.49349093e+31 -2.25179981e+15]
 [-5.97232515e+15  1.04284550e+15 -1.28274892e+15 -4.80154535e+15
  -2.74402402e+15  3.48963336e+14 -6.50097657e+14  2.26528152e+15
   2.39321582e+15 -1.55167349e+15 -3.19282295e+15  1.49349093e+31
  -1.49349093e+31  1.49349093e+31 -1.49349093e+31  2.25179981e+15]
 [ 5.97232515e+15 -1.04284550e+15  1.28274892e+15  4.80154535e+15
   2.74402402e+15 -3.48963336e+14  6.50097657e+14 -2.26528152e+15
  -2.39321582e+15  1.55167349e+15  3.19282295e+15 -1.49349093e+31
   1.49349093e+31 -1.49349093e+31  1.49349093e+31 -2.25179981e+15]
 [-5.97232515e+15  1.04284550e+15 -1.28274892e+15 -4.80154535e+15
  -2.74402402e+15  3.48963336e+14 -6.50097657e+14  2.26528152e+15
   2.39321582e+15 -1.55167349e+15 -3.19282295e+15  1.49349093e+31
  -1.49349093e+31  1.49349093e+31 -1.49349093e+31  2.25179981e+15]
 [ 1.50000000e-01 -1.25000000e-02  5.00000000e-02  1.50000000e-01
   7.50000000e-02  1.25000000e-02  2.77343750e-02 -6.25000000e-02
  -5.00000000e-02  5.00000000e-02  7.50000000e-02 -4.50359963e+14
   4.50359963e+14 -4.50359963e+14  4.50359963e+14  0.00000000e+00]]
/usr/local/lib/python2.7/dist-packages/oapackage/oahelper.py:1167: RuntimeWarning: invalid value encountered in sqrt
  return np.sqrt(m0), np.sqrt(m1), np.sqrt(m2)
output of designStandardError
(0.2288887948318749, array([       nan, 0.04360675, 0.2513571 , 0.26746609, 0.2813051 ]), array([       nan,        nan,        nan,        nan,        nan,
       0.        , 0.12943321, 0.27546362, 0.43557415, 0.48112925]))
0.228888794832
