### 2018-12-20 eendebakpt/oapackage#95 @gsagnol
There seems to be a number of entries of magnitude > 1e15 in the inverse of the information matrix, is this normal?
