### 2018-12-20 #1097 @gsagnol
@peendebak : Concerning the last "check-box", it seems that some doi's are missing in the software paper. For example,
The book "Orthogonal Arrays: Theory and Applications" of Hedayat et al. should have doi `https://doi.org/10.1007/978-1-4612-1478-6`

Moreover, this reference appears in two separate bib entries (once as a book, once as a website).
