### 2018-12-20 75c0393
add doi for Hedayat1999
