+++ docs/references.bib
@@ -384,7 +384,8 @@ address = "United Kingdom",
   title = {Orthogonal arrays : theory and applications},
   publisher = {Springer},
   year = {1999},
-  author = {Hedayat, A.S. and Sloane, N.J.A. and Stufken, J.}
+  author = {Hedayat, A.S. and Sloane, N.J.A. and Stufken, J.},
+  doi = {10.1007/978-1-4612-1478-6}
 }
 
 @ARTICLE{Hedayat1997a,
