+++ setup.py
@@ -164,6 +164,7 @@ except:
 
 #%% Test suite
 
+""" Run a limited set of tests for the package """
 class OATest(TestCommand):
     user_options = [('pytest-args=', 'a', "Arguments to pass to py.test")]
 
@@ -184,6 +185,7 @@ class OATest(TestCommand):
         print('## oapackage test: load package')
         # import here, cause outside the eggs aren't loaded
         import oapackage.tests
+        import oapackage.oahelper
         import oapackage.tests.test_oapackage
         print('## oapackage test: oalib version %s' % oapackage.version())
         print('## oapackage test: package compile options\n%s\n' % oapackage.oalib.compile_information())
