### 2018-12-20 87c19d6
make tests more quiet
