+++ oapackage/tests/test_oapackage.py
@@ -59,11 +59,11 @@ class TestMisc(unittest.TestCase):
 
     def test_exampleArray(self):
         # test a selection of the example arrays
-        al = oapackage.exampleArray(5)
+        al = oapackage.exampleArray(5, 0)
         assert(al.md5() == '3885c883d3bee0c7546511255bb5c3ae')
-        al = oapackage.exampleArray(20)
+        al = oapackage.exampleArray(20, 0)
         assert(np.array(al).shape == (24, 3))
-        al = oapackage.exampleArray(40)
+        al = oapackage.exampleArray(40, 0)
         assert(np.array(al).shape == (14, 5))
 
     def test_scanf(self):
@@ -116,7 +116,7 @@ def test_numpy_interface(verbose=0):
 def test_nauty(verbose=0):
     if verbose:
         print('test_nauty: test reduction to normal form')
-    al = oapackage.exampleArray(1, verbose)
+    al = oapackage.exampleArray(0, verbose)
     alr = al.randomperm()
     tr = oapackage.reduceOAnauty(alr)
     alx = tr.apply(alr)
@@ -161,7 +161,7 @@ def miscunittest(verbose=1):
     # DOP reduction
     if verbose >= 2:
         print('unittest: test delete-one-factor GWLP reduction')
-    al = oalib.exampleArray(5, 1)
+    al = oalib.exampleArray(5, verbose)
     al2 = al.randomperm()
 
     alr = al.reduceDOP()
@@ -199,10 +199,10 @@ class TestOAfiles(unittest.TestCase):
 
     def test_misc_file_operations(self):
         array_filename = tempfile.mktemp(suffix='.oa')
-        lst = [oapackage.exampleArray(4, 1)]
+        lst = [oapackage.exampleArray(4, 0)]
         oapackage.writearrayfile(array_filename, lst)
         assert(oapackage.oahelper.oaIsBinary(array_filename) is False)
-        oapackage.writearrayfile(array_filename, oapackage.exampleArray(4, 1), oapackage.ABINARY)
+        oapackage.writearrayfile(array_filename, oapackage.exampleArray(4, 0), oapackage.ABINARY)
         assert(oapackage.oahelper.oaIsBinary(array_filename))
 
         oapackage.oahelper.oainfo(array_filename)
@@ -212,7 +212,7 @@ class TestOAfiles(unittest.TestCase):
     def test_selectArrays(self):
         array_filename = tempfile.mktemp(suffix='.oa', dir=tempfile.tempdir)
         array_filename_out = tempfile.mktemp(suffix='.oa', dir=tempfile.tempdir)
-        oapackage.writearrayfile(array_filename, [oapackage.exampleArray(4, 1), oapackage.exampleArray(4, 1)])
+        oapackage.writearrayfile(array_filename, [oapackage.exampleArray(4, 0), oapackage.exampleArray(4, 1)])
         oapackage.oahelper.selectArrays(array_filename, array_filename_out, [
                                         1], afmode=oalib.ABINARY, verbose=1, cache=0)
         oapackage.oahelper.selectArrays(array_filename, array_filename_out,
@@ -220,13 +220,13 @@ class TestOAfiles(unittest.TestCase):
 
     def test_nArrayFile(self):
         array_filename = tempfile.mktemp(suffix='.oa', dir=tempfile.tempdir)
-        oapackage.writearrayfile(array_filename, [oapackage.exampleArray(4, 1)])
+        oapackage.writearrayfile(array_filename, [oapackage.exampleArray(4, 0)])
         self.assertEqual(oapackage.oahelper.nArrayFile(array_filename), 1)
[…]
