### 2018-12-20 5c669ea
add explicit install in appveyor testing section
