### 2018-12-20 98f4c83
make tests more quiet
