+++ appveyor.yml
@@ -108,7 +108,7 @@ install:
   - ps: $env:ispython36 = $env:PYTHON.Contains("Python36")
 
   # part of the testing requirements
-  - "%CMD_IN_ENV% pip install scanf"
+  - "%CMD_IN_ENV% pip install python-dateutil"
   - "%CMD_IN_ENV% pip install pytest"
 
   # for running the jupyter notebooks
