### 2018-12-20 b38366f
make tests more quiet
