### 2018-12-20 eendebakpt/oapackage#97 @gsagnol
There are not enough explanations about which optimal design problem is solved
I think the software paper should state more explicitly which problem your package can solve.
In particular, I don't like the term "optimal design", which is very general, while you are solving very particular
optimal design problems. Further, I don't think your software can guarantee global optimality.

In the documentation, I think it should be explained in details which problem you are solving. If I understand it
correctly, The `Doptimize()` function computes optimal designs with respect to
1) A quadratic response surface model
2) A compound criterion which mixes D, Ds and D1 optimality. Here, Ds seems to be D-optimality restricted to the subparameter systems of the linear terms in the quadratic response surface (If I guess correctly what "main effect robustness" means). I have no idea what D1 is, but I guess this is in fact a c-optimality criterion for the estimation of the intercept? These are probably well known things in your community, but I think it would be better to give more details, so your software can reach a broader audience.

Also, in the section "Data representation", you say that a D-optimal design maximizes $det(X_M X_M^T)^{1/p}$. I think it should be explained what the model matrix $X_M$ is, here. I had a look at what the function `get_modelMatrix()` is returning, and it seems that something special is done when
the number of factor levels is $s_i>2$ for some factor. This should probably be explained.

It should also be clearly written how your compound criterion is formed. If we denote by $\Phi_X$ each of your $D_X$ - criterions in det-root-p form (not in log-det form), are you maximizing $\Phi_D^{\alpha_1}  \Phi_{Ds}^{\alpha_2} \Phi_{D1}^{\alpha_3}$, or $\alpha_1 \Phi_D + \alpha_2 \Phi_{Ds} + \alpha_3 \Phi_{D1}$ ? I'am asking this question because I believe that both approaches coexist in the literature.
