### 2018-12-20 eendebakpt/oapackage#98 @gsagnol
Statement of need
The introduction text in the documentation is misleading. It says " large collection of arrays can be found on the Orthogonal Array package website [Een12]. An alternative collection of orthogonal arrays can be found on the website of Neil Sloane [Slo14]". Is it really related with your OApackage ?

* If yes, you should explain how these designs can be computed using your package. Maybe these are the designs stored as `oapackage.exampleArray()` ?

* Otherwise, this should probably moved somewhere else as in the introduction, which should solely be devoted to explain the features of your package.
