### 2018-12-20 eendebakpt/oapackage#99 @gsagnol
Funny behaviour of `strength`
* In some of the examples, you use ``strength = 0``. Does this mean that there are no strenghth restriction at all ?

* I think that an orthogonal array of strength $t$ with factors at $s$ different levels can only exist if the number of runs $N$ is a scalar multiple of $s^t$. `Doptim()` does not complain when we ask to optimize over
the class `oapackage.arraydata_t(2, 17, 2, 3)`, for example. In the same vein, does it mean something for a design with different number of levels for each factor to have strenght $t$ ? For example, OApackage does not issue any warning when we create an `arrayclass=oapackage.arraydata_t([2,3,4], 20, 2, 3)`.
