### 2018-12-20 eendebakpt/oapackage#100 @gsagnol
Doptim() is not documented enough
In the example "Calculation of D-optimal designs", you are using a parameter `method=oapackage.DOPTIM_UPDATE`.
This parameter does not seem to be used in that example, and it is not documented in `Doptim()` (and without the example, the user would not even know this parameter exists). It seems that `Doptim()` is one of the most important function of your package, so all its features should probably be documented.
