### 2018-12-20 eendebakpt/oapackage#101 @gsagnol
Full c++ signatures not necessary for python users
In the section *Generation of designs* of the documentation, I do not find
the way you are referring to functions using their complete c++ signature very elegant. I would find it more readable to simply write their name, with a link to the API documentation for a full explanation of the parameters.
