### 2018-12-20 eendebakpt/oapackage#102 @gsagnol
signature of conference_t ? Use of sphinx-doctest ?
`ctype=oapackage.conference_t(N=8, k=8)` raises an error
`TypeError: __init__() got an unexpected keyword argument 'k'`
but the statement `ctype=oapackage.conference_t(8,8,0)` seems to be working.

I tried to reproduce your example in the section Generation of designs -> Conference designs, but it is not working. I would strongly recommend that you use sphinx-doctest, in addition to your unit tests, to automatically check that the examples of the documentation work as expected. For example, in the documentation, your code includes a line
`>>> al.showarray() array: 0 1 1 1 0 -1 1 1 0 1 1 1 1 1 -1 1 -1 1 1 -1 1 1 -1 -1`
This is clearly wrong, and has been manually entered rather than generated by sphinx, because this is an output string (which should not start with ">>>"m and,  new lines "\" are missing.
