### 2018-12-20 eendebakpt/oapackage#103 @gsagnol
More details should be given on conference designs and Design analysis
I'm not very familiar with conference designs, and I find the documentation very scarce. I am not able to understand what `extend_conference` is doing, and I the example (in the section Generation of Designs -> Conference Designs) should provide more information about `create_root_three`, too. Moreover, the example uses `k=8` but it only computes designs with up to 6 columns.

The documentation  should also give some formulas, so that the user can understand what `PECSequence`, `PICSequence()`, `GWLP()` and `Jcharacteristics()` are doing. Usage examples would be helpful, too. For example,  what is the input `ii` of `Jcharacteristics_conference` ?
