### 2018-12-20 #1097 @gsagnol
@peendebak : I created a number of issues in the source repository. This can be considered as my list of requests for a *minor revision*.
