### 2018-12-20 4a3a273
add requirements for scanf module
