+++ setup.py
@@ -362,9 +362,9 @@ setup(name='OApackage',
       packages=packages,
       data_files=data_files,
       scripts=scripts,
-      tests_require=['numpy', 'nose>=1.3', 'coverage>=4.0', 'mock' , 'python-dateutil'],
+      tests_require=['numpy', 'nose>=1.3', 'coverage>=4.0', 'mock' , 'python-dateutil','backports.functools_lru_cache;python_version<"2.9"'],
       zip_safe=False,
-      install_requires=['numpy>=1.13', 'mock; python_version <"3.0"', 'python-dateutil'],
+      install_requires=['numpy>=1.13', 'mock; python_version <"3.0"', 'python-dateutil','backports.functools_lru_cache;python_version<"2.9"'],
       extras_require={
           'GUI':  ["qtpy", 'matplotlib'],
       },
