### 2018-12-20 1a6744b
fix appveyor tests
