### 2018-12-20 d18e2c9
fix doctests
