### 2018-12-20 59a9bb7
update docstrings; remove 2.7 notebooks in appveyor
