+++ appveyor.yml
@@ -165,9 +165,7 @@ test_script:
    - "%CMD_IN_ENV36% jupyter nbconvert --to notebook --execute  docs/examples/example_designs.ipynb "
    - "%CMD_IN_ENV36% jupyter nbconvert --to notebook --execute  docs/examples/example_conference_designs.ipynb"
 
-   # selected execution for python2.7
-   - "%CMD_IN_ENV276% jupyter nbconvert --to notebook --execute  docs/examples/example_oapackage.ipynb "
-     
+    
 after_test:
   - "%CMD_IN_ENV% %PYTHON%/python setup.py bdist_wheel bdist_wininst"
   - "%CMD_IN_ENV36% %PYTHON%/python setup.py sdist"
