### 2018-12-20 bf5a25e
make setup compatible with python 3.4; update doctests captions
