+++ docs/generation.rst
@@ -66,19 +66,23 @@ with entries 0, -1, +1 such that i) in each column the symbol 0 occurs
 exactly one time and ii) all columns are orthogonal to each other.
 For details see :cite:`Schoen2018dsd`, :cite:`wiki:ConferenceMatrix`.
 
-.. code-block:: python
+.. .. code-block:: python
  :caption: Generate conference designs with 8 rows
-                    
- >>> import oapackage
- >>> ctype=oapackage.conference_t(N=8, k=8)
- >>> al = ctype.create_root_three()
- >>> al.showarray() array: 0 1 1 1 0 -1 1 1 0 1 1 1 1 1 -1 1 -1 1 1 -1 1 1 -1 -1
- >>> l4=oapackage.extend_conference ([al], ctype, verbose=0)
- >>> l5=oapackage.extend_conference ( l4, ctype,verbose=0) 
- >>> l6=oapackage.extend_conference ( l5, ctype, verbose=0)
- >>>
- >>> print('number of non-isomorphic conference designs: number of conference designs: %d'  % len(l6) )
- non-isomorphic conference designs: 11
+
+.. admonition:: Generate conference designs with 8 rows
+
+   .. doctest::     
+                   
+     >>> import oapackage
+     >>> ctype=oapackage.conference_t(8, 8, 0) 
+     >>> al = ctype.create_root_three()
+     >>> al.showarray()
+     array: 0 1 1 1 0 -1 1 1 0 1 1 1 1 1 -1 1 -1 1 1 -1 1 1 -1 -1
+     >>> l4=oapackage.extend_conference ([al], ctype, verbose=0)
+     >>> l5=oapackage.extend_conference ( l4, ctype,verbose=0) 
+     >>> l6=oapackage.extend_conference ( l5, ctype, verbose=0)
+     >>> print('number of non-isomorphic conference designs: number of conference designs: %d'  % len(l6) )
+     non-isomorphic conference designs: 11
 
 
 The full interface for conference designs is available
@@ -96,21 +100,27 @@ with good properties for the :math:`D`-efficiency.
 
 A Python script to generate optimal designs with 40 runs and 7 factors is shown below.
 
-.. code-block:: python
- :caption: Doptimize
- 
- >>> N=40; s=2; k=7;
- >>> arrayclass=oapackage.arraydata_t(s, N, 0, k) 
- >>> print('We generate optimal designs with: %s' % arrayclass)
- We generate optimal designs with: arrayclass: N 40, k 7, strength 0, s 2,2,2,2,2,2,2, order 0.
- >>> alpha=[1,2,0] 
- >>> method=oapackage.DOPTIM_UPDATE 
- >>> scores, dds, designs, ngenerated = oapackage.Doptimize(arrayclass, nrestarts=40, optimfunc=alpha, selectpareto=True)
- Doptim: optimization class 40.2-2-2-2-2-2-2
- Doptimize: iteration 0/40
- Doptimize: iteration 39/40 Doptim: done (8 arrays, 0.6 [s]) 
- >>> print('Generated %d designs, the best D-efficiency is %.4f’ % (len(designs), dds[:,0].max() ))
- Generated 8 designs, the best D-efficiency is 0.9098
+.. admonition:: Example of Doptimize usage
+
+   .. testsetup::
+   
+       >>> import oapackage
+       
+   .. doctest::     
+
+
+     >>> N=40; s=2; k=7;
+     >>> arrayclass=oapackage.arraydata_t(s, N, 0, k) 
+     >>> print('We generate optimal designs with: %s' % arrayclass)
[…]
