### 2018-12-20 bdbb035
updatedocs to doctests
