### 2018-12-20 a2ad39f
doctest updates
