+++ docs/data.rst
@@ -74,18 +74,34 @@ It is also possible to convert to a Numpy array. The
 operations from packages such as Numpy work on the :meth:`~oalib.array_link`
 object.
 
-.. code-block:: python
- :caption: Array representation and indexing in Python
+.. admonition:: Array representation and indexing in Python
 
- >>> import oapackage
- >>> al=oapackage.exampleArray(0)
- >>> al.showarray() 
- array: 0 0 0 0 0 1 0 1 1 0 1 0 1 1 1 1 
- >>> al[2,1] 
- 1L 
- >>> X=np.array(al)
- >>> X 
- array([[0, 0], [0, 0], [0, 1], [0, 1], [1, 0], [1, 0], [1, 1], [1, 1]], dtype=int32)
+  .. doctest:: 
+  
+    >>> import oapackage
+    >>> al=oapackage.exampleArray(0)
+    >>> al.showarray() 
+    array: 
+      0   0
+      0   0
+      0   1
+      0   1
+      1   0
+      1   0
+      1   1
+      1   1
+    >>> al[2,1] 
+    1L 
+    >>> X=np.array(al)
+    >>> X 
+    array([[0, 0],
+           [0, 0],
+           [0, 1],
+           [0, 1],
+           [1, 0],
+           [1, 0],
+           [1, 1],
+           [1, 1]], dtype=int16)
 
 The C++ class is:
 
@@ -102,6 +118,10 @@ class.
 
 .. code-block:: python
    :caption: Write an array to disk
+
+.. admonition:: Write an array to disk
+
+  .. doctest:: 
 
    >>> import oapackage
    >>> al=oapackage.exampleArray()
