### 2018-12-20 865dadb
fix whitespace issues with doctest
