+++ docs/data.rst
@@ -74,14 +74,20 @@ It is also possible to convert to a Numpy array. The
 operations from packages such as Numpy work on the :meth:`~oalib.array_link`
 object.
 
+.. testsetup::
+   
+   import sys
+   # doctest: +NORMALIZE_WHITESPACE
+
+
 .. admonition:: Array representation and indexing in Python
 
   .. doctest:: 
   
-    >>> import oapackage
+    >>> import oapackage; import numpy as np
     >>> al=oapackage.exampleArray(0)
-    >>> al.showarray() 
-    array: 
+    >>> al.showarray() # doctest: +NORMALIZE_WHITESPACE
+    array:
       0   0
       0   0
       0   1
@@ -90,8 +96,9 @@ object.
       1   0
       1   1
       1   1
+      
     >>> al[2,1] 
-    1L 
+    1
     >>> X=np.array(al)
     >>> X 
     array([[0, 0],
@@ -115,9 +122,6 @@ Reading and writing arrays
 
 Reading and writing arrays to disk can be done with the :meth:`oalib.arrayfile_t`
 class. 
-
-.. code-block:: python
-   :caption: Write an array to disk
 
 .. admonition:: Write an array to disk
 
