### 2018-12-20 5d1920c
use full showarray
