### 2018-12-20 e1ca68e
add warning when Doptimize is used with strength unequal to zero
