### 2018-12-20 0065c10
with numerical rounding error in tests
