+++ oapackage/tests/test_cpplibrary.py
@@ -405,7 +405,7 @@ class TestCppLibrary(unittest.TestCase):
         pec_seq = oapackage.PECsequence(al)
         numpy.testing.assert_equal(pec_seq, (1.0,) * len(pec_seq))
         pic_seq = oapackage.PICsequence(al)
-        numpy.testing.assert_equal(pic_seq, (0.9985780064264659, 0.9965697009006985, 0.9906411254224957,
+        numpy.testing.assert_almost_equal(pic_seq, (0.9985780064264659, 0.9965697009006985, 0.9906411254224957,
                                              0.9797170906488152, 0.9635206782887167, 0.9421350381959234, 0.9162739059686846, 0.8879176205539139))
 
     def test_arraylink(self):
