### 2018-12-21 e5365a4
change order of tests
