### 2018-12-21 87aebf2
Address joss review comments (#96)
* include scanf code to prevent import conflicts with other scanf packages

* add missing file

* fix unittests

* for python 2.7 add explicit warning if mock is not installed

* fix environment marker format

* fix exception definition for python 2.7

* fix exception definition for python 2.7

* add mock to appveyor

* fix appveyor tests

* autopep

* fix appveyor file

* bump version

* remove scanf requirement

* fix python2 int division bug

* fix py27 tests

* fix matplotlib dep in tests

* fix appveyor tests

* updates to tests

* updates to tests

* tr
[…]
