+++ oapackage/oahelper.py
@@ -23,6 +23,7 @@ import warnings
 import webbrowser
 import tempfile
 import subprocess
+import dateutil.parser
 
 try:
     import matplotlib
@@ -239,8 +240,9 @@ def niceplot(ax, fig=None, despine=True, verbose=0, figurebg=True,
 
 def enlargelims(factor=1.05):
     """ Enlarge the limits of a plot
-    Example:
-      >>> enlargelims(1.1)
+    
+    Args:
+        factor (float): factor by which to make the plot margins wider
     """
     xl = plt.xlim()
     d = (factor - 1) * (xl[1] - xl[0]) / 2
@@ -834,10 +836,7 @@ def mkdirc(directory_name):
 def parseProcessingTime(logfile, verbose=0):
     """ Parse a log file to calculate the processing time """
 
-    try:
-        import dateutil.parser
-    except BaseException:
-        warnings.warn('oahelper: could not load dateutil package...')
+    import dateutil.parser
 
     fileinput.close()
     tstart = None
@@ -997,7 +996,9 @@ def runExtend(N, k, t=3, l=2, verbose=1, initsols=None, nums=[], algorithm=None)
     Returns:
         list: list of generated designs
 
-    >>> r = runExtend(16, 5, 3, verbose=0)
+    Example:
+       >>> import oapackage
+       >>> designs = oapackage.oahelper.runExtend(16, 5, 3, verbose=0)
     """
     if verbose:
         print('runExtend: N=%d, k=%d, t=%d' % (N, k, t))
