### 2018-12-21 e72305f
Update documentation links to website (#93)
* update documentation references to website

* add doi for Hedayat1999
