+++ docs/references.bib
@@ -58,12 +58,13 @@ url={http://www.pietereendebak.nl/oapage/},
 note={\url{http://www.pietereendebak.nl/oapage/}}
 }
 
-@TECHREPORT{EendebakOApackage,
+@TECHREPORT{EendebakOApackageResults,
 author={P. T. Eendebak},
-year={2013},
-title={The {O}rthogonal {A}rray package},
-url={http://www.pietereendebak.nl/oapage/software.html}
+year={2018},
+title={The {O}rthogonal {A}rray package: results},
+url={http://www.pietereendebak.nl/oapackage/}
 }
+
 
 @misc{EendebakDOF,
 author={P. T. Eendebak},
@@ -383,7 +384,8 @@ address = "United Kingdom",
   title = {Orthogonal arrays : theory and applications},
   publisher = {Springer},
   year = {1999},
-  author = {Hedayat, A.S. and Sloane, N.J.A. and Stufken, J.}
+  author = {Hedayat, A.S. and Sloane, N.J.A. and Stufken, J.},
+  doi = {10.1007/978-1-4612-1478-6}
 }
 
 @ARTICLE{Hedayat1997a,
