### 2018-12-21 e0f7236
update documentation of Doptimize methods
