### 2018-12-21 b97f41b
cleanup Deff.h and Deff.cpp
