+++ src/Deff.cpp
@@ -3,8 +3,10 @@
  *
  */
 
-#include <printfheader.h>
+#include <algorithm>
+#include <cmath>
 
+#include "printfheader.h"
 #include "arrayproperties.h"
 #include "arraytools.h"
 
@@ -27,6 +29,8 @@
 #define MATLABUPDATE
 #endif
 #endif
+
+const double NaN = std::numeric_limits< double >::quiet_NaN();
 
 DoptimReturn Doptimize (const arraydata_t &arrayclass, int nrestartsmax, std::vector< double > alpha, int verbose,
                         int method, int niter, double maxtime, int nabort) {
@@ -145,7 +149,6 @@ DoptimReturn DoptimizeMixed (const arraylist_t &sols, const arraydata_t &arraycl
                 array_link alu2 = optimDeff (alu, arrayclass, alpha, verbose >= 3, method2, niter, 0);
                 double score2 = scoreD (alu2.Defficiencies (), alpha);
 
-                //#pragma omp critical
                 dds[i] = alu2.Defficiencies ();
                 AA[i] = alu2;
 
@@ -182,7 +185,6 @@ array_link optimDeff (const array_link &A0, const arraydata_t &arrayclass, const
                 }
         }
 
-        int nx = 0;
         std::vector< int > factor_levels = arrayclass.factor_levels ();
         if (optimmethod == DOPTIM_UPDATE) {
                 if (arrayclass.is2level ())
@@ -202,7 +204,7 @@ array_link optimDeff (const array_link &A0, const arraydata_t &arrayclass, const
         // initialize score
         double current_score = scoreD (dd0, alpha);
 
-        int lc = 0; // index of last change to array
+        int last_update = 0; // index of last change to array
 
         // initialize arary with random permutation
         const int nn = N * k;
@@ -251,7 +253,6 @@ array_link optimDeff (const array_link &A0, const arraydata_t &arrayclass, const
                 }
                 // evaluate
                 std::vector< double > dd = A.Defficiencies ();
-                nx++;
                 double next_score = scoreD (dd, alpha);
 
                 if (verbose >= 4) {
@@ -263,9 +264,7 @@ array_link optimDeff (const array_link &A0, const arraydata_t &arrayclass, const
 
                 if (next_score >= current_score) {
                         if (next_score > current_score) {
-                                lc = ii;
-                                // std::random_shuffle ( updatepos.begin(), updatepos.end() );	// update the
-                                // sequence of positions to try
+                                last_update = ii;
 
                                 if (verbose >= 3)
                                         myprintf ("optimDeff: ii %d: %.6f -> %.6f\n", ii, current_score, next_score);
@@ -298,9 +297,9 @@ array_link optimDeff (const array_link &A0, const arraydata_t &arrayclass, const
                 }
 
                 // check progress
-                if ((ii - lc) > nabort) {
+                if ((ii - last_update) > nabort) {
                         if (verbose >= 2)
-                                myprintf ("optimDeff: early abort ii %d, lc %d: %d\n", ii, lc, (ii - lc));
+                                myprintf ("optimDeff: early abort ii %d, lc %d: %d\n", ii, last_update, (ii - last_update));
                         break;
                 }
         }
@@ -317,7 +316,6 @@ array_link optimDeff (const array_link &A0, const arraydata_t &arrayclass, const
 
 array_link optimDeff2level (const array_link &A0, const arraydata_t &arrayclass, std::vector< double > alpha,
                             int verbose, int optimmethod, int niter, int nabort) {
-        int nx = 0;
         int N = arrayclass.N;
         int k = arrayclass.ncols;
[…]
