### 2018-12-21 33376ea
add extra checks on arraydata_t constructor
