### 2018-12-21 757ac7d
add tests for arraydata_t oaindex; update docstrings
