### 2018-12-21 62fe51e
add warning for strength check; add tests
