### 2018-12-21 eendebakpt/oapackage#99 @eendebakpt
@gsagnol `strength = 0` does indeed mean there are no restrictions on the strength. The arrays in this class would normally not be called orthogonal arrays, since the orthogonal arrays have (by the regular definition) at least strength 2.

I added a warning to the D-optimize function to warn if a non-zero strength is used.

For mixed-level arrays the strength is properly defined, the definition is a straightforward generalization of the pure-level case. The package indeed does now warn when we create the class OA([4,3,2], 20, 2, 3). I did update the code so that for this case the index of the class (computed in the `.oaindex` variable) is zero, meaning that the strength and the level specifications do not match. I also issue a warning now if the factor levels are not ordered from large to small.
