### 2018-12-21 22cdd4d
fix print statement in complete_arraydata
