### 2018-12-21 e03d132
complete unittest of strength
