### 2018-12-21 6e25f34
fix compile error with older visual studio
