### 2018-12-21 c5c711d
fix compile error with older visual studio
