### 2018-12-21 e584a7f
cleanup code
