### 2018-12-21 9c75607
fix for old unittest on py27
