### 2018-12-21 19c8b17
fix py27 tests
