### 2018-12-21 159b40f
Add checks on strength in arraydata_t class (#106)
* add extra checks on arraydata_t constructor

* add tests for arraydata_t oaindex; update docstrings

* add warning for strength check; add tests

* fix print statement in complete_arraydata

* complete unittest of strength

* fix compile error with older visual studio

* fix compile error with older visual studio

* cleanup code

* fix for old unittest on py27

* fix py27 tests
