+++ docs/references.bib
@@ -58,7 +58,7 @@ url={http://www.pietereendebak.nl/oapage/},
 note={\url{http://www.pietereendebak.nl/oapage/}}
 }
 
-@TECHREPORT{EendebakOApackageResults,
+@misc{EendebakOApackageResults,
 author={P. T. Eendebak},
 year={2018},
 title={The {O}rthogonal {A}rray package: results},

[…]
