### 2018-12-21 53f9f41
remove unused variable
