### 2018-12-22 796f391
update docs of Doptimize
