+++ oapackage/Doptim.py
@@ -480,12 +480,13 @@ def Doptimize(arrayclass, nrestarts=10, optimfunc=[
               1, 0, 0], verbose=1, maxtime=180, selectpareto=True, nout=None, method=oalib.DOPTIM_UPDATE, niter=100000, nabort=0, dverbose=1):
     """ Calculate D-optimal designs
 
-   The optimization is performed multiple times (specified by the nrestarts parameter) to prevent finding a 
-   design in a local minmum of the target function. The method uses a coordinate-exchange algorithm to optimze
-   a target function defined by the optimziation paramaters.
+   The method uses a coordinate-exchange algorithm find a D-optimal design in the class specified by the
+   arrayclass. The optimality is defined in terms of the optimization parameters. The optimization is performed
+   multiple times (specified by the nrestarts parameter) to prevent finding a design in a local minmum of the
+   target function. 
 
-
-    For more details see the paper "Two-Level Designs to Estimate All Main
+    The optimization target and the Pareto optimality are defined in terms of the D-efficiency, main effect robustness
+    (or Ds-optimality) and the D1-efficiency of the design. For more details see the paper "Two-Level Designs to Estimate All Main
     Effects and Two-Factor Interactions", http://dx.doi.org/10.1080/00401706.2016.1142903
 
     Parameters
@@ -495,11 +496,11 @@ def Doptimize(arrayclass, nrestarts=10, optimfunc=[
     nrestarts : integer
         Number of restarts of the algorithm
     optimfunc : list with 3 floats
-        Gives the optimization weights
+        Gives the optimization weights :math:`\\alpha` of the target function :math:`\\alpha[0]*D+\\alpha[1]*D_s+\\alpha[2]*D_1`
     verbose : integer
-        A higher numer gives more output
+        Verbosity level. A higher numer gives more output
     maxtime: float
-        Maximum running time of the algorithm
+        Maximum running time of the algorithm. If this time is exceeded the algorithm is aborted.
     selectpareto : boolean, default is True
         If True then only the Pareto optimal designs are returned
     nout : integer, default None
