### 2018-12-22 95b16e5
update docs of Doptimize
