### 2018-12-22 18e1e7b
format docstring
