### 2018-12-22 bed5e47
Improve documentation of Doptimize (#107)
* update documentation of Doptimize methods

* cleanup Deff.h and Deff.cpp

* remove unused variable

* update docs of Doptimize

* update docs of Doptimize

* format docstring
