+++ paper/paper.bib
@@ -39,6 +39,7 @@ note={Published online}
   eprintx = { http://dx.doi.org/10.1080/00401706.2016.1142903},
   url = {http://dx.doi.org/10.1080/00401706.2016.1142903},
 }
+
 @misc{EendebakDOF,
   author = {Eendebak, P.T.},
   year = {2014},
