+++ src/Deff.cpp
@@ -3,8 +3,10 @@
  *
  */
 
-#include <printfheader.h>
+#include <algorithm>
+#include <cmath>
 
+#include "printfheader.h"
 #include "arrayproperties.h"
 #include "arraytools.h"
 
@@ -27,6 +29,8 @@
 #define MATLABUPDATE
 #endif
 #endif
+
+const double NaN = std::numeric_limits< double >::quiet_NaN();
 
 DoptimReturn Doptimize (const arraydata_t &arrayclass, int nrestartsmax, std::vector< double > alpha, int verbose,
                         int method, int niter, double maxtime, int nabort) {
@@ -145,7 +149,6 @@ DoptimReturn DoptimizeMixed (const arraylist_t &sols, const arraydata_t &arraycl
                 array_link alu2 = optimDeff (alu, arrayclass, alpha, verbose >= 3, method2, niter, 0);
                 double score2 = scoreD (alu2.Defficiencies (), alpha);
 
-                //#pragma omp critical
                 dds[i] = alu2.Defficiencies ();
                 AA[i] = alu2;
 
@@ -174,6 +177,7 @@ array_link optimDeff (const array_link &A0, const arraydata_t &arrayclass, const
         const int k = arrayclass.ncols;
 
         if (nabort <= 0) {
+			/// select a sane default
                 if (arrayclass.is2level ()) {
                         nabort = (N * k) * 2.5 + 1; // factor 2 to compensate for equal error switches
                 } else {
@@ -181,14 +185,13 @@ array_link optimDeff (const array_link &A0, const arraydata_t &arrayclass, const
                 }
         }
 
-        int nx = 0;
-        std::vector< int > s = arrayclass.factor_levels ();
+        std::vector< int > factor_levels = arrayclass.factor_levels ();
         if (optimmethod == DOPTIM_UPDATE) {
                 if (arrayclass.is2level ())
                         optimmethod = DOPTIM_FLIP;
         }
         array_link A = A0;
-        symmetry_group sg = symmetry_group (s);
+        symmetry_group sg = symmetry_group (factor_levels);
 
         std::vector< int > gidx = sg.gidx;
 
@@ -199,9 +202,9 @@ array_link optimDeff (const array_link &A0, const arraydata_t &arrayclass, const
         }
 
         // initialize score
-        double d = scoreD (dd0, alpha);
+        double current_score = scoreD (dd0, alpha);
 
-        int lc = 0; // index of last change to array
+        int last_update = 0; // index of last change to array
 
         // initialize arary with random permutation
         const int nn = N * k;
@@ -222,7 +225,7 @@ array_link optimDeff (const array_link &A0, const arraydata_t &arrayclass, const
 
                 // get values
                 array_t o = A._at (r, c);
-                array_t o2 = A._at (r2, c2); // no extra error checking
+                array_t o2 = A._at (r2, c2);
 
                 // update
 
@@ -235,7 +238,7 @@ array_link optimDeff (const array_link &A0, const arraydata_t &arrayclass, const
                         A._setvalue (r2, c2, o);
                         break;
                 case DOPTIM_UPDATE: { // random update
-                        int val = fastrandK (s[c]);
+                        int val = fastrandK (factor_levels[c]);
                         A._setvalue (r, c, val);
                         break;
                 }
@@ -250,35 +253,32 @@ array_link optimDeff (const array_link &A0, const arraydata_t &arrayclass, const
                 }
                 // evaluate
                 std::vector< double > dd = A.Defficiencies ();
-                nx++;
-                double dn = scoreD (dd, alpha);
+                double next_score = scoreD (dd, alpha);
 
                 if (verbose >= 4) {
[…]
