+++ src/Deff.h
@@ -20,8 +20,18 @@
  */
 double scoreD (const std::vector< double > efficiencies, const std::vector< double > alpha);
 
-/// different algorithms for the optimization routines
-enum { DOPTIM_UPDATE, DOPTIM_SWAP, DOPTIM_FLIP, DOPTIM_AUTOMATIC, DOPTIM_NONE };
+/// Different methods for the optimization. The default method DOPTIM_SWAP is a coordinate-exchange algorithms
+enum {
+	/// replace a random element with a random value
+	DOPTIM_UPDATE,
+	/// swap two elements at random
+	DOPTIM_SWAP,
+	/// randomly flip an element between 0 and 1
+	DOPTIM_FLIP,
+	/// automatically select one of the methods
+	DOPTIM_AUTOMATIC, 
+	/// perform no optimization
+	DOPTIM_NONE };
 
 /** Optimize a design according to the optimization function specified.
  *
@@ -52,20 +62,30 @@ struct DoptimReturn {
 
 /** Generates optimal designs for the specified class of designs
  *
+ * The method uses a coordinate-exchange algorithm to optimze a target function defined by the optimziation paramaters.
+ * The optimization is performed multiple times to prevent finding a design in a local minmum of the target function.
+ *
+ * The method is described in more detail in "Two-Level Designs to Estimate All Main Effects and Two-Factor Interactions",
+ * Eendebak et al., 2015, Technometrics, http://dx.doi.org/10.1080/00401706.2016.1142903.
+ *
  * \param arrayclass Class of designs to optimize
  * \param nrestarts Number of restarts to perform
- * \param alpha Optimization parameters
+ * \param alpha Optimization parameters. The target function is alpha_1 D + alpha_2 D_s + alpha D_1
  * \param verbose Verbosity level
  * \param method Method for optimization algorithm
  * \param niter Maximum number of iterations for each restart
  * \param maxtime Maximum calculation time. If this time is exceeded, the function is aborted
  * \param nabort Maximum number of iterations when no improvement is found
- *
+ * \returns A structure with the generated optimal designs
  */
 DoptimReturn Doptimize (const arraydata_t &arrayclass, int nrestarts, const std::vector< double > alpha, int verbose,
                         int method = DOPTIM_AUTOMATIC, int niter = 300000, double maxtime = 100000, int nabort = 5000);
 
-/// function to generate optimal designs with mixed optimization approach
+/** Function to generate optimal designs with mixed optimization approach
+ *
+ * This function is beta code. See Doptimize for detauls of the  parameters.
+ *
+ */
 DoptimReturn DoptimizeMixed (const arraylist_t &sols, const arraydata_t &arrayclass, const std::vector< double > alpha,
                              int verbose = 1, int nabort = -1);
 
