### 2018-12-22 f677f83
add name for doptim methods enum
