### 2018-12-22 eendebakpt/oapackage#95 @peendebak
@gsagnol The errors have been fixed in #105, no update of `unittest` should be necessary. The entries in the information matrix are large (so is the condition number of M), but this is fine since the input array is a special case used for testing (in the output you see `exampleArray: design in D(28, 2^5), D-efficiency is low`).
