### 2018-12-23 0bc49c0
add tests for exampleArray; fix compilation issue with DOPTIM_METHODS
