### 2018-12-23 31c4c47
Fix bug in examplearray (#109)
* fix missing line in exampleArray case 51

* add tests for exampleArray; fix compilation issue with DOPTIM_METHODS

* bump version
