### 2018-12-26 eendebakpt/oapackage#102 @eendebakpt
@gsagnol The use of keyword arguments is a little bit hard over different versions of python/swig. I updated the example to use the version without keyword arguments.

I didn't know about `sphinx-doctest`, thanks for the suggestion! I started using it, except for the examples where the algorithm uses random numbers and the output is not exactly reproducable.
