### 2018-12-26 d733803
enable doctest on appveyor
