### 2018-12-26 fb297a5
cleanup code
