### 2018-12-26 a5ba186
fix scanf warnings
