### 2018-12-26 ad84e3e
add make to appveyor
