### 2018-12-26 b8730ad
add sphinx to appveyor install
