+++ appveyor.yml
@@ -127,6 +127,9 @@ install:
   - "%CMD_IN_ENV% pip install mock"
   - "%CMD_IN_ENV% python -c \"import mock\""
 
+  # for running doctest
+  - "%CMD_IN_ENV36% pip install sphinx"
+
   # lint the code
   - "%CMD_IN_ENV36% pip install pylint"
   # fix issues with incompatible package versions
