### 2018-12-27 4e78331
add doctest to appveyor
