### 2018-12-27 1494bf8
fix bug in oalib.i due to rename
