### 2018-12-27 1447d1b
more informative error messages
