### 2018-12-27 fbc1e5a
fix appveyor file
