+++ appveyor.yml
@@ -156,7 +156,9 @@ test_script:
    - "%CMD_IN_ENV_ARCH64% %PYTHON%/python -c \"import oapackage; print(oapackage)\""
 
    - ps: cd docs
-   - "%CMD_IN_ENV_ARCH64% dir && make doctest"
+   - ps: dir
+   - "%CMD_IN_ENV_ARCH64% dir"
+   - "%CMD_IN_ENV_ARCH64% make doctest"
    - ps: cd ..
    
    # only run nbconvert for python3.6
