### 2018-12-27 edbd730
update docs
