### 2019-01-03 eendebakpt/oapackage#102 @gsagnol
For sphinx-doctest, you could also seed the random number generator (cf. `np.random.seed`) so the output is reproducible.)
