### 2019-01-06 2304b7d
disable doctest on appveyor
