+++ appveyor.yml
@@ -128,16 +128,14 @@ install:
   - "%CMD_IN_ENV% python -c \"import mock\""
 
   # for running doctest
-  - "%CMD_IN_ENV36% pip install sphinx"
+  #- "%CMD_IN_ENV36% pip install sphinx"
+  #- cinst -y make
 
-  # lint the code
-  - "%CMD_IN_ENV36% pip install pylint"
   # fix issues with incompatible package versions
   - "%CMD_IN_ENV35PLUS% pip install --upgrade prompt_toolkit"
   - "%CMD_IN_ENV35PLUS% %PYTHON%/python -c \"import prompt_toolkit; print(prompt_toolkit.__version__)\""
 
   - cinst -y swig
-  - cinst -y make
 
 build: off
 
@@ -145,8 +143,6 @@ build_script:
   # Build the compiled extension
   - "%CMD_IN_ENV% python setup.py build"
   - "%CMD_IN_ENV% %PYTHON%/python setup.py test"
-  - make -v
-  - "%CMD_IN_ENV_ARCH64% cd docs && make doctest"
 
 
 test_script:
