### 2019-01-06 ee39b19
Misc updates (#111)
* enable doctest on appveyor

* cleanup code

* test appveyor

* test appveyor

* fix scanf warnings

* add make to appveyor

* add sphinx to appveyor install

* add doctest to appveyor

* fix bug in oalib.i due to rename

* get info from appveyor

* more informative error messages

* fix appveyor file

* test appveyor

* test appveyor

* check make version

* update docs

* disable doctest on appveyor
