### 2019-01-10 112d2c6
refactor GWLP method
