+++ docs/generation.rst
@@ -4,13 +4,12 @@ Generation of designs
 The OApackage can be used to generate several classes of arrays. A collection of
 arrays generated by the package is available on the website http://www.pietereendebak.nl/oapackage/index.html.
 
+.. toctree::
+ :maxdepth: 2
+ 
 
 The main functions:
 
-.. autosummary::
-
-        oapackage.Doptim.Doptimize
-        oapackage.oalib.generateConferenceExtensions
         
         
 Generation of orthogonal arrays
@@ -96,8 +95,11 @@ For details see :cite:`Schoen2018dsd`, :cite:`wiki:ConferenceMatrix`.
 The full interface for conference designs is available
 in the :ref:`Interface for conference designs`.
 
-.. comment 
-    .. doxygenfile:: conference.h
+xx .. autosummary::
+
+        oapackage.Doptim.Doptimize
+        oapackage.oalib.generateConferenceExtensions
+
 
 Calculation of D-optimal designs
 --------------------------------
@@ -114,20 +116,22 @@ A Python script to generate optimal designs with 40 runs and 7 factors is shown 
    
        import oapackage
        
-   .. code-block:: python
+   .. .. code-block:: python
+
+   .. doctest::  
 
      >>> N=40; s=2; k=7;
      >>> arrayclass=oapackage.arraydata_t(s, N, 0, k) 
      >>> print('We generate optimal designs with: %s' % arrayclass)
-     We generate optimal designs with: arrayclass: N 40, k 7, strength 0, s 2,2,2,2,2,2,2, order 0.
+     We generate optimal designs with: arrayclass: N 40, k 7, strength 0, s {2,2,2,2,2,2,2}, order 0
      >>> alpha=[1,2,0] 
-     >>> scores, dds, designs, ngenerated = oapackage.Doptimize(arrayclass, nrestarts=40, optimfunc=alpha, selectpareto=True)
-     Doptim: optimization class 40.2-2-2-2-2-2-2
+     >>> oapackage.seedfastrand(123)
+     >>> oapackage.set_srand(123)
+     >>> scores, dds, designs, ngenerated = oapackage.Doptimize(arrayclass, nrestarts=40, optimfunc=alpha, selectpareto=True, verbose=0)
      Doptimize: iteration 0/40
      Doptimize: iteration 39/40
-     Doptim: done (8 arrays, 0.6 [s]) 
-     >>> print('Generated %d designs, the best D-efficiency is %.4f’ % (len(designs), dds[:,0].max() ))
-     Generated 8 designs, the best D-efficiency is 0.9098
+     >>> print('Generated %d designs, the best D-efficiency is %.4f' % (len(designs), dds[:,0].max() ))
+     Generated 10 designs, the best D-efficiency is 0.9198
 
 The parameters of the :meth:`~oapackage.Doptim.Doptimize` function are documented in the code.
 
@@ -141,14 +145,16 @@ to calculate the :math:`D`-, :math:`D_s`- and :math:`D_1`-efficiency.
 For details see :cite:`EendebakSO`.
 
 The OApackage also allows to compute the projection estimation 
-capacity (PEC) sequence for orthogonal arrays :cite:`loeppky2004ranking`. This can be calculated with the following command:
+capacity (PEC) sequence for orthogonal arrays :cite:`loeppky2004ranking`.
+This can be calculated with the following commands :meth:`~oalib.PECsequence` and :meth:`~oalib.array_link.PECsequence`.
 
-.. doxygenfunction:: PECsequence(const array_link&, int)
-    :no-link:
-    :outline:
-.. doxygenfunction:: array_link::PECsequence()
-    :no-link:
-    :outline:
+.. comment
[…]
