+++ docs/generation.rst
@@ -3,14 +3,7 @@ Generation of designs
 
 The OApackage can be used to generate several classes of arrays. A collection of
 arrays generated by the package is available on the website http://www.pietereendebak.nl/oapackage/index.html.
-
-.. toctree::
- :maxdepth: 2
- 
-
-The main functions:
-
-        
+      
         
 Generation of orthogonal arrays
 -------------------------------
@@ -65,9 +58,6 @@ with entries 0, -1, +1 such that i) in each column the symbol 0 occurs
 exactly one time and ii) all columns are orthogonal to each other.
 For details see :cite:`Schoen2018dsd`, :cite:`wiki:ConferenceMatrix`.
 
-.. .. code-block:: python
- :caption: Generate conference designs with 8 rows
-
 .. admonition:: Generate conference designs with 8 rows
 
    .. doctest::     
@@ -92,13 +82,15 @@ For details see :cite:`Schoen2018dsd`, :cite:`wiki:ConferenceMatrix`.
     number of non-isomorphic conference designs: 11
 
 
+An example notebook with more functionality is
+:ref:`Generation and analysis of conference designs`.
 The full interface for conference designs is available
 in the :ref:`Interface for conference designs`.
 
-xx .. autosummary::
 
-        oapackage.Doptim.Doptimize
-        oapackage.oalib.generateConferenceExtensions
+The low-level functions for generating extension columns of conference and double conference designs
+are :meth:`~oapackage.oalib.generateConferenceExtensions` and
+:meth:`~oapackage.oalib.generateDoubleConferenceExtensions`.
 
 
 Calculation of D-optimal designs
@@ -146,7 +138,7 @@ For details see :cite:`EendebakSO`.
 
 The OApackage also allows to compute the projection estimation 
 capacity (PEC) sequence for orthogonal arrays :cite:`loeppky2004ranking`.
-This can be calculated with the following commands :meth:`~oalib.PECsequence` and :meth:`~oalib.array_link.PECsequence`.
+This can be calculated with the following commands :meth:`~oalib.PECsequence` and :meth:`oalib.array_link.PECsequence`.
 
 .. comment
     .. doxygenfunction:: PECsequence(const array_link&, int)
