### 2019-01-10 9a38e34
update docstrings
