### 2019-01-11 206b292
update docstrings
