### 2019-01-11 19e6c8d
update docstrings
