### 2019-01-11 8993509
Improve documentation (#113)
* update generation.rst

* update documentation

* update docstrings

* update docstrings

* update docstrings

* update docstrings
