+++ docs/generation.rst
@@ -3,15 +3,7 @@ Generation of designs
 
 The OApackage can be used to generate several classes of arrays. A collection of
 arrays generated by the package is available on the website http://www.pietereendebak.nl/oapackage/index.html.
-
-
-The main functions:
-
-.. autosummary::
-
-        oapackage.Doptim.Doptimize
-        oapackage.oalib.generateConferenceExtensions
-        
+      
         
 Generation of orthogonal arrays
 -------------------------------
@@ -66,9 +58,6 @@ with entries 0, -1, +1 such that i) in each column the symbol 0 occurs
 exactly one time and ii) all columns are orthogonal to each other.
 For details see :cite:`Schoen2018dsd`, :cite:`wiki:ConferenceMatrix`.
 
-.. .. code-block:: python
- :caption: Generate conference designs with 8 rows
-
 .. admonition:: Generate conference designs with 8 rows
 
    .. doctest::     
@@ -93,11 +82,16 @@ For details see :cite:`Schoen2018dsd`, :cite:`wiki:ConferenceMatrix`.
     number of non-isomorphic conference designs: 11
 
 
+An example notebook with more functionality is
+:ref:`Generation and analysis of conference designs`.
 The full interface for conference designs is available
 in the :ref:`Interface for conference designs`.
 
-.. comment 
-    .. doxygenfile:: conference.h
+
+The low-level functions for generating extension columns of conference and double conference designs
+are :meth:`~oapackage.oalib.generateConferenceExtensions` and
+:meth:`~oapackage.oalib.generateDoubleConferenceExtensions`.
+
 
 Calculation of D-optimal designs
 --------------------------------
@@ -114,20 +108,22 @@ A Python script to generate optimal designs with 40 runs and 7 factors is shown 
    
        import oapackage
        
-   .. code-block:: python
+   .. .. code-block:: python
+
+   .. doctest::  
 
      >>> N=40; s=2; k=7;
      >>> arrayclass=oapackage.arraydata_t(s, N, 0, k) 
      >>> print('We generate optimal designs with: %s' % arrayclass)
-     We generate optimal designs with: arrayclass: N 40, k 7, strength 0, s 2,2,2,2,2,2,2, order 0.
+     We generate optimal designs with: arrayclass: N 40, k 7, strength 0, s {2,2,2,2,2,2,2}, order 0
      >>> alpha=[1,2,0] 
-     >>> scores, dds, designs, ngenerated = oapackage.Doptimize(arrayclass, nrestarts=40, optimfunc=alpha, selectpareto=True)
-     Doptim: optimization class 40.2-2-2-2-2-2-2
+     >>> oapackage.seedfastrand(123)
+     >>> oapackage.set_srand(123)
+     >>> scores, dds, designs, ngenerated = oapackage.Doptimize(arrayclass, nrestarts=40, optimfunc=alpha, selectpareto=True, verbose=0)
      Doptimize: iteration 0/40
      Doptimize: iteration 39/40
-     Doptim: done (8 arrays, 0.6 [s]) 
-     >>> print('Generated %d designs, the best D-efficiency is %.4f’ % (len(designs), dds[:,0].max() ))
-     Generated 8 designs, the best D-efficiency is 0.9098
+     >>> print('Generated %d designs, the best D-efficiency is %.4f' % (len(designs), dds[:,0].max() ))
+     Generated 10 designs, the best D-efficiency is 0.9198
 
 The parameters of the :meth:`~oapackage.Doptim.Doptimize` function are documented in the code.
 
@@ -141,14 +137,16 @@ to calculate the :math:`D`-, :math:`D_s`- and :math:`D_1`-efficiency.
 For details see :cite:`EendebakSO`.
 
[…]
