### 2019-01-12 1b4588e
improve documentation
