### 2019-01-12 #1097 @eendebakpt
@danielskatz We started addressing the remaining comments.
