### 2019-01-12 18f2f02
link to c++ docs instead of python docs
