### 2019-01-12 3b01c5b
Improve documentation (#114)
* improve documentation

* replace docs with autogenerated docs

* link to c++ docs instead of python docs

* link to c++ docs instead of python docs
