+++ src/conference.h
@@ -21,6 +21,11 @@ For more information see:
 #include "arraytools.h"
 #include "graphtools.h"
 
+#ifdef SWIG
+%ignore CandidateGeneratorBase;
+#endif
+
+
 /// print a candidate extension
 void print_column(const conference_column &column, const char *msg = 0);
 
@@ -38,22 +43,6 @@ void showCandidates (const std::vector< conference_column > &column_candidates);
  * \param add_zeros If True, then append a row of zeros
  */
 array_link conference2DSD(const array_link &conference_design, bool add_zeros = 1);
-
-/// structure to cache a list of candidate extensions
-struct conf_candidates_t {
-      public:
-        /// list of candidate extentions for each number of columns
-        std::vector< std::vector< conference_column > > ce;
-
-        /// print information about the set of candidate extentions
-        void info (int verbose = 1) const {
-                for (int i = 2; i < (int)ce.size (); i++) {
-                        if (verbose) {
-                                myprintf ("generateCandidateExtensions: k %d: %d candinates\n", i, (int)ce[i].size ());
-                        }
-                }
-        }
-};
 
 /// Structure representing the type of conference designs
 class conference_t {
@@ -255,7 +244,7 @@ arraylist_t selectLMC0doubleconference (const arraylist_t &list, int verbose, co
 /// select conference arrays in LMC0 form
 arraylist_t selectLMC0 (const arraylist_t &list, int verbose, const conference_t &ctype);
 
-/** Generate candidate extensions (wrapper function)
+/** Generate candidate extensions for a conference design
  *
  * \param array Design to be extended
  * \param conference_type Class of conference designs
@@ -263,23 +252,24 @@ arraylist_t selectLMC0 (const arraylist_t &list, int verbose, const conference_t
  * \param verbose Verbosity level
  * \param filtersymm If True, filter based on symmetry
  * \param filterj2 If True, filter based on J2 values
+ * \return List of generated extensions
  */
 std::vector< conference_column > generateConferenceExtensions (const array_link &array, const conference_t &conference_type,
                                                    int zero_index, int verbose = 1, int filtersymm = 1,
                                                    int filterj2 = 1);
 
 /** Generate candidate extensions for restricted isomorphism classes */
-std::vector< conference_column > generateConferenceRestrictedExtensions (const array_link &al, const conference_t &conference_type,
+std::vector< conference_column > generateConferenceRestrictedExtensions (const array_link &array, const conference_t &conference_type,
                                                              int zero_index, int verbose = 1, int filtersymm = 1,
                                                              int filterip = 1);
 
-// generate extensions for double conference matrices in LMC0 form
-std::vector< conference_column > generateDoubleConferenceExtensions (const array_link &al, const conference_t &conference_type,
+/// generate extensions for double conference matrices in LMC0 form
+std::vector< conference_column > generateDoubleConferenceExtensions (const array_link &array, const conference_t &conference_type,
                                                          int verbose = 1, int filtersymm = 1, int filterip = 1,
                                                          int filterJ3 = 0, int filtersymminline = 1);
 
-// generate extensions for conference matrices in LMC0 form
-std::vector< conference_column > generateSingleConferenceExtensions (const array_link &al, const conference_t &conference_type,
+/// generate extensions for conference matrices in LMC0 form
+std::vector< conference_column > generateSingleConferenceExtensions (const array_link &array, const conference_t &conference_type,
                                                          int zero_index, int verbose, int filtersymm, int filterj2,
                                                          int filterj3, int filtersymminline = 0);
 
@@ -353,7 +343,7 @@ class DconferenceFilter {
         symmdata sd;
 
       public:
[…]
