+++ docs/rtd-condalist.yml
@@ -11,6 +11,7 @@ dependencies:
 - swig
 - ipython
 - jupyter
+- sphinx=1.7
 
 - pip:
   - scanf
