+++ paper/paper.bib
@@ -5,12 +5,19 @@
   PUBLISHER ={Wiley}
 }
 
-@article{Schoen2018dsd,
-title={A Classification Criterion for Definitive Screening Designs},
-author={Eric D. Schoen and Pieter Thijs Eendebak and Peter Goos},
-journal = {The Annals of Statistics},
-year={2018},
-note={Published online}
+@article{Schoen2019,
+author = "Schoen, Eric D. and Eendebak, Pieter T. and Goos, Peter",
+doi = "10.1214/18-AOS1723",
+fjournal = "The Annals of Statistics",
+journal = "Ann. Statist.",
+month = "04",
+number = "2",
+pages = "1179--1202",
+publisher = "The Institute of Mathematical Statistics",
+title = "A classification criterion for definitive screening designs",
+url = "https://doi.org/10.1214/18-AOS1723",
+volume = "47",
+year = "2019"
 }
 
 @Article{Goos2011,
@@ -113,15 +120,15 @@ eprint = {/oup/backfile/content_public/journal/biomet/82/3/10.1093/biomet/82.3.5
 
 @online{EendebakOAwebsite,
   author = {P.T. Eendebak},
-  title = {The Orthogonal Array Package},
+  title = {The {O}rthogonal {A}rray {P}ackage: results},
   year = 2018,
   url = {http://www.pietereendebak.nl/oapackage/},
-  urldate = {2018-11-12}
+  urldate = {2019-1-1}
 }
 
 @online{SloaneOAwebsite,
   author = {N.J.A. Sloane},
-  title = {Orthogonal Arrays: Theory and Applications},
+  title = {{O}rthogonal {A}rrays: {T}heory and {A}pplications},
   year = 1999,
   url = {http://neilsloane.com/doc/OA.html},
   urldate = {2018-11-12}
