### 2019-01-24 8b63d8f
update docs on lmc check
