+++ utils/oacheck.cpp
@@ -173,41 +173,43 @@ int main (int argc, char *argv[]) {
                 index = afile->read_array (al);
                 array_t *array = al.array;
 
-                arraydata_t ad = arraylink2arraydata (al, 0, strength);
-                ad.lmc_overflow_check ();
+                arraydata_t arrayclass = arraylink2arraydata (al, 0, strength);
+                arrayclass.lmc_overflow_check ();
 
-                dyndata_t dynd = dyndata_t (ad.N);
                 OAextend oaextend;
 
                 lmc_t result = LMC_NONSENSE;
 
-                LMCreduction_t *reduction = new LMCreduction_t (&ad);
-                LMCreduction_t *randtest = new LMCreduction_t (&ad);
-                array_t *testarray = clone_array (array, ad.N, ad.ncols);
+                LMCreduction_t *reduction = new LMCreduction_t (&arrayclass);
+                LMCreduction_t *randtest = new LMCreduction_t (&arrayclass);
+				array_link test_arraylink = al.clone();
+				array_t *testarray = test_arraylink.array;
                 randtest->transformation->randomize ();
                 reduction->setArray ( al); 
                 /* variables needed within the switch statement */
                 switch (mode) {
                 case MODE_CHECK:
-                        /* LMC test with reduction code */
-                        reduction->mode = OA_TEST;
-                        result = LMCreduce (array, array, &ad, &dynd, reduction, oaextend);
-                        break;
+				{
+					/* LMC test with reduction code */
+					reduction->mode = OA_TEST;
+					result = LMCcheck(array, arrayclass, oaextend, *reduction);
+					break;
+				}
                 case MODE_CHECKJ4: {
                         /* LMC test with special code */
                         reduction->mode = OA_TEST;
                         reduction->init_state = COPY;
-                        array_link al (array, ad.N, ad.ncols, -10);
-                        oaextend.setAlgorithm (MODE_J4, &ad);
-                        result = LMCcheck (al, ad, oaextend, *reduction);
+                        array_link al (array, arrayclass.N, arrayclass.ncols, -10);
+                        oaextend.setAlgorithm (MODE_J4, &arrayclass);
+                        result = LMCcheck (al, arrayclass, oaextend, *reduction);
                         break;
                 }
                 case MODE_REDUCEJ4: {
                         /* LMC test with special code */
                         reduction->mode = OA_REDUCE;
-                        array_link al (array, ad.N, ad.ncols, -10);
+                        array_link al (array, arrayclass.N, arrayclass.ncols, -10);
                         reduction->setArray (al);
-                        result = LMCcheckj4 (al, ad, *reduction, oaextend);
+                        result = LMCcheckj4 (al, arrayclass, *reduction, oaextend);
                         break;
                 }
                 case MODE_CHECK_SYMMETRY: {
@@ -215,37 +217,37 @@ int main (int argc, char *argv[]) {
 					exit(1);
                 }
                 case MODE_CHECKJ5X: {
-                        oaextend.setAlgorithm (MODE_J5ORDERX, &ad);
+                        oaextend.setAlgorithm (MODE_J5ORDERX, &arrayclass);
 
                         /* LMC test with special code */
                         reduction->mode = OA_TEST;
                         reduction->init_state = COPY;
 
-                        result = LMCcheck (al, ad, oaextend, *reduction);
+                        result = LMCcheck (al, arrayclass, oaextend, *reduction);
                         break;
                 }
                 case MODE_CHECKJ5XFAST: {
-                        oaextend.setAlgorithm (MODE_J5ORDERX, &ad);
+                        oaextend.setAlgorithm (MODE_J5ORDERX, &arrayclass);
 
                         /* LMC test with special code */
[…]
