### 2019-01-25 f99ec70
add unit tests for LMCcheck; remove private functions from public interface; update docs
