### 2019-01-25 154c4f9
fix doctest
