### 2019-01-25 ff9bf11
fix issue with arrays with small number of columns
