### 2019-01-25 5835d51
cleanup interface; update docs
