### 2019-01-25 a3724d7
fix doctest
