### 2019-01-25 827c8f1
Improve LMCcheck (#117)
* update docs on lmc check

* add unit tests for LMCcheck; remove private functions from public interface; update docs

* fix doctest

* debug corner case

* fix issue with arrays with small number of columns

* cleanup interface; update docs

* update docs

* fix doctest
