+++ src/lmc.cpp
@@ -1322,6 +1322,10 @@ int fastj3 (carray_t *array, rowindex_t N, const int J, const colindex_t *pp) {
         return (jval);
 }
 
+// forward declaration
+lmc_t LMCreduction(array_t const *original, array_t const *array, const arraydata_t *ad, const dyndata_t *dyndata,
+	LMCreduction_t *reduction, const OAextend &oaextend);
+
 lmc_t LMCcheckj4 (array_link const &al, arraydata_t const &adin, LMCreduction_t &reduction, const OAextend &oaextend,
                   int jj) {
         LMCreduction_helper_t &tmpStatic = * (acquire_LMCreduction_object () );
@@ -1352,7 +1356,7 @@ lmc_t LMCcheckj4 (array_link const &al, arraydata_t const &adin, LMCreduction_t 
                 // fall back to original method
                 dyndata_t dyndata (adin.N);
 
-                lmc_t ret = LMCreduce (al.array, al.array, &adin, &dyndata, &reduction, oaextend);
+                lmc_t ret = LMCreduction (al.array, al.array, &adin, &dyndata, &reduction, oaextend);
                 return ret;
         }
 
@@ -1563,7 +1567,7 @@ int jj45split (carray_t *array, rowindex_t N, int jj, const colperm_t comb, cons
                 adfix.show_colgroups ();
         }
 
-        ret = LMCreduce (array, array, &adfix, &dyndata, &reduction, oaextend);
+        ret = LMCreduction (array, array, &adfix, &dyndata, &reduction, oaextend);
 
         if (verbose)
                 myprintf ("  ret %d\n", ret);
@@ -1580,6 +1584,39 @@ inline double jj452double (const double *ww) {
         for (size_t i = 0; i < 6; i++)
                 val = val * 64 + ww[i];
         return val;
+}
+
+/// return true if target is in root form, otherwise return false
+inline bool check_root_form(const array_t *array, const arraydata_t &arrayclass) {
+	array_t *root = create_array(arrayclass.N, arrayclass.strength);
+	create_root(root, &arrayclass);
+	if (std::equal(array, array + arrayclass.N * arrayclass.strength, root)) {
+		destroy_array(root);
+		return true;
+	}
+	else {
+		destroy_array(root);
+		return false;
+	}
+}
+
+/// return 0 if target is equal to original, otherwise return 1 and copy root initialization + 1
+inline int check_root_update(carray_t *original, const arraydata_t &arrayclass, array_t *target) {
+	int changed = 0;
+
+	array_t *root = create_array(arrayclass.N, arrayclass.strength);
+	create_root(root, &arrayclass);
+	if (!std::equal(original, original + arrayclass.N * arrayclass.strength, root)) {
+		copy_array(root, target, arrayclass.N, arrayclass.strength);
+		if (arrayclass.ncols>arrayclass.strength) {
+		  for (int j = 0; j < arrayclass.N; j++)
+			  target[arrayclass.N * arrayclass.strength + j] = arrayclass.s[arrayclass.strength] + 100;
+	}
+		changed = 1;
+	}
+	destroy_array(root);
+
+	return changed;
 }
 
 /// helper function to calculate J-values
@@ -1619,8 +1656,7 @@ jj45_t jj45val (carray_t *array, rowindex_t N, int jj, const colperm_t comb, int
 }
 
 
-lmc_t LMCcheckj5 (array_link const &al, arraydata_t const &adin, LMCreduction_t &reduction, const OAextend &oaextend,
-                  int hack) {
+lmc_t LMCcheckj5 (array_link const &al, arraydata_t const &adin, LMCreduction_t &reduction, const OAextend &oaextend) {
         const int dverbose = 0;
[…]
