### 2019-01-26 e414f65
more informative error messages
