### 2019-01-26 b5c6386
more informative error messages
