### 2019-01-26 bd87cba
cleanup code
