### 2019-01-26 61a6742
Cleanup code and documentation (#118)
* more informative error messages

* more informative error messages

* fix unit tests

* cleanup code

* cleanup docs
