+++ docs/properties.rst
@@ -1,10 +1,37 @@
 Properties of designs
 =====================
 
-This section shows the statistical properties of the arrays and designs generated by the Orthogonal Array package. Most properties of an array can be calculated using the :class:`~oalib.array_link`
-object. Some properties are calculated using a function from the package
+This section shows the structural and statistical properties of the orthogonal arrays, conference designs and D-efficient designs generated by the Orthogonal Array package. The properties of the arrays and designs are calculated using the :class:`~oalib.array_link` 
+object or functions from the package.
 
-For example, to calculate the rank of an array and determine whether the array is a foldover array, one can use:
+Before introducing the structural and statistical properties, we define orthogonal arrays, conference designs and D-efficient designs:
+
+An orthogonal array (OA) of strength :math:`{t}`, :math:`{N}` runs and
+:math:`{n}` factors at :math:`{s}` levels is an :math:`{N}\times {n}`
+array of symbols :math:`0,
+\ldots,({s}-1)`, such that for every subset of :math:`{t}` columns,
+every :math:`{t}`-tuple occurs equally
+often :cite:`Rao1947`. The set of all strength-:math:`{t}` OAs with 
+:math:`{N}` runs and :math:`{n}` factors at :math:`{s}` levels 
+is denoted by :math:`{\operatorname{OA}({N}; {t}; {s}^{n})}`. If :math:`{s=2}`, the OA is called a two-level OA and the set of all strength-:math:`{t}` two-level OAs with 
+:math:`{N}` runs and :math:`{n}` factors is denoted as :math:`{\operatorname{OA}({N}; {t}; {2}^{n})}`.  
+
+For :math:`{N}` even, a conference design :cite:`Schoen2018dsd` :math:`C` is 
+an :math:`{N}\times {n}` array which satisfies :math:`{C}^{T}C = (n-1) I_{n}`,
+with :math:`{C}_{ii} = 0` and :math:`{C}_{ij} \in \{-1,1\}`, for 
+:math:`{i} \neq {j}` and :math:`{i}, {j} = 1, \ldots, n`. A :math:`{N}\times {N}` conference design :math:`E` such that :math:`E{E}^{T} = (n-1) I_{n}` is called a conference matrix; see :cite:`Elster1995`, :cite:`Colbourn2006` and :cite:`Xiao2012`. 
+
+Let :math:`{X}` be the :math:`{N}\times {p}` interaction model matrix consisting of a column of ones and the contrast vectors associated to the main and two-factor interactions of :math:`{n}` factors, where :math:`{p = 1 + n + (n)(n-1)/2}`. A D-efficient design :cite:`Donev2007` (:math:`D`) is an :math:`{N}\times {n}` array that maximizes the :math:`D`-efficiency, define as :math:`{(\operatorname{det}({X}^{T}{X})^{1/p})/N}`, where the model matrix :math:`{X}` is constructed using :math:`D`.
+
+Structural properties of an array
+----------------------------------
+
+The OApackage can calculate the rank of an array, defined as the maximum number of linearly independent column or row vectors in the array. The rank of an array is useful for several other functions in the package. For two-level arrays, the OApackage can also check if the arrays are foldover arrays. A two-level array is called a foldover array if half of its runs are mirror images of the other half, in the sense that the factor levels are swapped. 
+
+.. doxygenfunction:: array_link::rank
+.. doxygenfunction:: array_link::foldover
+
[…]
