+++ docs/references.bib
@@ -1,641 +1,610 @@
-% This file was created with JabRef 2.7b.
-% Encoding: UTF-8
+%% This BibTeX bibliography file was created using BibDesk.
+%% http://bibdesk.sourceforge.net/
 
-@PREAMBLE{
-"\def\cprime{$'$} "
-}
 
-@STRING{SV = {Springer-Verlag}}
+%% Created for Alan R Vazquez at 2019-01-31 11:46:27 +0100 
+
+
+%% Saved with string encoding Unicode (UTF-8) 
+
+
+@preamble{"\def\cprime{$'$} "}
+
+
+@string{sv = {Springer-Verlag}}
+
+
+@article{Loeppky2007,
+	Author = {Loeppky, J. L. and Sitter, R. R. and Tang, B.},
+	Date-Added = {2019-01-31 11:45:08 +0100},
+	Date-Modified = {2019-01-31 11:46:27 +0100},
+	Journal = {Technometrics},
+	Pages = {454-467},
+	Title = {Nonregular Designs With Desirable Projection Properties},
+	Volume = {49},
+	Year = {2007}}
+
+@article{Deng2002,
+	Author = {Deng, L.Y. and Tang, B.},
+	Date-Added = {2019-01-15 13:18:21 +0100},
+	Date-Modified = {2019-01-15 13:19:26 +0100},
+	Journal = {Technometrics},
+	Pages = {173-184},
+	Title = {Design Selection and Classification for Hadamard Matrices using Generalized Minimum Aberration Criteria},
+	Volume = {44},
+	Year = {2002}}
+
+@book{Colbourn2006,
+	Author = {Colbourn, C. J. and Dinitz, J. H.},
+	Date-Added = {2019-01-15 10:38:54 +0100},
+	Date-Modified = {2019-01-15 10:39:54 +0100},
+	Publisher = {CRC Press},
+	Title = {Handbook of Combinatorial Designs},
+	Year = {2006}}
+
+@article{Schoen2018,
+	Author = {Schoen E. D., Eendebak, P. T., Goos, P.},
+	Date-Added = {2019-01-15 10:31:31 +0100},
+	Date-Modified = {2019-01-15 10:32:28 +0100}}
 
 @misc{wiki:ConferenceMatrix,
-   author = "Wikipedia",
-   title = "Conference matrix --- Wikipedia{,} The Free Encyclopedia",
-   year = "2016",
-   url = "\url{https://en.wikipedia.org/w/index.php?title=Conference_matrix&oldid=702633334}",
-   note = "[Online; accessed 4-April-2016]"
- }
+	Author = {Wikipedia},
+	Note = {[Online; accessed 4-April-2016]},
+	Title = {Conference matrix --- Wikipedia{,} The Free Encyclopedia},
+	Url = {\url{https://en.wikipedia.org/w/index.php?title=Conference_matrix&oldid=702633334}},
+	Year = {2016},
+	Bdsk-Url-1 = {https://en.wikipedia.org/w/index.php?title=Conference_matrix&oldid=702633334}}
 
 @book{Wu2009,
-  author={Wu, C.F.J. and Hamada, M.S.},
-  year={2009},
-  title={Experiments: Planning, Analysis and Optimization},
-  PUBLISHER ={Wiley}
-}
+	Author = {Wu, C.F.J. and Hamada, M.S.},
+	Publisher = {Wiley},
+	Title = {Experiments: Planning, Analysis and Optimization},
+	Year = {2009}}
 
- 
-@ARTICLE{Angelopoulos2007,
-  author = {Angelopoulos, P. and Evangalaras, H. and Koukouvinos, C. and Lappas,
-	E.},
-  title = {An effective step-down algorithm for the construction and the identification
-	of nonisomorphic orthogonal arrays},
-  journal = {Metrika},
[…]
