### 2019-02-01 dc51d54
fix documentation building; add PICsequence
