+++ docs/properties.rst
@@ -21,17 +21,25 @@ an :math:`{N}\times {n}` array which satisfies :math:`{C}^{T}C = (n-1) I_{n}`,
 with :math:`{C}_{ii} = 0` and :math:`{C}_{ij} \in \{-1,1\}`, for 
 :math:`{i} \neq {j}` and :math:`{i}, {j} = 1, \ldots, n`. A :math:`{N}\times {N}` conference design :math:`E` such that :math:`E{E}^{T} = (n-1) I_{n}` is called a conference matrix; see :cite:`Elster1995`, :cite:`Colbourn2006` and :cite:`Xiao2012`. 
 
-Let :math:`{X}` be the :math:`{N}\times {p}` interaction model matrix consisting of a column of ones and the contrast vectors associated to the main and two-factor interactions of :math:`{n}` factors, where :math:`{p = 1 + n + (n)(n-1)/2}`. A D-efficient design :cite:`Donev2007` (:math:`D`) is an :math:`{N}\times {n}` array that maximizes the :math:`D`-efficiency, define as :math:`{(\operatorname{det}({X}^{T}{X})^{1/p})/N}`, where the model matrix :math:`{X}` is constructed using :math:`D`.
+Let :math:`{X}` be the :math:`{N}\times {p}` interaction model matrix consisting of a column of
+ones and the contrast vectors associated to the main and two-factor interactions
+of :math:`{n}` factors, where :math:`{p = 1 + n + (n)(n-1)/2}`.
+A D-efficient design :cite:`Donev2007` (:math:`D`) is an :math:`{N}\times {n}` array that maximizes
+the :math:`D`-efficiency, define as :math:`{(\operatorname{det}({X}^{T}{X})^{1/p})/N}`,
+where the model matrix :math:`{X}` is constructed using :math:`D`.
 
 Structural properties of an array
 ----------------------------------
 
-The OApackage can calculate the rank of an array, defined as the maximum number of linearly independent column or row vectors in the array. The rank of an array is useful for several other functions in the package. For two-level arrays, the OApackage can also check if the arrays are foldover arrays. A two-level array is called a foldover array if half of its runs are mirror images of the other half, in the sense that the factor levels are swapped. 
+The OApackage can calculate the rank of an array, defined as the maximum number of linearly independent column
+or row vectors in the array. The rank of an array is useful for several other functions in the package.
+For two-level arrays, the OApackage can also check if the arrays are foldover arrays.
+A two-level array is called a foldover array if half of its runs are mirror images of the other half,
+in the sense that the factor levels are swapped. 
 
-.. doxygenfunction:: array_link::rank
-.. doxygenfunction:: array_link::foldover
-
-For example, to calculate the rank of a two-level orthogonal array and determine whether the array is a foldover array, one can use:
+For example, to calculate the rank of a two-level orthogonal array and determine whether
+the array is a foldover array, one can use
+the methods :cpp:func:`array_link::rank` and :cpp:func:`array_link::foldover`:
 
 .. testsetup::
    
@@ -41,44 +49,43 @@ For example, to calculate the rank of a two-level orthogonal array and determine
 
   .. doctest:: 
    
-     >>> array = oapackage.exampleArray(1) # Select an example two-level orthogonal array
-     >>> array.showarray() # Show the two-level orthogonal array
-     array:  
-     0   0   0   0   0
-     0   0   0   0   0
-     0   0   0   1   1
-     0   0   1   0   1
-     0   1   0   1   0
-     0   1   1   0   0
-     0   1   1   1   1
-     0   1   1   1   1
-     1   0   0   1   1
[…]
