+++ docs/references.bib
@@ -1,13 +1,3 @@
-%% This BibTeX bibliography file was created using BibDesk.
-%% http://bibdesk.sourceforge.net/
-
-
-%% Created for Alan R Vazquez at 2019-01-31 11:46:27 +0100 
-
-
-%% Saved with string encoding Unicode (UTF-8) 
-
-
 @preamble{"\def\cprime{$'$} "}
 
 
@@ -42,10 +32,6 @@
 	Title = {Handbook of Combinatorial Designs},
 	Year = {2006}}
 
-@article{Schoen2018,
-	Author = {Schoen E. D., Eendebak, P. T., Goos, P.},
-	Date-Added = {2019-01-15 10:31:31 +0100},
-	Date-Modified = {2019-01-15 10:32:28 +0100}}
 
 @misc{wiki:ConferenceMatrix,
 	Author = {Wikipedia},
