+++ docs/properties.rst
@@ -1,10 +1,14 @@
 Properties of designs
 =====================
 
-This section shows the structural and statistical properties of the orthogonal arrays, conference designs and D-efficient designs generated by the Orthogonal Array package. The properties of the arrays and designs are calculated using the :class:`~oalib.array_link` 
+This section shows the structural and statistical properties of the
+orthogonal arrays, conference designs and D-efficient designs generated by
+the Orthogonal Array package. The properties of the arrays and designs are
+calculated using the :cpp:class:`array_link` 
 object or functions from the package.
 
-Before introducing the structural and statistical properties, we define orthogonal arrays, conference designs and D-efficient designs:
+Before introducing the structural and statistical properties,
+we define orthogonal arrays, conference designs and D-efficient designs:
 
 An orthogonal array (OA) of strength :math:`{t}`, :math:`{N}` runs and
 :math:`{n}` factors at :math:`{s}` levels is an :math:`{N}\times {n}`
@@ -140,7 +144,7 @@ Moreover, the package can calculate the well-known :math:`A`- and :math:`E`-opti
 on Optimal Experimental Design :cite:`Donev2007`.
 The functions to perform the calulcations are
 :cpp:func:`array_link::Defficiency`,
-:cpp:func:`array_link::DsEfficiency`,
+:cpp:func:`array_link::Defficiencies`,
 :cpp:func:`array_link::Aefficiency`,
 :cpp:func:`array_link::Eefficiency`.
 
@@ -152,15 +156,15 @@ The following example shows how to calculate the :math:`D`-, :math:`{D}_{s}`-, :
      
      # Select an array that can estimate the interaction model
      >>> al = oapackage.exampleArray(11, 1)
-     exampleArray: D-optimal array in OA(44, 2^8)
+     exampleArray 11: D-optimal array in OA(44, 2^8)
      >>> print('D-efficiency: %.4f' % al.Defficiency())
-     0.8879
+     D-efficiency: 0.8879
      >>> print('Ds-efficiency (Eendebak and Schoen, 2017): %.4f' % al.Dsefficiency()) 
-     0.8879
+     Ds-efficiency (Eendebak and Schoen, 2017): 0.8879
      >>> print('A-efficiency for the interaction model: %.4f' % al.Aefficiency())
-     0.7906
+     A-efficiency for the interaction model: 0.7906
      >>> print('E-efficiency for the interaction model: %.4f' % al.Eefficiency())  
-     0.3602
+     E-efficiency for the interaction model: 0.3602
 
 .. topic:: Calculation of :math:`D`-, :math:`A`- and :math:`E`-efficiency
    :name: DAE
@@ -195,10 +199,13 @@ include the so-called projection estimation capacity (PEC) and
 projection information capacity (PIC) :cite:`Loeppky2007`. These criteria focus on the projections of the two-level design onto a smaller number of factors. More specifically, the PEC and PIC summarize the performance of all the :math:`N`-run subdesigns with :math:`l \leq k` factors in terms of the capacity to estimate the interaction model and the :math:`D`-efficiency for this model, respectively. 
 
 The PEC and PIC are based on the so-called PEC and PIC sequences, which are formally defined as follows.
[…]
