### 2019-02-01 ebffde3
update docs on external packages
