+++ docs/references.bib
@@ -569,6 +569,26 @@
 	Year = {2012},
 	Bdsk-Url-1 = {http://docs.scipy.org/doc/numpy/reference/}}
 
+
+@misc{WK19,
+	Author = {Warren Kuhfeld},
+	Note = {\url{http://www.stat.ucla.edu/~hqxu/nsf/}},
+	Title = {SAS},
+	Year = {2018}}
+
+@misc{HongquanXuOnline,
+	Author = {Hongquan Xu },
+	Note = {\url{http://www.stat.ucla.edu/~hqxu/nsf/}},
+	Title = {DoE.base: Full Factorials, Orthogonal Arrays and Base Utilities for DoE Packages},
+	Year = {2018}}
+
+@misc{DoEbase,
+	Author = {Ulrike Groemping and Boyko Amarov and Hongquan Xu},
+	Note = {\url{https://cran.r-project.org/package=DoE.base}},
+	Title = {DoE.base: Full Factorials, Orthogonal Arrays and Base Utilities for DoE Packages},
+	Year = {2018}}
+
+
 @misc{Sloanewebsite,
 	Author = {Sloane, N.J.A.},
 	Note = {\url{http://neilsloane.com/oadir/index.html}},
@@ -697,3 +717,5 @@
 	Url = {http://www.ietf.org/rfc/rfc1952.txt},
 	Year = 1996,
 	Bdsk-Url-1 = {http://www.ietf.org/rfc/rfc1952.txt}}
+
+
