+++ docs/references.bib
@@ -572,7 +572,7 @@
 
 @misc{WK19,
 	Author = {Warren Kuhfeld},
-	Note = {\url{http://www.stat.ucla.edu/~hqxu/nsf/}},
+	Note = {\url{http://support.sas.com/techsup/technote/ts723.html}},
 	Title = {SAS},
 	Year = {2018}}
 
