### 2019-02-01 2e2d8d1
fix docs; cleanup code
