+++ src/arraytools.cpp
@@ -3554,31 +3554,6 @@ void jstructconference_t::calcJvalues(int N, int jj) {
 }
 void jstructconference_t::calc(const array_link &al) { values = Jcharacteristics_conference(al, this->jj); }
 
-/// create J2 table as intermediate result for J-characteristic calculations for conference matrices
-array_link createJ2tableConference (const array_link &confmatrix) {
-        const int nr = confmatrix.n_rows;
-        const int nc = (confmatrix.n_columns + 1) * confmatrix.n_columns / 2;
-        // fill double column table
-        array_link dtable (nr, nc, -1);
-
-        // loop over all column pairs
-        int idx = 0;
-        for (int i = 0; i < confmatrix.n_columns; i++) {
-                for (int j = 0; j <= i; j++) {
-                        // loop over all rows of original array
-                        const array_t *p1 = confmatrix.array + confmatrix.n_rows * i;
-                        const array_t *p2 = confmatrix.array + confmatrix.n_rows * j;
-                        array_t *pout = dtable.array + idx * dtable.n_rows;
-                        for (int x = 0; x < nr; x++) {
-                                pout[x] = p1[x] * p2[x];
-                        }
-                        idx++;
-                }
-        }
-
-        return dtable;
-}
-
 /// create J2 table as intermediate result for J-characteristic calculations
 array_link createJdtable (const array_link &al) {
         const int nr = al.n_rows;
