### 2019-02-01 eendebakpt/oapackage#98 @eendebakpt
@gsagnol We created a separate section that points to other sources of orthogonal arrays.
