### 2019-02-01 4a901e2
Update introduction with section or other source of designs (#122)
* update docs on external packages

* update url

* update intro
