+++ src/md5.h
@@ -5,6 +5,6 @@
 #pragma once
 
 /// calculate md5 sum of a data block in memory
-std::string md5 (void *data, int numbytes);
+std::string md5 (void *data, int number_of_bytes);
 /// calculate md5 sum of a file on disk
 std::string md5 (const std::string filename);
