### 2019-02-01 8e5a896
improve writearrayfile interface; update docs
