+++ src/arraytools.cpp
@@ -4756,9 +4756,9 @@ int readarrayfile (const char *fname, arraylist_t *arraylist, int verbose, int *
         return i;
 }
 
-int writearrayfile (const char *fname, const arraylist_t arraylist, arrayfile::arrayfilemode_t mode, int nrows,
+int writearrayfile (const char *filename, const arraylist_t &arraylist, arrayfile::arrayfilemode_t mode, int nrows,
                     int ncols) {
-        return writearrayfile (fname, &arraylist, mode, nrows, ncols);
+        return writearrayfile (filename, &arraylist, mode, nrows, ncols);
 }
 
 /**
@@ -5415,10 +5415,10 @@ void vectorvector2binfile(const std::string fname, const std::vector< std::vecto
 	fclose(fid);
 }
 
-int writearrayfile (const char *fname, const array_link &al, arrayfile::arrayfilemode_t mode) {
+int writearrayfile (const char *filename, const array_link &array, arrayfile::arrayfilemode_t mode) {
         arraylist_t s;
-        s.push_back (al);
-        return writearrayfile (fname, &s, mode);
+        s.push_back (array);
+        return writearrayfile (filename, s, mode);
 }
 
 /// append a single array to an array file. creates a new file if no file exists
