### 2019-02-02 ca0dba4
cleanup code
