### 2019-02-02 837a8f2
update doctests
