### 2019-02-02 30202be
Release 2.5.7 (#123)
* Update introduction with section or other source of designs (#122)

* update docs on external packages

* update url

* update intro

* release 2.5.7

* update docs

* add missing header

* fix docs

* improve writearrayfile interface; update docs

* cleanup code

* update doctests

* fix build

* update docs
