### 2019-02-02 877a001
misc fixes
