### 2019-02-03 5ab9921
Fix compilation error under unix (#124)
* misc fixes

* misc fixes

* fix compilation

* cleanup docs

* bump version
