+++ docs/properties.rst
@@ -165,7 +165,7 @@ The following example shows how to calculate the :math:`D`-, :math:`{D}_{s}`-, :
      >>> print('D-efficiency: %.4f' % al.Defficiency())
      D-efficiency: 0.8879
      >>> print('Ds-efficiency (Eendebak and Schoen, 2017): %.4f' % al.DsEfficiency()) 
-     Ds-efficiency (Eendebak and Schoen, 2017): 0.8879
+     Ds-efficiency (Eendebak and Schoen, 2017): 0.8227
      >>> print('A-efficiency for the interaction model: %.4f' % al.Aefficiency())
      A-efficiency for the interaction model: 0.7906
      >>> print('E-efficiency for the interaction model: %.4f' % al.Eefficiency())  
