### 2019-02-03 5b49a2e
refactor conference_t
