### 2019-02-03 47557eb
fix compilation; update docs
