+++ docs/generation.rst
@@ -12,7 +12,7 @@ A list of arrays in LMC form (i.e., lexicographically minimum in columns) can be
 form with one additional column. Details about the algorithm are described
 in :cite:`Eendebak2009`.
 
-The main function for array extension is the function :cpp:function:`extend_arraylist`. The signature is
+The main function for array extension is the function :cpp:func:`extend_arraylist`. The signature is
 
 
 .. doxygenfunction:: extend_arraylist(const arraylist_t&, arraydata_t&, OAextend const&)
@@ -23,8 +23,6 @@ arrays and :cpp:class:`OAextend` contains various options for the algorithm.
 
 An example of a session that extends a designs is:
 
-.. .. code-block:: python
-   :caption: Extend an array
 .. doctest::
    
    >>> import oapackage
@@ -60,7 +58,7 @@ For details see :cite:`Schoen2018dsd`, :cite:`wiki:ConferenceMatrix`.
                    
     >>> import oapackage
     >>> ctype=oapackage.conference_t(8, 8, 0) 
-    >>> al = ctype.create_root_three()
+    >>> al = ctype.create_root_three_columns()
     >>> al.showarray()
     array:
       0   1   1
