+++ src/arrayproperties.h
@@ -77,35 +77,43 @@ std::vector< double > PICsequence(const array_link &array, int verbose = 0);
  */
 std::vector< double > distance_distribution (const array_link &array);
 
-/// Calculate J-characteristics of matrix (the values are signed)
-std::vector< int > Jcharacteristics (const array_link &al, int jj = 4, int verbose = 0);
+/** Calculate Jk-characteristics of a matrix
+ *
+ * The calcualted Jk-values are signed.
+ * 
+ * \param array Array to calculate Jk-characteristics for
+ * \param number_of_columns Number of columns
+ * \param verbose Verbosity level
+ * \returns Vector with calculated Jk-characteristics
+ */
+std::vector< int > Jcharacteristics (const array_link &array, int number_of_columns = 4, int verbose = 0);
 
 /** @brief Calculate GWLP (generalized wordlength pattern)
  *
  * The method used for calculation is from Xu and Wu (2001), "Generalized minimum aberration for asymmetrical
  * fractional factorial desings"
  * For non-symmetric arrays see "Algorithmic Construction of Efficient Fractional Factorial Designs With Large Run
- * Sizes", Xu
+ * Sizes", Xu, Technometrics, 2009.
  *
  * \param array Array to calculate the GWLP value for
  * \param verbose Verbosity level
  * \param truncate If True then round values near zero to solve double precision errors
+ * \returns Vector with calculated generalized wordlength pattern
  */
 std::vector< double > GWLP (const array_link &array, int verbose = 0, int truncate = 1);
 
 /** @brief Calculate GWLP (generalized wordlength pattern) for mixed-level arrays
 *
-*  The method used for calculation is from Xu and Wu (2001), "Generalized minimum aberration for asymmetrical
-* fractional factorial desings"
-*  The non-symmetric arrays see "Algorithmic Construction of Efficient Fractional Factorial Designs With Large Run
-* Sizes", Xu
+*  The method used for calculation is from "Algorithmic Construction of Efficient Fractional Factorial Designs With Large Run
+* Sizes", Xu, Technometrics, 2009.
 *
-* \param al Array to calculate the GWLP value for
+* \param array Array to calculate the GWLP value for
 * \param verbose Verbosity level
 * \param truncate If True then round values near zero to solve double precision errors
+* \returns Vector with calculated generalized wordlength pattern
 *
 */
-std::vector< double > GWLPmixed (const array_link &al, int verbose = 0, int truncate = 1);
+std::vector< double > GWLPmixed (const array_link &array, int verbose = 0, int truncate = 1);
 
 // SWIG has some issues with typedefs, so we use a define
 #define GWLPvalue mvalue_t< double >
@@ -313,9 +321,9 @@ inline mvalue_t< long > F4 (const array_link &al, int verbose = 1) {
 }
 
 template < class IndexType >
-/** Add array to list of Pareto optimal arrays
+/** Calculate properties of an array and create a Pareto element
  *
- * The values to be optimized are:
+ * The values calculated are:
  *
  * 1) Rank (higher is better)
  * 2) A3, A4 (lower is better)
@@ -324,13 +332,13 @@ template < class IndexType >
  * Valid for 2-level arrays of strength at least 3
  *
  * */
-typename Pareto< mvalue_t< long >, IndexType >::pValue calculateArrayParetoRankFA (const array_link &al,
+typename Pareto< mvalue_t< long >, IndexType >::pValue calculateArrayParetoRankFA (const array_link &array,
                                                                                           int verbose) {
-        int N = al.n_rows;
[…]
