### 2019-02-03 df4bc69
fix notebooks; add unit tests
