### 2019-02-03 eendebakpt/oapackage#101 @eendebakpt
@gsagnol The documentation was modified to link to the functions instead of including the full signature.
