### 2019-02-03 3814723
Update documentation on conference designs (#126)
* fix compilation; update docs

* update documentation

* cleanup docs

* update documentation

* update code

* update docs

* fix notebooks; add unit tests

* fix notebooks
