+++ src/conference.h
@@ -90,14 +90,14 @@ class conference_t {
         /// create the unique representative of the 2 column conference design in LMC0 form
         array_link create_root () const;
 
-        /// create the unique representative of the 3 column conference design  in LMC0 form
-        array_link create_root_three () const;
+        /// create the unique representative of the 3 column conference design in LMC0 form
+        array_link create_root_three_columns () const;
 
         /// create the root arrays with 1 column for the double conference matrices
         arraylist_t createDoubleConferenceRootArrays () const;
 
-        /// returnt the list of root arrays for the class of conference designs
-		arraylist_t createRootArrays (arraylist_t) const;
+        /// return the list of root arrays for the class of conference designs
+	arraylist_t createRootArrays () const;
 
         /// return string representation of the object
         std::string __repr__ () const {
@@ -223,6 +223,14 @@ class CandidateGeneratorDouble : public CandidateGeneratorBase {
 };
 
 /** Extend a list of conference designs with a single column.
+ * 
+ * The list of conference designs is extended by adding each design with the candidate extentions generated by @ref CandidateGenerator.
+ * 
+ * \param lst List of conference designs
+ * \param conference_type Type specification for the conference designs
+ * \param verbose Verbosity level
+ * \param select_isomorphism_classes If True then select only a single design for each isomorphism class specified by the conference type.
+ * \returns List of generated conference designs 
  *
  */
 arraylist_t extend_conference (const arraylist_t &lst, const conference_t conference_type, int verbose,
@@ -235,7 +243,18 @@ arraylist_t extend_conference_plain (const arraylist_t &lst, const conference_t 
 /** Extend a list of conference designs with a single column */
 arraylist_t extend_conference_restricted (const arraylist_t &lst, const conference_t conference_type, int verbose);
 
-// extend a list of double conference matrices with an additional column
+/** Extend a list of double conference matrices with an additional column
+ *
+ * The list of designs is extended by adding each design with the candidate extentions generated by @ref CandidateGeneratorDouble.
+ * 
+ * \param lst List of double conference designs
+ * \param conference_type Type specification for the conference designs
+ * \param conference_type Type specification for the double conference designs
+ * \param verbose Verbosity level
+ * \param select_isomorphism_classes If True then select only a single design for each isomorphism class specified by the conference type.
+ * \returns List of generated double conference designs 
+ *
+ */
 arraylist_t extend_double_conference (const arraylist_t &lst, const conference_t conference_type, int verbose);
 
 /// select representatives for the isomorphism classes of a list of conference arrays

[…]
