### 2019-02-04 #1097 @eendebakpt
@gsagnal @danielskatz

Thank you for the extensive review of our package. We tried to addressed the issues you raised on the OApackage issue tracker. In particular:

https://github.com/eendebakpt/oapackage/issues/92: The github site now points to https://oapackage.readthedocs.io/. The webssite with generated results in mentioned on the readthedocs documentation.
https://github.com/eendebakpt/oapackage/issues/95: Issues have been fixed in https://github.com/eendebakpt/oapackage/issues/96 (make tests python 2.7 compatible, resolve issue with multiple scanf versions)
https://github.com/eendebakpt/oapackage/issues/97: The documentation for Doptimize was improved in several PRs. The referee points out that our software does not find globally optimal solutions. We now explicitly mention in the documentation that our software does not guarantee to find the global optimum. The generation of the model matrix still needs better documentation. The compound criterium is explained in detail in the paper to which we refer.
https://github.com/eendebakpt/oapackage/issues/98: We updated the introducion text (https://github.com/eendebakpt/oapackage/issues/122)
https://github.com/eendebakpt/oapackage/issues/99, https://github.com/eendebakpt/oapackage/issues/100: We added warnings to the Doptim function to let the user know that this method does not use the strength variable. The DOPTIM_UPDATE variable was indeed not used in the example, so we removed it. See https://github.com/eendebakpt/oapackage/issues/104, https://github.com/eendebakpt/oapackage/issues/106, https://github.com/eendebakpt/oapackage/issues/107
https://github.com/eendebakpt/oapackage/issues/101: We agree with the author that the complete c++ signature is not needed in the documentation (this style of documentation was started before we could properly generate the documentation with https://breathe.readthedocs.io). We updated the documentation (https://github.com/eendebakpt/oapackage/issues/113, https://github.com/eendebakpt/oapackage/issues/114, https://github.com/eendebakpt/oapackage/issues/119)
https://github.com/eendebakpt/oapackage/issues/102: We use doctest now in the documentation which guarantees the code snippets in the documentation will run without errors. https://github.com/eendebakpt/oapackage/issues/111, https://github.com/eendebakpt/oapackage/issues/113
https://github.com/eendebakpt/oapackage/issues/103: We updated the documentation https://github.com/eendebakpt/oapackage/issues/121

We still need to improve the documentation on the model matrix and the statistics of conference designs. We will do this in the next weeks. Feel free to already look and comment on the improvements we made. A new release with the improvements has already been made on https://pypi.org/project/OApackage/
