+++ src/arraytools.h
@@ -410,26 +410,35 @@ struct array_link {
 		*
 		*/
 		array_link ();
-        /** @copydoc arraydata_t::arraydata_t()
+        /** @copydoc array_link::array_link()
 		 *
-		 * The arary is iitialized with zeros.
+		 * The array is intialized with zeros.
 		 */
         array_link (rowindex_t nrows, colindex_t ncols, int index);
-		/** @copydoc arraydata_t::arraydata_t()
+		/** @copydoc array_link::array_link()
 		*
 		* Initialize with data from a pointer.
 		*/
         array_link (rowindex_t nrows, colindex_t ncols, int index, carray_t *data);
-		/** @copydoc arraydata_t::arraydata_t()
+		/** @copydoc array_link::array_link()
 		*
 		* Initialize with data from another array_link object.
 		*/
         array_link (const array_link &);
-		/** @copydoc arraydata_t::arraydata_t()
+		/** @copydoc array_link::array_link()
 		*
 		* Initialize with data from anEigen matrix.
 		*/
         array_link (Eigen::MatrixXd &eigen_matrix);
+		/// @copydoc array_link::array_link()
+		array_link(const array_link &, const std::vector< int > &colperm);
+		/// @copydoc array_link::array_link()
+		array_link(const array_t *array, rowindex_t nrows, colindex_t ncols, int index = 0);
+		/// @copydoc array_link::array_link()
+		array_link(const array_t *array, rowindex_t nrows, colindex_t ncolsorig, colindex_t ncols, int index);
+		/// @copydoc array_link::array_link()
+		array_link(const std::vector< int > &v, rowindex_t nrows, colindex_t ncols, int index = 0);
+
         ~array_link ();
 
 #ifdef SWIGCODE
@@ -537,10 +546,10 @@ struct array_link {
         std::vector< int > FvaluesConference (int jj) const;
 
         /** Calculate the Jk-characteristics of matrix (the values are signed)
-	 * 
-	 * 
-	 * 
-	 */
+		 * 
+		 * \param jj Number of columns to use
+		 *	\returns Vector with calculated Jk values
+		 */
         std::vector< int > Jcharacteristics (int jj = 4) const;
 
         /// Calculate the projective estimation capacity sequence
@@ -569,7 +578,7 @@ struct array_link {
         // return value of maximum element in array
         array_t max () const;
 
-        /** calculate centered L2 discrepancy
+        /** Calculate centered L2 discrepancy
          *
          * The method is from "A connection between uniformity and aberration in regular fractions of two-level
          * factorials", Fang and Mukerjee, 2000
@@ -607,18 +616,17 @@ struct array_link {
         /// elementwise addition
         array_link operator+ (const array_link &) const;
         /// elementwise addition
-        array_link operator+ (array_t v) const;
+        array_link operator+ (array_t value) const;
         array_link operator- (const array_link &) const;
-        array_link operator- (array_t v) const;
+        array_link operator- (array_t value) const;
 
         /// elementwise multiplication
         array_link operator* (const array_link &rhs) const;
 
-        array_link operator* (array_t val) const;
-
-        array_link operator*= (array_t val);
-        array_link operator+= (array_t val);
-        array_link operator-= (array_t val);
+        array_link operator* (array_t value) const;
+        array_link operator*= (array_t value);
+        array_link operator+= (array_t value);
+        array_link operator-= (array_t value);
 
         /// get element from array, no error checking, inline version
         inline const array_t &atfast (const rowindex_t r, const colindex_t c) const {
@@ -700,12 +708,6 @@ struct array_link {
 
         /// set column to values
         void setcolumn (int target_column, const array_link &source_array, int source_column = 0) const;
-
-      public:
-        array_link (const array_link &, const std::vector< int > &colperm);
[…]
