### 2019-02-05 e205051
Misc updates (#129)
* fix hyperlink

* update documentation

* fix docstrings

* fix build
