+++ docs/properties.rst
@@ -230,9 +230,25 @@ The following example shows how to compute the PEC and PIC sequences of a two-le
 Properties of conference designs
 --------------------------------
 
+In :cite:`Schoen2018dsd` it is shown that the :math:`F_4` vector is usefull in
+classifying definitieve screening designs that are generated by folding over a conference design.
+
+.. admonition:: Calculate the F4 vector for a conference design
+
+  .. doctest:: 
+
+    >>> import oapackage
+    >>> array=oapackage.exampleArray(47,1)
+    exampleArray 47: third conference design in C(20,8)
+    >>> F4=array.FvaluesConference(4)
+    >>> print(F4)
+    (0, 2, 4, 51, 13)
+    >>> definitive_screening_design = oapackage.conference2DSD(array)
+
+The individual :math:`J_k`-characteristics can be calculated with the method :cpp:func:`Jcharacteristics_conference`.
+
 For conference designs we can calculate the projection statistics using
-:meth:`conferenceProjectionStatistics`. The Jk-characteristics can be calculated with the method
-:cpp:func:`Jcharacteristics_conference`.
+:meth:`~oalib.conferenceProjectionStatistics`.
 
 .. admonition:: Calculate projection statistics for conference designs
 
