### 2019-02-08 #1097 @gsagnol
@eendebakpt
Nice to see the progress ! I will wait for the improved documentation to complete my review.

I still have a concern about eendebakpt/oapackage#97: I really think writing the formula for the criterion you optimize is necessary, for several reasons:
1. This is a software, and my philosophy is that software users should not need to read a full scientific paper to use it. Of course, the article must be cited in the doc, but at least the formula you use in your software to evaluate the quality of a design should be in the doc.
2. This is a journal about *open source* software, and you are referring to an article hosted on a commercial website.
