### 2019-02-08 1ab22cd
update documentation for model matrix; fix documentation building; add arrayproperties to docs
