+++ docs/cppdocs/arrayproperties.rst
@@ -1,0 +1,5 @@
+Interface for array properties
+==============================
+
+.. doxygenfile:: arrayproperties.h
+
