+++ src/arrayproperties.h
@@ -180,11 +180,12 @@ array_link conference_design2modelmatrix(const array_link & conference_design, c
 
 /** Convert orthogonal array or conference design to model matrix
  *
- * Intercept, main effects, interaction effects, quadratics
- * The order in the interaction effects is (c1, c2)=(0,0), (1,0), (2,0), (2,1), ... with c2<c1
+ * The model matrix consists of the intercept, main effects and (optionally) the interaction effects, quadratic effects.
+ * The order in the interaction effects is (c1, c2)=(0,0), (1,0), (2,0), (2,1), ... with c2<c1 for columns c1, c2.
+ * The size of the model matrix calculated by this function is given by @ref array2modelmatrix_sizes.
  *
  * \param array Orthogonal array or conference design
- * \param mode Can be 'm' for main effects, 'i' for interaction effects or 'q' for quadratic effects
+ * \param mode Type of model matrix to calculate. Can be 'm' for main effects, 'i' for interaction effects or 'q' for quadratic effects
  * \param verbose Verbosity level
  * \returns Calculated model matrix
  */
