+++ docs/properties.rst
@@ -86,6 +86,7 @@ Model matrices
 
 For orthogonal arrays and conference designs we can calculate model matrices. For 2-level orthogonal arrays the array is
 converted to a -1,+1 array and then the model matrix is calculated. For mixel-level orthogonal arrays Helmert contrasts are used.
+For more details on the calculation of the model matrices see :cpp:func:`array2modelmatrix`.
 
 .. admonition:: Calculate interaction effects model matrix 
 
