### 2019-02-08 a7afeac
cleanup code
