### 2019-02-09 #1097 @djmitche
Perhaps @danielskatz can chime in here as to the requirements of JOSS?

My understanding is that it's a bridge between the world of publishing and the world of OSS, and as such part of its *purpose* is to inter-reference between articles on commercial websites and open-source applications.

I do agree that the documentation should be self-sufficient in understanding how to *use* the software, but that the documentation need not go into detail as to the implementation.  By comparison, I would not expect the documentation for GIMP to include formal models for its "magic wand" boundary detection algorithm.   Including a formula in oapackage's docs, for comparison to the source code, but referring the user to the paper for the explanation, seems a good compromise.
