### 2019-02-09 47048c8
cleanup array2eigenModelMatrixMixed
