### 2019-02-09 #1097 @danielskatz
> My understanding is that it's a bridge between the world of publishing and the world of OSS, and as such part of its _purpose_ is to inter-reference between articles on commercial websites and open-source applications.

I don't quite agree.  The goal of JOSS is really to give software authors a product that can be cited, and to improve the software along the way. How internally complete the documentation is versus how much it references externally (and potentially commercial) articles is not not so much as factor, as long as a potential user can access as much documentation as they need to use the software.

> Including a formula in oapackage's docs, for comparison to the source code, but referring the user to the paper for the explanation, seems a good compromise.

I think this seems reasonable.

As a part of this compromise, regarding the "article hosted on a commercial website," perhaps a pre-print would be allowed to be deposited on arxiv or similar?
