+++ oapackage/oahelper.py
@@ -240,7 +240,7 @@ def niceplot(ax, fig=None, despine=True, verbose=0, figurebg=True,
 
 def enlargelims(factor=1.05):
     """ Enlarge the limits of a plot
-    
+
     Args:
         factor (float): factor by which to make the plot margins wider
     """
@@ -254,6 +254,50 @@ def enlargelims(factor=1.05):
     plt.ylim(yl)
 
 # %%
+
+
+def helmert_contrasts(number_of_levels, verbose=0):
+    """ Calculate Helmert contrasts for a given number of levels for a number_of_levels
+
+    Args:
+        number_of_levels: number of levels in the number_of_levels
+    Returns:
+        array: array with calculated Helmert contrasts
+    """
+    N = number_of_levels
+    meoffset = 0
+    md = number_of_levels - 1
+
+    main_effects = np.zeros((number_of_levels, md))
+    Z = np.zeros((number_of_levels, md + 1))
+
+    for value in range(number_of_levels):
+        for ii in range(0, md + 1):
+            Z[value, ii] = value > ii - 1
+
+        # make Helmert contrasts (these are automatically orthogonal)
+        Z[value, 0] = 1
+        if (value > 0):
+            Z[value, value] = value
+
+        for q in range(1, value):
+            Z[value, q] = 0
+        for q in range(value + 1, md + 1):
+            Z[value, q] = -1
+
+    if verbose:
+        print('helmert_contrasts: %d\n' % (number_of_levels))
+        print('Z (initial creation)')
+        print(Z)
+
+    # normalize the contrasts
+    for ii in range(0, md):
+        normalization = Z[:, (ii + 1)].T.dot(Z[:, ii + 1])
+        if verbose:
+            print('helmert_contrasts: normalize number_of_levels tmp: %s ' % (normalization,))
+        main_effects[:, meoffset + ii:(meoffset + ii + 1)] = np.sqrt((N)) * Z[:, (ii + 1):(ii + 2)] / np.sqrt((normalization))
+
+    return main_effects
 
 
 def selectParetoArrays(array_list, pareto_object):
@@ -426,7 +470,7 @@ def array2latex(X, header=1, hlines=[], floatfmt='%g', comment=None, hlinespace=
     ss = ''
     if comment is not None:
         if isinstance(comment, list):
-            for line in comment:    
+            for line in comment:
                 ss += '%% %s\n' % str(line)
         else:
             ss += '%% %s\n' % str(comment)
@@ -1002,7 +1046,7 @@ def runExtend(N, k, t=3, l=2, verbose=1, initsols=None, nums=[], algorithm=None)
     """
     if verbose:
         print('runExtend: N=%d, k=%d, t=%d' % (N, k, t))
-    if isinstance(l, list):  
+    if isinstance(l, list):
         ll = l
     else:
         ll = [l]
@@ -1038,7 +1082,7 @@ def runExtend(N, k, t=3, l=2, verbose=1, initsols=None, nums=[], algorithm=None)
 
 def compressOAfile(afile, decompress=False, verbose=1):
     """ Compress an OA array file 
-    
+
     Args:
         afile (str): array to compress
         decompress (bool): If True then decompress
@@ -1047,10 +1091,10 @@ def compressOAfile(afile, decompress=False, verbose=1):
     af = oalib.arrayfile_t(afile, 0)
     if decompress:
         raise NotImplementedError('decompressing file not implemted')
-        
+
     if verbose >= 2:
         print('file %s: binary %s' % (afile, af.isbinary()))
-    if not (sys.platform == 'linux2' or sys.platform=='linux'):
+    if not (sys.platform == 'linux2' or sys.platform == 'linux'):
         if verbose:
             print('compressOAfile: not compressing file %s (platform not supported)' %
                   afile)
