### 2019-02-09 8d3bcd5
fix bug in DsEfficiency calculation; add tests
