### 2019-02-09 275ec36
update docstrings
