### 2019-02-09 #1097 @eendebakpt
@danielskatz @djmitche @gsagnol We all agree that the software must be usable without references, so I will add to the documentation.

I do believe that referencing the paper is the right thing to do. To understand the ideas and motivations behind the code one needs to read the paper, and it doesn't make sense to rewrite the contents of the paper in the documentation of the oapackage.

The link is to a commercial publisher, but using google and the title of the paper as search term a PDF can be found easily.
