### 2019-02-09 21b7a8d
autopep; fixes to tests
