+++ oapackage/tests/test_cpplibrary.py
@@ -41,38 +41,40 @@ import oapackage
 
 
 class TestMinimalFormCheck(unittest.TestCase):
+
     def test_LMCcheck(self):
         array = oapackage.exampleArray(1)
-    
+
         lmc_type = oapackage.LMCcheck(array)
         self.assertEqual(lmc_type, oapackage.LMC_MORE)
 
-        array2 = array.selectColumns([0,1,2,4,3])
+        array2 = array.selectColumns([0, 1, 2, 4, 3])
         lmc_type = oapackage.LMCcheck(array2)
         self.assertEqual(lmc_type, oapackage.LMC_LESS)
 
-        array2 = array.selectColumns([2,1,0,3,4])
+        array2 = array.selectColumns([2, 1, 0, 3, 4])
         lmc_type = oapackage.LMCcheck(array2)
         self.assertEqual(lmc_type, oapackage.LMC_LESS)
 
         arrayclass = oapackage.arraylink2arraydata(array)
-        reduction =oapackage.LMCreduction_t(arrayclass)
-        oaextend=oapackage.OAextend(arrayclass)
-        lmc_type=oapackage.LMCcheck (array2.array, arrayclass, oaextend, reduction)
+        reduction = oapackage.LMCreduction_t(arrayclass)
+        oaextend = oapackage.OAextend(arrayclass)
+        lmc_type = oapackage.LMCcheck(array2.array, arrayclass, oaextend, reduction)
         self.assertEqual(lmc_type, oapackage.LMC_LESS)
-       
+
     def test_LMCcheckOriginal(self):
         array = oapackage.exampleArray(1)
         lmc_type = oapackage.LMCcheckOriginal(array)
         self.assertEqual(lmc_type, oapackage.LMC_MORE)
 
-        array2 = array.selectColumns([0,1,2,4,3])
+        array2 = array.selectColumns([0, 1, 2, 4, 3])
         lmc_type = oapackage.LMCcheck(array2)
         self.assertEqual(lmc_type, oapackage.LMC_LESS)
 
-        array = oapackage.exampleArray(5,0)
+        array = oapackage.exampleArray(5, 0)
         with self.assertRaises(RuntimeError):
             lmc_type = oapackage.LMCcheckOriginal(array)
+
 
 class TestReductions(unittest.TestCase):
 
@@ -85,11 +87,11 @@ class TestReductions(unittest.TestCase):
         alr = oapackage.reduceLMCform(al)
         self.assertTrue(alr == al)
 
-        array0 = oapackage.exampleArray(1,1).selectFirstColumns(3)
+        array0 = oapackage.exampleArray(1, 1).selectFirstColumns(3)
         array = array0.randomperm()
         reduced_array = oapackage.reduceLMCform(array)
         self.assertEqual(array0, reduced_array)
-    
+
     @only_python3
     def test_DOP(self):
         al = oapackage.exampleArray(1, 0)
@@ -143,34 +145,34 @@ class TestModelmatrix(unittest.TestCase):
         self.assertTrue(model_matrix.shape[1] == sizes[3])
 
     def test_array2eigenModelMatrixMixed(self):
-        array=oapackage.exampleArray(0,0);
-        r=oapackage.array2eigenModelMatrixMixed(array.selectFirstColumns(1), verbose=0)
+        array = oapackage.exampleArray(0, 0)
+        r = oapackage.array2eigenModelMatrixMixed(array.selectFirstColumns(1), verbose=0)
         main_effects = r[0]
         np.testing.assert_array_equal(main_effects, np.array([[-1., -1., -1., -1.,  1.,  1.,  1.,  1.]]).T)
-    
[…]
