### 2019-02-09 9ced3db
fix tests; fix Ds calculation
