### 2019-02-09 b42a1c5
Update of model matrix code (#130)
* update documentation for model matrix; fix documentation building; add arrayproperties to docs

* update docs

* update docs

* cleanup code

* remove unnecessary warning

* add tests

* cleanup array2eigenModelMatrixMixed

* add helmert contrasts

* fix typo

* fix bug in DsEfficiency calculation; add tests

* update docstrings

* autopep; fixes to tests

* doc update

* fix tests; fix Ds calculation
