+++ src/conference.h
@@ -78,8 +78,8 @@ class conference_t {
 		/** @copydoc conference_t::conference_t()
 		 *
 		 * \param N Number of rows
-		 * \param k Numbe of columns
-		 * \param 
+		 * \param k Number of columns
+		 * \param j1zero If True then require the J1-characteristics to be zero
 		 */
         conference_t (int N, int k, int j1zero);
         conference_t (const conference_t &rhs);
@@ -248,10 +248,8 @@ arraylist_t extend_conference_restricted (const arraylist_t &lst, const conferen
  * The list of designs is extended by adding each design with the candidate extentions generated by @ref CandidateGeneratorDouble.
  * 
  * \param lst List of double conference designs
- * \param conference_type Type specification for the conference designs
  * \param conference_type Type specification for the double conference designs
  * \param verbose Verbosity level
- * \param select_isomorphism_classes If True then select only a single design for each isomorphism class specified by the conference type.
  * \returns List of generated double conference designs 
  *
  */
