+++ src/arrayproperties.h
@@ -15,8 +15,6 @@
 #include "oaoptions.h"
 #include "tools.h"
 
-#define stringify(name) #name
-
 /// Calculate D-efficiency and VIF-efficiency and E-efficiency values using SVD
 void DAEefficiencyWithSVD (const Eigen::MatrixXd &secondorder_interaction_matrix, double &Deff, double &vif, double &Eeff, int &rank, int verbose);
 
@@ -25,14 +23,22 @@ void DAEefficiencyWithSVD (const Eigen::MatrixXd &secondorder_interaction_matrix
  * The model is the intercept, main effects and interaction effects
  * The rank, D-efficiency, VIF-efficiency and E-efficiency are appended to the second argument
  *
- * The vector ret is filled with the rank, Defficiency, VIF efficiency and Eefficiency
+ * The return vector is filled with the rank, Defficiency, VIF efficiency and Eefficiency
  */
 int array2rank_Deff_Beff (const array_link &al, std::vector< double > *ret = 0, int verbose = 0);
 
 /// Calculate D-efficiency for a 2-level array using symmetric eigenvalue decomposition
-double Defficiency (const array_link &al, int verbose = 0);
+double Defficiency (const array_link &orthogonal_array, int verbose = 0);
 
-std::vector< double > Defficiencies (const array_link &al, const arraydata_t &arrayclass, int verbose = 0,
+/** Calculate efficiencies for an array
+ *
+ * \param array Array to use in calculation
+ * \param arrayclass
+ * \param verbose Verbosity level
+ * \param addDs0 If True, then add the Ds0-efficiency to the output
+ * \returns Vector with the calculate D-efficiency, the main effect robustness (or Ds-optimality) and D1-efficiency for an orthogonal array
+ */
+std::vector< double > Defficiencies (const array_link &array, const arraydata_t &arrayclass, int verbose = 0,
                                      int addDs0 = 0);
 
 /// Calculate VIF-efficiency of matrix
@@ -49,12 +55,12 @@ std::vector< double > Aefficiencies (const array_link &orthogonal_array, int ver
 
 /** Calculate D-efficiencies for all projection designs
 *
-* \param al Design to calculate D-efficiencies for
+* \param array Design to calculate D-efficiencies for
 * \param number_of_factors Number of factors into which to project
 * \param verbose Verbosity level
 * \returns Vector with calculated D-efficiencies
 */
-std::vector< double > projDeff (const array_link &al, int number_of_factors, int verbose = 0);
+std::vector< double > projDeff (const array_link &array, int number_of_factors, int verbose = 0);
 
 /** Calculate the projection estimation capacity sequence for a design
 *
@@ -118,6 +124,7 @@ std::vector< double > GWLPmixed (const array_link &array, int verbose = 0, int t
 // SWIG has some issues with typedefs, so we use a define
 #define GWLPvalue mvalue_t< double >
 
+/// delete-one-factor projection value
 typedef mvalue_t< double > DOFvalue;
 
 /// calculate delete-one-factor GWLP (generalized wordlength pattern) projections
@@ -133,7 +140,7 @@ std::vector< double > projectionGWLPdoublevalues (const array_link &al);
  * The method is from "A connection between uniformity and aberration in regular fractions of two-level factorials",
  * Fang and Mukerjee, 2000
  */
-double CL2discrepancy (const array_link &al);
+double CL2discrepancy (const array_link &array);
 
 /** Calculate second order interaction model for 2-level array
 *
@@ -149,6 +156,19 @@ array_link array2secondorder (const array_link &array);
  */
 array_link array2xf (const array_link &array);
 
+enum model_matrix_t {
+	/// only the intercept
+	MODEL_CONSTANT,
[…]
