### 2019-02-11 3ada46d
refactor Doptim code; update docs of efficiencies
