+++ oapackage/Doptim.py
@@ -15,8 +15,12 @@ import time
 import logging
 import warnings
 
-class MissingMatplotLibException(Exception):
-    pass
+import oalib
+
+import oapackage.markup as markup
+import oapackage.oahelper as oahelper
+from oapackage.markup import oneliner as e
+
 
 try:
     import matplotlib
@@ -25,18 +29,20 @@ try:
 except BaseException:
     matplotlib = None
 
-import oalib
+class MissingMatplotLibException(Exception):
+    pass
 
-import oapackage.markup as markup
-import oapackage.oahelper as oahelper
-
-from oapackage.markup import oneliner as e
 
 # %%
 
 
-def array2Dtable(sols, verbose=1, titlestr=None):
-    """ Generate HTML table with information about for a list of designs """
+def array2Dtable(array_list, verbose=1, titlestr=None):
+    """ Generate HTML table with information about for a list of designs 
+    
+    Args:
+        array_list (list): list of arrays
+        verbose (int): verbosity level
+    """
     page = markup.page()
     page.table(style=' border-collapse: collapse;')
     page.tr(style='font-weight: bold; border-bottom: solid 1px black;')
@@ -45,7 +51,7 @@ def array2Dtable(sols, verbose=1, titlestr=None):
             style='padding-right:14px;')
     page.th(('GWLP'), style='padding-right:14px;')
     page.tr.close()
-    for ii, al in enumerate(sols):
+    for ii, al in enumerate(array_list):
         aidx = ii
         (D, Ds, D1) = al.Defficiencies()
         gwlp = al.GWLP()
@@ -284,13 +290,27 @@ def generateDpage(outputdir, arrayclass, dds, allarrays, fig=20, optimfunc=[1, 0
     return outfile
 
 
+def scoreDn(dds, optimfunc):
+    """ Calculate scores from various efficiencies
+
+    Args:
+        dds (array): calculated D-efficiencies
+        optimfunc (list): parameters for optimization function
+    Returns:
+        scores (array)
+    """
+    scores = np.array([oalib.scoreD(dd, optimfunc) for dd in dds])
+    return scores
+
 def calcScore(dds, optimfunc):
-    """ Calculate D-efficiency score using multiple efficiencies and a weight factor """
-    n = dds.shape[0]
-    scores = np.zeros(n, )
-    for ii, d in enumerate(dds):
-        alpha = optimfunc
-        scores[ii] = alpha[0] * d[0] + alpha[1] * d[1] + alpha[2] * d[2]
+    """ Calculate D-efficiency score using multiple efficiencies and a weight factor
+    
+    Args:
+        dds (array): the rows contains the effieciencies for various designs
+        optimfunc (array): aray with the weight factors for the efficiencies
+    """  
+    scores = np.array(dds).dot(np.array(optimfunc))
+        
     return scores
 
 
@@ -420,19 +440,6 @@ def filterPareto(scores, dds, sols, verbose=0):
     return pscores, pdds, psols
 
 
-def scoreDn(dds, optimfunc):
-    """ Calculate scores from various efficiencies
-
-    Args:
-        dds (array): calculated D-efficiencies
-        optimfunc (list): parameters for optimization function
-    Returns:
-        scores (array)
-    """
-    scores = np.array([oalib.scoreD(dd, optimfunc) for dd in dds])
-    return scores
-
-
 def selectDn(scores, dds, sols, nout=1, sortfull=True):
     """ Select best arrays according to given scores
         The resulting data is sorted
@@ -485,9 +492,6 @@ def Doptimize(arrayclass, nrestarts=10, optimfunc=[
     multiple times (specified by the nrestarts parameter) to prevent finding a design in a local minmum of the
[…]

[…]
