### 2019-02-11 29dd395
improve docs and tests
