### 2019-02-11 ef4d7bd
use decicated type for coordinate-exchange method parameter
