+++ docs/generation.rst
@@ -83,10 +83,10 @@ The low-level functions for generating candidate extension columns of conference
 are :cpp:func:`generateConferenceExtensions` and :cpp:func:`generateDoubleConferenceExtensions`.
 
 
-Calculation of D-optimal designs
---------------------------------
+Calculation of D-efficient designs
+----------------------------------
 
-D-optimal designs can be calculated with the function :py:meth:`oapackage.Doptim.Doptimize`.
+D-efficient designs (sometimes called D-optimal designs) can be calculated with the function :py:meth:`oapackage.Doptim.Doptimize`.
 This function uses a coordinate-exchange algorithm to generate designs
 with good properties for the :math:`D`-efficiency.
 
@@ -116,25 +116,13 @@ A Python script to generate optimal designs with 40 runs and 7 factors is shown 
 The parameters of the :meth:`~oapackage.Doptim.Doptimize` function are documented in the code.
 
 To calculate properties of designs we can use the following functions.
-We can use the following command:
-
-.. doxygenfunction:: array_link::Defficiencies(int)
-    :no-link:
-
-to calculate the :math:`D`-, :math:`D_s`- and :math:`D_1`-efficiency.
-For details see :cite:`EendebakSO`.
+We can use the method :cpp:func:`Defficiencies`, to calculate the :math:`D`-, :math:`D_s`- and :math:`D_1`-efficiency.
+For details of the efficiencies see the section `Optimality criteria for D-efficient designs`_ and :cite:`EendebakSO`.
 
 The OApackage also allows to compute the projection estimation 
 capacity (PEC) sequence for orthogonal arrays :cite:`loeppky2004ranking`.
 This can be calculated with the following commands :meth:`~oalib.PECsequence` and :meth:`oalib.array_link.PECsequence`.
 
-.. comment
-    .. doxygenfunction:: PECsequence(const array_link&, int)
-        :no-link:
-        :outline:
-    .. doxygenfunction:: array_link::PECsequence()
-        :no-link:
-        :outline:
     
 In :cite:`EendebakSO` it is shown that one can optimize a linear combination of the
 :math:`D`-efficiency and :math:`D_s`-efficiency to generate a rich 
