+++ docs/references.bib
@@ -108,7 +108,7 @@
 	Number = {1},
 	Pages = {69-79},
 	Title = {Two-Level Designs to Estimate All Main Effects and Two-Factor Interactions},
-	Url = {https://doi.org/10.1080/00401706.2016.1142903},
+	_Url = {https://doi.org/10.1080/00401706.2016.1142903},
 	Volume = {59},
 	Year = {2017},
 	Bdsk-Url-1 = {https://doi.org/10.1080/00401706.2016.1142903}}
