### 2019-02-11 56793b6
fix hyperlinks in references
