+++ docs/references.bib
@@ -37,7 +37,7 @@
 	Author = {Wikipedia},
 	Note = {[Online; accessed 4-April-2016]},
 	Title = {Conference matrix --- Wikipedia{,} The Free Encyclopedia},
-	Url = {\url{https://en.wikipedia.org/w/index.php?title=Conference_matrix&oldid=702633334}},
+	Url = {https://en.wikipedia.org/w/index.php?title=Conference_matrix&oldid=702633334},
 	Year = {2016},
 	Bdsk-Url-1 = {https://en.wikipedia.org/w/index.php?title=Conference_matrix&oldid=702633334}}
 
@@ -299,8 +299,7 @@
 	Title = {Projection properties of certain three level orthogonal arrays},
 	Url = {http://ideas.repec.org/a/spr/metrik/v62y2005i2p241-257.html},
 	Volume = {62},
-	Year = {2005},
-	Bdsk-Url-1 = {http://ideas.repec.org/a/spr/metrik/v62y2005i2p241-257.html}}
+	Year = {2005} }
 
 @article{Xu2001,
 	Author = {Hongquan Xu and C. F. J. Wu},
@@ -319,7 +318,6 @@
 	Author = {Donev, {A. N.} and Atkinson, {A. C.} and Tobias, {R. D.}},
 	Isbn = {ISBN-13: 978-0-19-929659-0},
 	Language = {English},
-	Note = {Received the Ziegel price from ASA for 2009: http://ukcatalogue.oup.com/product/9780199296590.do?keyword=donev&sortby=bestMatches},
 	Publisher = {Oxford University Press},
 	Series = {Oxford Statistical Science Series},
 	Title = {Optimum Experimental Designs, with SAS},
@@ -328,16 +326,15 @@
 @article{Xiao2012,
 	Author = {Lili Xiao and Dennis K. J. Lin and Fengshan Bai},
 	Doi = {10.1080/00224065.2012.11917877},
-	Eprint = {https://doi.org/10.1080/00224065.2012.11917877},
+	_Eprint = {https://doi.org/10.1080/00224065.2012.11917877},
 	Journal = {Journal of Quality Technology},
 	Number = {1},
 	Pages = {2-8},
 	Publisher = {Taylor \& Francis},
 	Title = {Constructing Definitive Screening Designs Using Conference Matrices},
-	Url = {https://doi.org/10.1080/00224065.2012.11917877},
+	_Url = {https://doi.org/10.1080/00224065.2012.11917877},
 	Volume = {44},
-	Year = {2012},
-	Bdsk-Url-1 = {https://doi.org/10.1080/00224065.2012.11917877}}
+	Year = {2012} }
 
 @article{Elster1995,
 	Abstract = {Screening experiments aim to identify the relevant variables within some process potentially depending on a large number of variables. In this paper we introduce a new class of experimental designs called edge designs. These designs allow a model-independent estimate of the set of relevant variables, thus providing more robustness than traditional designs. We give a bound on the determinant of the information matrix of certain edge designs, and show that a large class of edge designs meeting this bound can be constructed from conference matrices. We also show that the resulting conference designs have an optimal space exploration property which guards against unexpected nonlinearities. We survey the existence of and constructions for conference matrices, and, for $n < 50$ variables, give explicit such matrices when n is a prime, and references to explicit constructions otherwise.},
@@ -511,7 +508,6 @@
 	Pages = {123--140},
 	Publisher = {Wiley Subscription Services, Inc., A Wiley Company},
[…]
