+++ docs/properties.rst
@@ -292,7 +292,6 @@ classifying definitieve screening designs that are generated by folding over a c
     >>> definitive_screening_design = oapackage.conference2DSD(array)
 
 The individual :math:`J_k`-characteristics can be calculated with the method :cpp:func:`Jcharacteristics_conference`.
-
 For conference designs we can calculate the projection statistics using
 :meth:`~oalib.conferenceProjectionStatistics`.
 
