### 2019-02-11 f3c66ea
Improve Doptimize interface (#131)
* fix hyperlink

* refactor Doptim code; update docs of efficiencies

* improve docs and tests

* use decicated type for coordinate-exchange method parameter

* update generation docs

* cleanuorefactor code

* fix notebook

* fix py27 tests

* update docs

* update docs

* fix hyperlinks in references

* update docs

* update docs
