+++ src/arrayproperties.cpp
@@ -533,7 +533,7 @@ std::vector< double > PICsequence(const array_link &array, int verbose) {
 
 	int N = array.n_rows;
 	int number_of_factors = array.n_columns;
-	std::vector< double > pec(number_of_factors);
+	std::vector< double > pic(number_of_factors);
 
 	if (number_of_factors >= 20) {
 		throw_runtime_exception("PICsequence: error: not implemented for 20 or more columns\n");
@@ -544,21 +544,21 @@ std::vector< double > PICsequence(const array_link &array, int verbose) {
 #pragma omp parallel for
 #endif
 	for (int i = 0; i < number_of_factors; i++) {
-		int kp = i + 1;
-		int m = 1 + kp + kp * (kp - 1) / 2;
+		int number_projection_factors = i + 1;
+		int m = 1 + number_projection_factors + number_projection_factors * (number_projection_factors - 1) / 2;
 
 		if (m > N) {
 			// if size of model is larger than number of runs the model cannot be estimated
-			pec[i] = 0;
+			pic[i] = 0;
 		}
 		else {
-			std::vector< double > dd = projDeff(array, kp, verbose >= 2);
+			std::vector< double > dd = projDeff(array, number_projection_factors, verbose >= 2);
 
-			pec[i] = average(dd);
+			pic[i] = average(dd);
 		}
 	}
 
-	return pec;
+	return pic;
 }
 
 void round_GWLP_zero_values(std::vector<double> &gma, int N)
@@ -1518,15 +1518,15 @@ double detXtXfloat (const MyMatrixf &mymatrix, int verbose) {
 // typedef Eigen::MatrixXd MyMatrix;
 typedef MatrixFloat EigenMatrixFloat;
 
-std::vector< double > Defficiencies (const array_link &al, const arraydata_t &arrayclass, int verbose, int addDs0) {
-        if ((al.n_rows > 500) || (al.n_columns > 500)) {
+std::vector< double > Defficiencies (const array_link &array, const arraydata_t &arrayclass, int verbose, int addDs0) {
+        if ((array.n_rows > 500) || (array.n_columns > 500)) {
                 myprintf ("Defficiencies: array size not supported\n");
                 return std::vector< double > (3);
         }
 
-        int k = al.n_columns;
-        int k1 = al.n_columns + 1;
-        int nrows = al.n_rows;
+        int k = array.n_columns;
+        int k1 = array.n_columns + 1;
+        int nrows = array.n_rows;
         int m = 1 + k + k * (k - 1) / 2;
 
         EigenMatrixFloat X;
@@ -1534,39 +1534,39 @@ std::vector< double > Defficiencies (const array_link &al, const arraydata_t &ar
 		/// number of 2-factor interactions in contrast matrix
         int n2fi = -1; 
         /// number of main effects in contrast matrix
-        int nme = -1;  
+        int size_main_effects = -1;  
 
         if (arrayclass.is2level ()) {
 
-                X = array2eigenModelMatrix (al);
+                X = array2eigenModelMatrix (array);
 
                 n2fi = k * (k - 1) / 2;
-                nme = k;
+                size_main_effects = k;
         } else {
                 if (verbose >= 2)
                         myprintf ("Defficiencies: mixed design!\n");
-                std::pair< EigenMatrixFloat, EigenMatrixFloat > mm = array2eigenModelMatrixMixed (al, 0);
+                std::pair< EigenMatrixFloat, EigenMatrixFloat > mm = array2eigenModelMatrixMixed (array, 0);
                 const EigenMatrixFloat &X1 = mm.first;
[…]

[…]
