### 2019-02-12 5fc7351
update docs; add extra example array
