+++ src/arraytools.cpp
@@ -1410,7 +1410,7 @@ array_link exampleArray (int idx, int verbose) {
         }
         case 10: {
                 if (verbose) {
-                        myprintf ("exampleArray %d: array in OA(9, 3^2)\n", idx);
+                        myprintf ("exampleArray %d: array in OA(9, 3^3)\n", idx);
                 }
                 array_link al (9, 3, 0);
                 int tmp[] = {0, 0, 0, 1, 1, 1, 2, 2, 2, 0, 1, 2, 1, 1, 2, 0, 0, 2, 2, 0, 2, 0, 2, 1, 0, 1, 1};
@@ -2039,6 +2039,16 @@ array_link exampleArray (int idx, int verbose) {
 			 }
 			 array_link array(12, 4, 0);
 			 int array_data_tmp[] = { 0,1,1,1,1,1,1,1,1,1,1,1,1,0,1,1,1,1,1,-1,-1,-1,-1,-1,1,-1,0,1,1,-1,-1,1,1,1,-1,-1,1,1,-1,-1,-1,1,1,0,1,1,-1,-1 };
+			 array.setarraydata(array_data_tmp, array.n_rows * array.n_columns);
+			 return array;
+		 }
+		 case 54: {
+			 dstr = "root array in OA(6,2,3^1 2^1)";
+			 if (verbose) {
+				 myprintf("exampleArray %d: %s\n", idx, dstr.c_str());
+			 }
+			 array_link array(6, 2, 0);
+			 int array_data_tmp[] = { 0,0,1,1,2,2,0,1,0,1,0,1 };
 			 array.setarraydata(array_data_tmp, array.n_rows * array.n_columns);
 			 return array;
 		 }
