### 2019-02-12 96d0618
update docs on mixlevel_modelmatrix
