### 2019-02-12 cd45d50
cleanup headers; update doctest script for conference designs
