### 2019-02-12 fd18708
add LMC0 definition
