+++ docs/normalform.rst
@@ -125,3 +125,55 @@ The function :py:meth:`~oalib.reduceOAnauty` reduces an orthogonal array to Naut
       1   1
       1   1
 
+Normal forms for conference designs
+-----------------------------------
+
+For conference designs a convenient normal form is the LMC0 ordering (sometimes also called L0 ordering) :cite:`Schoen2018dsd`.
+
+.. admonition:: LMC0 ordering
+
+ The LMC0 ordering for conference designs is defined in three steps:
+  
+  Definition 1: *Order of elements*
+    The LMC0 order of the factor levels -1, 0 and +1 is 0 < +1 < -1.
+
+  Definition 2: *Order of columns*
+    A column a is smaller than a column b according to L0 ordering, notated as a < b, if either
+    of the following conditions hold:
+       1. If we replace the values -1 by +1 in both columns, then the first element where the
+          columns differ is smaller in a than in b according to Definition 1.
+       2. The zeros in column a are in the same position as the zeros in column b, and the first element where the
+          columns differ is smaller in a than in b according to Definition 1.
+          
+  Definition 3: *Order of designs*
+    Conference design A is smaller than conference design B according to LMC0 ordering, notated
+    as A < B, if the first column where the designs differ is smaller in A than in B.
+
+
+The definition implies that the ordering of designs is column-by-column and that the position of zeros in the columns is dominant over the values +1, -1.
+
+.. admonition:: Conference design in normal form
+ 
+  .. testsetup::
+     
+     import oapackage
+     
+  .. doctest::
+    
+    >>> array = oapackage.exampleArray(53,1)
+    exampleArray 53: third array in C(12,4)
+    >>> array.showarray()
+    array:
+      0   1   1   1
+      1   0  -1   1
+      1   1   0  -1
+      1   1   1  -1
+      1   1   1  -1
+      1   1  -1   1
+      1   1  -1   1
+      1  -1   1   0
+      1  -1   1   1
+      1  -1   1   1
+      1  -1  -1  -1
+      1  -1  -1  -1
+  
