+++ docs/generation.rst
@@ -78,9 +78,10 @@ in the :ref:`Interface for conference designs`.
 
 The main functions to extend conference and double conference designs are
 :cpp:func:`extend_conference` and :cpp:func:`extend_double_conference`.
-
 The low-level functions for generating candidate extension columns of conference and double conference designs
 are :cpp:func:`generateConferenceExtensions` and :cpp:func:`generateDoubleConferenceExtensions`.
+
+The conference designs are generated in :ref:`LMC0` form.
 
 
 Calculation of D-efficient designs
