### 2019-02-12 31a051f
fix compile warnings
