+++ src/arraytools.cpp
@@ -4239,7 +4239,7 @@ arrayfile_t::arrayfile_t (const std::string fname, int nrows, int ncols, int nar
         this->gzfid = 0;
 #endif
 
-        createfile (fname, nrows, ncols, narrays_, m, number_of_bits);
+        createfile (fname, nrows, ncols, narrays_, mode, number_of_bits);
 }
 
 int arrayfile_t::headersize () const { return 8 * sizeof (int32_t); }
