+++ src/conference.h
@@ -38,6 +38,7 @@ void showCandidates (const std::vector< conference_column > &column_candidates);
  *
  * \param conference_design Array with the conference design
  * \param add_zeros If True, then append a row of zeros
+ * \return The DSD generated from the conference design
  */
 array_link conference2DSD(const array_link &conference_design, bool add_zeros = 1);
 
@@ -221,7 +222,7 @@ class CandidateGeneratorDouble : public CandidateGeneratorBase {
 
 /** Extend a list of conference designs with a single column.
  * 
- * The list of conference designs is extended by adding each design with the candidate extentions generated by @ref CandidateGenerator.
+ * The list of conference designs is extended by adding to each design the candidate extentions generated by @ref CandidateGenerator.
  * 
  * \param lst List of conference designs
  * \param conference_type Type specification for the conference designs
@@ -229,15 +230,23 @@ class CandidateGeneratorDouble : public CandidateGeneratorBase {
  * \param select_isomorphism_classes If True then select only a single design for each isomorphism class specified by the conference type.
  * \returns List of generated conference designs 
  *
+ * The extension algorithm tried to generate designs in LMC0 normal form and prune any designs that are not in LMC0 form.
+ * 
  */
 arraylist_t extend_conference (const arraylist_t &lst, const conference_t conference_type, int verbose,
                                int select_isomorphism_classes = 0);
 
-/// Extend a list of conference designs with a single column, plain version without caching
+/** Extend a list of conference designs with a single column, plain version without caching
+ *
+ * Research function.
+ */
 arraylist_t extend_conference_plain (const arraylist_t &lst, const conference_t conference_type, int verbose,
                                      int select_isomorphism_classes = 0);
 
-/** Extend a list of conference designs with a single column */
+/** Extend a list of conference designs with a single column 
+ *
+ * Research function.
+ */
 arraylist_t extend_conference_restricted (const arraylist_t &lst, const conference_t conference_type, int verbose);
 
 /** Extend a list of double conference matrices with an additional column
@@ -260,10 +269,22 @@ arraylist_t selectConferenceIsomorpismClasses (const arraylist_t &list, int verb
 std::vector< int > selectConferenceIsomorpismIndices (const arraylist_t &lst, int verbose,
                                                       matrix_isomorphism_t itype = CONFERENCE_ISOMORPHISM);
 
-/// select double conference arrays in LMC0 form
+/** Select double conference designs in LMC0 form
+ *
+ * \param list List of double conference designs
+ * \param verbose Verbosity level
+ * \param ctype Specifiation of the class of designs
+ * \return List with only the designs in the input list that are in LMC0 normal form.
+ **/
 arraylist_t selectLMC0doubleconference (const arraylist_t &list, int verbose, const conference_t &ctype);
 
-/// select conference arrays in LMC0 form
+/** Select conference designs in LMC0 form
+*
+* \param list List of conference designs
+* \param verbose Verbosity level
+* \param ctype Specifiation of the class of designs
+* \return List with only the designs in the input list that are in LMC0 normal form.
+**/
 arraylist_t selectLMC0 (const arraylist_t &list, int verbose, const conference_t &ctype);
 
 /** Generate candidate extensions for a conference design
@@ -301,10 +322,10 @@ std::vector< conference_column > generateSingleConferenceExtensions (const array
  */
 int maxz (const array_link &al, int column_index = -1);
 
-/** Return true of the array is smaller in LMC-0 ordering
+/** Return true if the first array is smaller in LMC-0 ordering than the second array
  *
  */
-bool compareLMC0 (const array_link &array_left, const array_link &array_right);
[…]
