### 2019-02-13 f3893b7
Misc updates (#132)
* update documentation

* fix whitespace

* fix hyperlink

* update docs; add extra example array

* rename function arguments

* update docs on mixlevel_modelmatrix

* update docs

* cleanup headers; update doctest script for conference designs

* fix test

* add LMC0 definition

* update docs

* fix compile warnings

* update docs
