### 2019-02-13 c894a29
rewrite model matrix docs
