### 2019-02-13 eaced9a
update docs (#134)
* update docs

* rewrite model matrix docs
