### 2019-02-14 b22f288
update example notebooks
