### 2019-02-14 e637765
add unit tests for Doptim
