### 2019-02-14 4a6e2c4
update docstrings
