### 2019-02-15 264ddef
add unit tests; update performance test script
