### 2019-02-15 fd16c94
Update example notebooks and add tests (#137)
* update example notebooks

* fix notebook

* add unit tests for Doptim

* update gitignore file

* update docstrings

* cleanup code

* cleanup code

* fix py27 compile

* add unit tests; update performance test script
