### 2019-02-15 02a2ecb
remove dead code
