### 2019-02-15 7344e08
Feat/misc updates3 (#139)
* remove dead code

* cleanup code

*  update docs

* fix hyperlink
