+++ oapackage/tests/test_Doptim.py
@@ -35,6 +35,7 @@ def only_python3(function):
             return None
     return only_python3_function
 
+
 class TestDoptimize(unittest.TestCase):
 
     def setUp(self):
@@ -54,7 +55,7 @@ class TestDoptimize(unittest.TestCase):
         def optimfunc(x): return x[0] + x[1] + x[2]
         with mock.patch('sys.stdout', new_callable=io.StringIO) as mock_stdout:
             scores, dds, sols, n = oapackage.Doptim.Doptimize(self.arrayclass, nrestarts=2, optimfunc=optimfunc, verbose=1,
-                                                          maxtime=18, selectpareto=False, nout=None, method=oalib.DOPTIM_UPDATE, niter=1000, nabort=0, dverbose=0)
+                                                              maxtime=18, selectpareto=False, nout=None, method=oalib.DOPTIM_UPDATE, niter=1000, nabort=0, dverbose=0)
         self.assertEqual(len(scores), n)
         self.assertEqual(len(dds), n)
         self.assertEqual(len(sols), n)
@@ -65,24 +66,25 @@ class TestDoptimize(unittest.TestCase):
     def test_Doptimize_nonzero_strength(self):
         arrayclass = oapackage.arraydata_t(2, 16, 2, 6)
         with self.assertWarns(UserWarning):
-            scores, dds, sols, _ = oapackage.Doptim.Doptimize(arrayclass, nrestarts=1, verbose=0, maxtime=9, selectpareto=False, nout=None, niter=1000, dverbose=0)
-        
+            scores, dds, sols, _ = oapackage.Doptim.Doptimize(
+                arrayclass, nrestarts=1, verbose=0, maxtime=9, selectpareto=False, nout=None, niter=1000, dverbose=0)
+
     def test_Doptimize_selectDn(self):
         scores, dds, sols, _ = oapackage.Doptim.Doptimize(self.arrayclass, nrestarts=10, optimfunc=[
                                                           1, 0, 0], verbose=0, maxtime=9, selectpareto=False, nout=None, method=oalib.DOPTIM_UPDATE, niter=1000, nabort=0, dverbose=0)
 
         result = oapackage.Doptim.selectDn(scores, dds, sols, nout=1, sortfull=True)
-        self.assertTrue(len(result[2])==1)
+        self.assertTrue(len(result[2]) == 1)
 
     @only_python3
     def test_optimDeffPython(self):
         al = oapackage.exampleArray(2, 0)
         with mock.patch('sys.stdout', new_callable=io.StringIO) as mock_stdout:
             _, al = oapackage.Doptim.optimDeffPython(
-                    al, arrayclass=None, niter=1000, nabort=1500, verbose=1, alpha=[1, 0, 0], method=0)
+                al, arrayclass=None, niter=1000, nabort=1500, verbose=1, alpha=[1, 0, 0], method=0)
             stdout = mock_stdout.getvalue()
             self.assertTrue(stdout.startswith('optimDeff: initial Deff 0.0000'))
-            
+
         for method in [oapackage.oalib.DOPTIM_SWAP, oapackage.oalib.DOPTIM_FLIP, oapackage.oalib.DOPTIM_UPDATE]:
             r, al = oapackage.Doptim.optimDeffPython(
                 al, arrayclass=None, niter=100, nabort=200, verbose=0, alpha=[1, 0, 0], method=method)
@@ -91,9 +93,9 @@ class TestDoptimize(unittest.TestCase):
     def test_generateDscatter(self):
         if self.guitest:
             fig = 100
-            with mock.patch('sys.stdout', new_callable=io.StringIO) as mock_stdout:                
[…]
