+++ oapackage/tests/test_cpplibrary.py
@@ -122,12 +122,12 @@ class TestModelmatrix(unittest.TestCase):
 
     def test_modelmatrix_main_effects(self):
         # test model matrix main effects are helmert contrasts
-        array=oapackage.array_link( np.array([[0,1,2,3]]).T)
-        
-        M=oapackage.array2modelmatrix(array, 'm')
-        hc=oapackage.oahelper.helmert_contrasts(4)
-        np.testing.assert_array_equal(hc, M[:,1:])
-    
+        array = oapackage.array_link(np.array([[0, 1, 2, 3]]).T)
+
+        M = oapackage.array2modelmatrix(array, 'm')
+        hc = oapackage.oahelper.helmert_contrasts(4)
+        np.testing.assert_array_equal(hc, M[:, 1:])
+
     def test_modelmatrix(self):
         al = oapackage.exampleArray(1, 0)
 
@@ -156,30 +156,30 @@ class TestModelmatrix(unittest.TestCase):
         array = oapackage.exampleArray(0, 0)
         r = oapackage.array2eigenModelMatrixMixed(array.selectFirstColumns(1), verbose=0)
         main_effects = r[0]
-        np.testing.assert_array_equal(main_effects, np.array([[-1., -1., -1., -1.,  1.,  1.,  1.,  1.]]).T)
+        np.testing.assert_array_equal(main_effects, np.array([[-1., -1., -1., -1., 1., 1., 1., 1.]]).T)
 
         array = oapackage.exampleArray(10, 0).selectFirstColumns(2)
         r = oapackage.array2eigenModelMatrixMixed(array, verbose=0)
         main_effects = r[0]
         interaction_model = r[1]
         np.testing.assert_array_almost_equal(main_effects, np.array([[-1.22474487, -0.70710678, -1.22474487, -0.70710678],
-                                                                     [-1.22474487, -0.70710678,  1.22474487, -0.70710678],
-                                                                     [-1.22474487, -0.70710678,  0.,  1.41421356],
-                                                                     [1.22474487, -0.70710678,  1.22474487, -0.70710678],
-                                                                     [1.22474487, -0.70710678,  1.22474487, -0.70710678],
-                                                                     [1.22474487, -0.70710678,  0.,  1.41421356],
-                                                                     [0.,  1.41421356, -1.22474487, -0.70710678],
-                                                                     [0.,  1.41421356, -1.22474487, -0.70710678],
-                                                                     [0.,  1.41421356,  0.,  1.41421356]]))
-        np.testing.assert_array_almost_equal(interaction_model, np.array([[1.5,  0.8660254,  0.8660254,  0.5],
-                                                                          [-1.5,  0.8660254, -0.8660254,  0.5],
+                                                                     [-1.22474487, -0.70710678, 1.22474487, -0.70710678],
+                                                                     [-1.22474487, -0.70710678, 0., 1.41421356],
+                                                                     [1.22474487, -0.70710678, 1.22474487, -0.70710678],
+                                                                     [1.22474487, -0.70710678, 1.22474487, -0.70710678],
+                                                                     [1.22474487, -0.70710678, 0., 1.41421356],
+                                                                     [0., 1.41421356, -1.22474487, -0.70710678],
+                                                                     [0., 1.41421356, -1.22474487, -0.70710678],
+                                                                     [0., 1.41421356, 0., 1.41421356]]))
+        np.testing.assert_array_almost_equal(interaction_model, np.array([[1.5, 0.8660254, 0.8660254, 0.5],
+                                                                          [-1.5, 0.8660254, -0.8660254, 0.5],
[…]
