### 2019-02-15 #1097 @eendebakpt
@danielskatz @gsagnol @djmitche @alanrvazquez Thank you again for reviewing our package and providing valuable feedback. In a series of PRs we addressed the comments you had. The changes have been merged to the dev branch and there a a new release on pypi. Can you take the time to review the package once more?
