### 2019-02-17 3d620b7
refactor Fvalues
