### 2019-02-17 dedef98
refactor Fvalues
