+++ src/arraytools.cpp
@@ -3489,18 +3489,19 @@ int jstruct_t::maxJ () const {
         return vmax;
 }
 
-std::vector< int > jstruct_t::number_J_values(int strength) const {
+int jstruct_t::number_J_values(int strength) const {
 	assert(strength >= 1);
-	int x = pow((double)2, strength + 1);
-	int nn = floor((double)N / x) + 1;
+	int Jstep = pow((double)2, strength + 1);
+	int nn = floor((double)N / Jstep) + 1;
 	return nn;
 }
 
 std::vector< int > jstruct_t::Fval (int strength) const {
+		int Jstep = pow((double)2, strength + 1);
 		int nn = this->number_J_values(strength);
-        std::vector< int > Fv (nn);
+		std::vector< int > Fv (nn);
         for (int i = 0; i < nn; i++) {
-                Fv[i] = N - x * i;
+                Fv[i] = N - Jstep * i;
         }
         return Fv;
 }
@@ -3509,10 +3510,11 @@ std::vector< int > jstruct_t::calculateF (int strength) const {
         int Nmax = N;
 
 		int nn = this->number_J_values(strength);
+		int Jstep = pow((double)2, strength + 1);
 		std::vector< int > F (nn);
 
         for (int i = 0; i < nc; i++) {
-                int fi = (N - abs (values[i])) / x;
+                int fi = (N - abs (values[i])) / Jstep;
                 F[fi]++;
         }
         return F;
