### 2019-02-18 #1097 @danielskatz
👋 @gsagnol - please let us know your thoughts as well:
Does this now allow you to check all the boxes?
If not, what else do you think needs to be done?
