### 2019-02-20 #1097 @gsagnol
It's OK for me, all boxes checked !
