### 2019-02-20 db8c691
Minor corrections and additions. Pieter needs to include a link.
