### 2019-02-20 #1097 @danielskatz
@eendebakpt - are you happy with everything now?  The paper, the software version in the repo, the software that is archived in zenodo?
