### 2019-02-20 f33b758
safe initialization of arraydata_t from vector
