### 2019-02-20 590ecd7
add check on initialization of arraydata_t
