### 2019-02-20 97acf75
add check on initialization of arraydata_t
