+++ src/version.h
@@ -1,1 +1,1 @@
-const char *__version__ = "2.6.1";
+const char *__version__ = "2.6.2";
