### 2019-02-20 4cd4d86
Final Revision [Alan] (#142)
* Correct typos.

* Minor corrections and additions. Pieter needs to include a link.

* Minor addition.

* Minor additions. Pieter needs to include a link.

* Minor modifications.

* Minor modifications.

* Minor modifications and additions. Pieter needs to include two examples.

* More details in mixed-level example. Capital letters for examples. Pieter needs to add two examples.

* Minor modifications. Pieter needs to include link.

* Remove please. Pieter needs to add link.

* Minor modifications. Pieter needs to incldue link.

* Minor addition.

* Minor additio
[…]
