### 2019-02-21 4abe58e
Include section of definitions and add definition of optimal design
