+++ docs/data.rst
@@ -4,8 +4,8 @@
 Data representation
 ===================
 
-All designs handled by the OApackage are integer valued. The designs (whether these are orthogonal arrays, optimal designs or conferences designs)
-are stored in an :meth:`array_link` object.
+All designs handled by the Orthogonal Array package are integer valued. The designs (whether these are orthogonal arrays, optimal designs or conferences designs)
+are stored in an :meth:`array_link` object. The formal definitions of orthogonal arrays, optimal designs and conference designs are included in the section :ref:`Definitions of arrays and designs`.
 
 An orthogonal array (OA) of strength :math:`{t}`, :math:`{N}` runs and
 :math:`{n}` factors at :math:`{s}` levels is an :math:`{N}\times {n}`
