### 2019-02-21 886b596
Minor edits. Pieter needs to include a link to a new section in Properties.
