+++ docs/data.rst
@@ -7,32 +7,11 @@ Data representation
 All designs handled by the Orthogonal Array package are integer valued. The designs (whether these are orthogonal arrays, optimal designs or conferences designs)
 are stored in an :meth:`array_link` object. The formal definitions of orthogonal arrays, optimal designs and conference designs are included in the section :ref:`Definitions of arrays and designs`.
 
-An orthogonal array (OA) of strength :math:`{t}`, :math:`{N}` runs and
-:math:`{n}` factors at :math:`{s}` levels is an :math:`{N}\times {n}`
-array of symbols :math:`0,
-\ldots,({s}-1)`, such that for every subset of :math:`{t}` columns,
-every :math:`{t}`-tuple occurs equally
-often :cite:`Rao1947`. The set of all strength-:math:`{t}` OAs with 
-:math:`{N}` runs and :math:`{n}` factors at :math:`{s}` levels 
-is denoted by :math:`{\operatorname{OA}({N}; {t}; {s}^{n})}`. 
-The OAs are represented by arrays (data in memory is stored in column-major form).
-
-A D-optimal design :cite:`Donev2007` (:math:`X`) is an :math:`{N}\times {n}` array 
-that maximizes the :math:`{(\operatorname{det}({X}^{T}_{M}{X}^{\phantom{T}}_{M})^{1/p})/N}`,
-for a given :math:`{N}\times {p}` model matrix :math:`{X}_{M}` (for details see :ref:`Model matrices`).
-An orthogonal array is called D-optimal if it provides the largest determinant among all comparable orthogonal arrays.
-
-For :math:`{N}` even, a conference design :math:`C` is 
-an :math:`{N}\times {n}` array which satisfies :math:`{C}^{T}C = (n-1) I_{n}`,
-with :math:`{C}_{ii} = 0` and :math:`{C}_{ij} \in \{-1,1\}`, for 
-:math:`{i} \neq {j}` and :math:`{i}, {j} = 1, \ldots, n`.
-See :cite:`Xiao2012`, :cite:`Wu2009`.
-
 
 Data structures
 ---------------
 
-The package contains several data structures. Here we describe the main
+The package contains several data structures. Here, we describe the main
 structures and their use.
 
   :meth:`~oalib.array_link`
@@ -41,7 +20,7 @@ structures and their use.
     :meth:`~oalib.arraylist_t` object, which is implemented as a :code:`std::deque` container.
 
   :meth:`~oalib.arrayfile_t`
-    This is an object that allows for reading and writing of arrays to
+    This object allows for reading and writing of arrays to
     disk.
 
   :meth:`~oalib.arraydata_t`
@@ -52,12 +31,12 @@ structures and their use.
     The structure describing a certain class of conference designs.
 
   :meth:`~oalib.array_transformation_t`
-    This describes a transformation of an orthogonal array, which 
+    The structure describing a transformation of an orthogonal array, which 
     includes row and column permutations, as well as permutations 
     to the symbols in one or more columns.
 
   :meth:`~oalib.conference_transformation_t`
-    This describes a transformation of conference design or double conference design, which includes 
+    The structure describing a transformation of conference design or double conference design, which includes 
     row and column permutations, as well as sign switches to the elements
     in one or more rows and columns.
 
@@ -66,7 +45,7 @@ Representing arrays
 
 The structure containing an orthogonal array is called the
 :cpp:class:`array_link` structure. It consists of a specified number of rows and
-columns, the data (integer valued) and an index. In the Python interface the :meth:`array_link` object can be indexed just as
+columns, the data (integer valued) and an index. In the Python interface, the :meth:`array_link` object can be indexed just as
 normal arrays. 
 
 It is also possible to convert to a Numpy array. The
@@ -116,8 +95,7 @@ Classes of arrays
 -----------------
 
[…]
