### 2019-02-21 ce5f665
Minor edits. Pieter needs to include a link to a new section in Properties.
