### 2019-02-21 cfe4695
Minor edits.
