+++ docs/generation.rst
@@ -1,8 +1,7 @@
 Generation of designs
 =====================
 
-The OApackage can be used to generate several classes of arrays. A collection of
-arrays generated by the package is available on the website http://www.pietereendebak.nl/oapackage/index.html.
+The Orthogonal Array package can be used to generate several classes of arrays and designs. A collection of arrays and designs generated by the package is available on the website http://www.pietereendebak.nl/oapackage/index.html.
       
         
 Generation of orthogonal arrays
@@ -16,13 +15,13 @@ The main function for array extension is the function :cpp:func:`extend_arraylis
 to extend, a specification of the class of arrays in :cpp:class:`arraydata_t` and the
 options :cpp:class:`OAextend` for the algorithm.
 
-An example of a session that extends a designs is:
+An example of a session that extends an array is:
 
 .. doctest::
    
    >>> import oapackage
-   >>> N=8; ncols=3;
-   >>> arrayclass=oapackage.arraydata_t(2, N, 2, ncols)
+   >>> nrows=8; ncols=3;
+   >>> arrayclass=oapackage.arraydata_t(2, nrows, 2, ncols)
    >>> root_array=arrayclass.create_root() 
    >>> root_array.showarraycompact()
    00
@@ -37,15 +36,16 @@ An example of a session that extends a designs is:
    >>> print('found %d extensions of the root array' % len(array_list))
    found 2 extensions of the root array
 
-A more detailed example is :ref:`Enumerate orthogonal arrays`.
+A more detailed example is included in :ref:`Enumerate orthogonal arrays`.
 
 Conference designs
 ------------------
 
-An :math:`n\times k` conference design is an :math:`N\times k` matrix
+A conference design is an :math:`N\times k` matrix
 with entries 0, -1, +1 such that i) in each column the symbol 0 occurs
 exactly one time and ii) all columns are orthogonal to each other.
-For details see :cite:`Schoen2018dsd`, :cite:`wiki:ConferenceMatrix`.
+For details on conference designs, see the section 
+:ref:`Statistical properties of orthogonal arrays` and :cite:`Schoen2018dsd`. An example of a session to generate conference designs is the following:
 
 .. admonition:: Generate conference designs with 8 rows
 
@@ -71,15 +71,16 @@ For details see :cite:`Schoen2018dsd`, :cite:`wiki:ConferenceMatrix`.
     number of non-isomorphic conference designs with 6 columns: 11
 
 
-An example notebook with more functionality is
+An example notebook with more functionality is included in
 :ref:`Generation and analysis of conference designs`.
 The full interface for conference designs is available
 in the :ref:`Interface for conference designs`.
 
 The main functions to extend conference and double conference designs are
-:cpp:func:`extend_conference` and :cpp:func:`extend_double_conference`.
+:cpp:func:`extend_conference` and :cpp:func:`extend_double_conference`, respectively.
 The low-level functions for generating candidate extension columns of conference and double conference designs
-are :cpp:func:`generateConferenceExtensions` and :cpp:func:`generateDoubleConferenceExtensions`.
+are :cpp:func:`generateConferenceExtensions` and 
+:cpp:func:`generateDoubleConferenceExtensions`, respectively.
 
 The conference designs are generated in :ref:`LMC0 <LMC0>` form.
 
@@ -89,14 +90,13 @@ Calculation of D-efficient designs
 
 D-efficient designs (sometimes called D-optimal designs) can be calculated with the function :py:meth:`oapackage.Doptim.Doptimize`.
 This function uses a coordinate-exchange algorithm to generate designs
-with good properties for the :math:`D`-efficiency.
[…]
