### 2019-02-21 f44eaaf
Correction link of properties of designs.
