### 2019-02-21 30b334e
Major edits and include definition of F4 vector for conference designs.
