### 2019-02-21 0fb5ed9
Major edits and include definition of F4 vector for conference designs.
