+++ docs/properties.rst
@@ -247,7 +247,7 @@ The following example shows how to calculate the :math:`D`-, :math:`{D}_{s}`-, :
    OApackage calculates the :math:`D_{s}`-criterion using the eigenvalues of the SVD of the matrices :math:`{X}` and :math:`{X_{01}}`.
    Finally, the package calculates the :math:`D_{s}`-efficiency of :math:`D` as :math:`D_{s,\text{eff}}(A) = D_{s,\text{crit}}(A)^{1/m}`, where :math:`m` is the number of factors. 
    
-   In a similar way the :math:`D_{1}`-efficiency of a design :math:`{A}` with :math:`n` factors and model matrix of intercept and main effects :math:`{X01}`,  is defined as
+   In a similar way the :math:`D_{1}`-efficiency of a design :math:`{A}` with :math:`n` factors and model matrix of intercept and main effects :math:`{X_{01}}`,  is defined as
 
    .. math::
     
@@ -266,7 +266,7 @@ the interaction model in :math:`l` factors, that is, the model including the
 intercept, all :math:`l` main effects and all :math:`l(l-1)/2` two-factor
 interactions. The PEC sequence is the vector :math:`(PEC_{1}, PEC_{2}, \ldots, PEC_{k})`. Now, let :math:`PIC_{l}` denote the average :math:`D`-efficiency for the interaction model in :math:`l` factors accross all :math:`N`-run :math:`l`-factor subdesigns. The PIC sequence is the vector :math:`(PIC_{1}, PIC_{2}, \ldots, PIC_{k})`.
 The OApackage can calculate the PEC and PIC sequences of two-level designs with
-:cpp:func:`PECsequence` and :cpp:func:`PICsequence`.
+:cpp:func:`PECsequence` and :cpp:func:`PICsequence`, respectively.
     
 
 The following example shows how to compute the PEC and PIC sequences of a two-level orthogonal array using the OApackage.
@@ -287,8 +287,16 @@ The following example shows how to compute the PEC and PIC sequences of a two-le
 Properties of conference designs
 --------------------------------
 
-In :cite:`Schoen2018dsd` it is shown that the :math:`F_4` vector is usefull in
-classifying definitieve screening designs that are generated by folding over a conference design.
+In :cite:`Schoen2018dsd`, it is shown that the :math:`F_4` vector is useful for
+classifying definitive screening designs :cite:`Xiao2012` that are generated by folding over a conference design. To calculate the :math:`F_4` vector, we first need to compute the :math:`J_4`-characteristics of the defnitive screening design. The calculations for 
+the :math:`J_k`-characteristics of conference designs are similar as 
+for orthogonal arrays; see :ref:`Statistical properties of orthogonal arrays`. 
+Consider a definitive screening design with :math:`N` runs and 
+more than four factors. The :math:`F_4` vector of this design collects 
+the frequencies of the :math:`J_4`-characteristics of 
+:math:`{2N − 8\lambda}` for :math:`{\lambda = 1,\ldots,N/4}` 
+when :math:`N` is a multiple of 4, or :math:`{\lambda = 1, \ldots, (N − 2)/4}`
+when :math:`N` is an odd multiple of 2.
 
 .. admonition:: Calculate the F4 vector for a conference design
 
@@ -303,7 +311,7 @@ classifying definitieve screening designs that are generated by folding over a c
     >>> definitive_screening_design = oapackage.conference2DSD(array)
 
 The individual :math:`J_k`-characteristics can be calculated with the method :cpp:func:`Jcharacteristics_conference`.
-For conference designs we can calculate the projection statistics using
+For conference designs, we can calculate the projection statistics using
 :meth:`~oalib.conferenceProjectionStatistics`.
 
 .. admonition:: Calculate projection statistics for conference designs
