+++ docs/mixedlevel_modelmatrix.rst
@@ -2,8 +2,8 @@ Model matrices for mixed-level orthogonal arrays
 ================================================
 
 
-For mixed-level orthogonal arrays we generate model matrices using Helmert contrasts.
-First we define an example array and calculate the Helmert contrasts for 2 and 3 levels.
+For mixed-level orthogonal arrays, we generate model matrices using Helmert contrasts.
+First, we define an example array and calculate the Helmert contrasts for 2 and 3 levels.
 
 The Helmert contrasts for a given number of levels can be generated with :meth:`~oapackage.oahelper.helmert_contrasts`.
 The rows of the array generated by the method correspond to the levels.
@@ -67,7 +67,7 @@ yields the same results as the direct calculation for a 2-level design.
          
 The interaction effects are created by determining for all pair-wise combinations of columns the 
 product between the elements of the contrast vectors.
-For this array there is only one column combination: one 3-level factor and one 2-level factor.
+For this array, there is only one column combination: one 3-level factor and one 2-level factor.
 The number of columns in the interaction part is therefore :math:`(3-1)(2-1) = 2`.
          
 .. admonition:: Interaction model and information matrix
