### 2019-02-21 43e5be7
Include definitions of model matrices for conference designs. Pieter needs to add link.
