### 2019-02-21 39453e0
Include definitions of model matrices for conference designs. Pieter needs to add link.
