### 2019-02-21 2cfd87a
Update calculation of F values (#141)
* refactor Fvalues

* refactor Fvalues

* update docstrings
