### 2019-02-21 cc7ae03
Final updates from Alan (#145)
* Minor edits. Pieter needs to include a link to a new section in Properties.

* Include new section and definitions.

* Minor edits.

* Minor edits.

* Major edits and include definition of F4 vector for conference designs.

* Added a few commas.

* Include definitions of model matrices for conference designs. Pieter needs to add link.

* add missing text

* fix merge conflicts
