### 2019-02-22 e23b9dc
update documentation; add tests; add check on number of rows in Fvalues calculation
