### 2019-02-22 a1bd771
update documentation; add tests; add check on number of rows in Fvalues calculation
