+++ docs/properties.rst
@@ -281,7 +281,7 @@ The following example shows how to calculate the :math:`D`-, :math:`{D}_{s}`-, :
 
    .. math::
     
-       D_{s,\text{eff}}(A) = ( \operatorname{det}((X_{01})^{T}(X_{01}) )^{1/(n+1)}  \label{formula:D1efficiency}
+       D_{1,\text{eff}}(A) = ( \operatorname{det}((X_{01})^{T}(X_{01}) )^{1/(n+1)}  \label{formula:D1efficiency}
    
 Projection capacities
 ---------------------
@@ -318,15 +318,21 @@ Properties of conference designs
 --------------------------------
 
 In :cite:`Schoen2018dsd`, it is shown that the :math:`F_4` vector is useful for
-classifying definitive screening designs :cite:`Xiao2012` that are generated by folding over a conference design. To calculate the :math:`F_4` vector, we first need to compute the :math:`J_4`-characteristics of the defnitive screening design. The calculations for 
+classifying definitive screening designs :cite:`Xiao2012` that are generated by folding over
+a conference design. To calculate the :math:`F_4` vector, we first need to compute
+the :math:`J_4`-characteristics of the conference design. The calculations for 
 the :math:`J_k`-characteristics of conference designs are similar as 
 for orthogonal arrays; see :ref:`Statistical properties of orthogonal arrays`. 
-Consider a definitive screening design with :math:`N` runs and 
-more than four factors. The :math:`F_4` vector of this design collects 
+Consider a definitive screening design constructed from an :math:`N`-run conference design with at least
+four factors. The :math:`F_4` vector of this design collects 
 the frequencies of the :math:`J_4`-characteristics of 
 :math:`{2N − 8\lambda}` for :math:`{\lambda = 1,\ldots,N/4}` 
 when :math:`N` is a multiple of 4, or :math:`{\lambda = 1, \ldots, (N − 2)/4}`
 when :math:`N` is an odd multiple of 2.
+
+Note: the :math:`J_4`-values of a definitive screening design generated by folding over a conference design
+are twice the value of the :math:`J_4`-values of the conference design. The :math:`F_4` vector of
+a conference design and the corresponding definitive screening design are equal.
 
 .. admonition:: Calculate the F4 vector for a conference design
 
