### 2019-02-22 e20a0fc
update documentation; add tests; add check on number of rows in Fvalues calculation (#148)
