### 2019-02-22 c433dd1
fix docstring
