### 2019-02-23 070d071
update docstrings
