### 2019-02-23 #1097 @eendebakpt
@danielskatz We are happy with the paper. Could you update the software version to the latest (2.6.2). We made some small updates to the documentation, but did not change the paper.
What do you mean exactly by the software that is archived in zenodo?
