### 2019-02-23 #1097 @danielskatz
@eendebakpt - you need to make an archive of this version of the software, in zenodo or in another repository.

See the third bullet in https://joss.readthedocs.io/en/latest/submitting.html#the-review-process

(And see https://guides.github.com/activities/citable-code/ for how to do this for Zenodo)

Once you have done so, paste the DOI of the archive here.
