### 2019-02-27 #1097 @eendebakpt
@danielskatz To create a zenodo archive I updated the version number of the package, it is now 2.6.3

The DOI of the archive is 10.5281/zenodo.2579025
