### 2019-02-27 #1097 @eendebakpt
@danielskatz @gsagnol @djmitche Thanks for reviewing the package. It helped us to improve the package structure.
