### 2019-02-27 #1097 @danielskatz
🎉🎉🎉 This was the 500th accepted paper in JOSS! 🎉🎉🎉
